@@ -1,10 +1,10 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.28.1
-// 	protoc        v5.26.1
+// 	protoc-gen-go v1.36.6
+// 	protoc        v5.29.3
 // source: charts.proto
 
-package chartsProto
+package charts
 
 import (
 	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
@@ -12,6 +12,7 @@ import (
 	timestamppb "google.golang.org/protobuf/types/known/timestamppb"
 	reflect "reflect"
 	sync "sync"
+	unsafe "unsafe"
 )
 
 const (
@@ -74,21 +75,18 @@ func (ChartPeriod) EnumDescriptor() ([]byte, []int) {
 }
 
 type GetPoolStatsRequest struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Period        ChartPeriod            `protobuf:"varint,1,opt,name=period,proto3,enum=charts.ChartPeriod" json:"period,omitempty"`
+	Solo          bool                   `protobuf:"varint,2,opt,name=solo,proto3" json:"solo,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Period ChartPeriod `protobuf:"varint,1,opt,name=period,proto3,enum=charts.ChartPeriod" json:"period,omitempty"`
-	Solo   bool        `protobuf:"varint,2,opt,name=solo,proto3" json:"solo,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *GetPoolStatsRequest) Reset() {
 	*x = GetPoolStatsRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_charts_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_charts_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *GetPoolStatsRequest) String() string {
@@ -99,7 +97,7 @@ func (*GetPoolStatsRequest) ProtoMessage() {}
 
 func (x *GetPoolStatsRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_charts_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -129,23 +127,20 @@ func (x *GetPoolStatsRequest) GetSolo() bool {
 }
 
 type PoolStatsPoint struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Hashrate      []byte                 `protobuf:"bytes,1,opt,name=hashrate,proto3" json:"hashrate,omitempty"`
+	AvgHashrate   []byte                 `protobuf:"bytes,2,opt,name=avg_hashrate,json=avgHashrate,proto3" json:"avg_hashrate,omitempty"`
+	MinersCount   uint32                 `protobuf:"varint,3,opt,name=miners_count,json=minersCount,proto3" json:"miners_count,omitempty"`
+	Date          *timestamppb.Timestamp `protobuf:"bytes,4,opt,name=date,proto3" json:"date,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Hashrate    []byte                 `protobuf:"bytes,1,opt,name=hashrate,proto3" json:"hashrate,omitempty"`
-	AvgHashrate []byte                 `protobuf:"bytes,2,opt,name=avg_hashrate,json=avgHashrate,proto3" json:"avg_hashrate,omitempty"`
-	MinersCount uint32                 `protobuf:"varint,3,opt,name=miners_count,json=minersCount,proto3" json:"miners_count,omitempty"`
-	Date        *timestamppb.Timestamp `protobuf:"bytes,4,opt,name=date,proto3" json:"date,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *PoolStatsPoint) Reset() {
 	*x = PoolStatsPoint{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_charts_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_charts_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *PoolStatsPoint) String() string {
@@ -156,7 +151,7 @@ func (*PoolStatsPoint) ProtoMessage() {}
 
 func (x *PoolStatsPoint) ProtoReflect() protoreflect.Message {
 	mi := &file_charts_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -200,20 +195,17 @@ func (x *PoolStatsPoint) GetDate() *timestamppb.Timestamp {
 }
 
 type PoolStatsPoints struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Points        []*PoolStatsPoint      `protobuf:"bytes,1,rep,name=points,proto3" json:"points,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Points []*PoolStatsPoint `protobuf:"bytes,1,rep,name=points,proto3" json:"points,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *PoolStatsPoints) Reset() {
 	*x = PoolStatsPoints{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_charts_proto_msgTypes[2]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_charts_proto_msgTypes[2]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *PoolStatsPoints) String() string {
@@ -224,7 +216,7 @@ func (*PoolStatsPoints) ProtoMessage() {}
 
 func (x *PoolStatsPoints) ProtoReflect() protoreflect.Message {
 	mi := &file_charts_proto_msgTypes[2]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -247,20 +239,17 @@ func (x *PoolStatsPoints) GetPoints() []*PoolStatsPoint {
 }
 
 type GetNetworkDifficultiesRequest struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Period        ChartPeriod            `protobuf:"varint,1,opt,name=period,proto3,enum=charts.ChartPeriod" json:"period,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Period ChartPeriod `protobuf:"varint,1,opt,name=period,proto3,enum=charts.ChartPeriod" json:"period,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *GetNetworkDifficultiesRequest) Reset() {
 	*x = GetNetworkDifficultiesRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_charts_proto_msgTypes[3]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_charts_proto_msgTypes[3]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *GetNetworkDifficultiesRequest) String() string {
@@ -271,7 +260,7 @@ func (*GetNetworkDifficultiesRequest) ProtoMessage() {}
 
 func (x *GetNetworkDifficultiesRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_charts_proto_msgTypes[3]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -294,22 +283,19 @@ func (x *GetNetworkDifficultiesRequest) GetPeriod() ChartPeriod {
 }
 
 type NetworkDifficultiesPoint struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
+	state         protoimpl.MessageState `protogen:"open.v1"`
 	Difficulty    []byte                 `protobuf:"bytes,1,opt,name=difficulty,proto3" json:"difficulty,omitempty"`
 	AvgDifficulty []byte                 `protobuf:"bytes,2,opt,name=avg_difficulty,json=avgDifficulty,proto3" json:"avg_difficulty,omitempty"`
 	Date          *timestamppb.Timestamp `protobuf:"bytes,3,opt,name=date,proto3" json:"date,omitempty"`
+	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *NetworkDifficultiesPoint) Reset() {
 	*x = NetworkDifficultiesPoint{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_charts_proto_msgTypes[4]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_charts_proto_msgTypes[4]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *NetworkDifficultiesPoint) String() string {
@@ -320,7 +306,7 @@ func (*NetworkDifficultiesPoint) ProtoMessage() {}
 
 func (x *NetworkDifficultiesPoint) ProtoReflect() protoreflect.Message {
 	mi := &file_charts_proto_msgTypes[4]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -357,20 +343,17 @@ func (x *NetworkDifficultiesPoint) GetDate() *timestamppb.Timestamp {
 }
 
 type NetworkDifficultiesPoints struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState      `protogen:"open.v1"`
+	Points        []*NetworkDifficultiesPoint `protobuf:"bytes,1,rep,name=points,proto3" json:"points,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Points []*NetworkDifficultiesPoint `protobuf:"bytes,1,rep,name=points,proto3" json:"points,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *NetworkDifficultiesPoints) Reset() {
 	*x = NetworkDifficultiesPoints{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_charts_proto_msgTypes[5]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_charts_proto_msgTypes[5]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *NetworkDifficultiesPoints) String() string {
@@ -381,7 +364,7 @@ func (*NetworkDifficultiesPoints) ProtoMessage() {}
 
 func (x *NetworkDifficultiesPoints) ProtoReflect() protoreflect.Message {
 	mi := &file_charts_proto_msgTypes[5]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -404,20 +387,17 @@ func (x *NetworkDifficultiesPoints) GetPoints() []*NetworkDifficultiesPoint {
 }
 
 type GetRoundsRequest struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Period        ChartPeriod            `protobuf:"varint,1,opt,name=period,proto3,enum=charts.ChartPeriod" json:"period,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Period ChartPeriod `protobuf:"varint,1,opt,name=period,proto3,enum=charts.ChartPeriod" json:"period,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *GetRoundsRequest) Reset() {
 	*x = GetRoundsRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_charts_proto_msgTypes[6]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_charts_proto_msgTypes[6]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *GetRoundsRequest) String() string {
@@ -428,7 +408,7 @@ func (*GetRoundsRequest) ProtoMessage() {}
 
 func (x *GetRoundsRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_charts_proto_msgTypes[6]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -451,23 +431,20 @@ func (x *GetRoundsRequest) GetPeriod() ChartPeriod {
 }
 
 type RoundsPoint struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
+	state            protoimpl.MessageState `protogen:"open.v1"`
 	RoundsCount      uint32                 `protobuf:"varint,1,opt,name=rounds_count,json=roundsCount,proto3" json:"rounds_count,omitempty"`
 	MinersCount      uint32                 `protobuf:"varint,2,opt,name=miners_count,json=minersCount,proto3" json:"miners_count,omitempty"`
 	AvgRoundInterval float64                `protobuf:"fixed64,3,opt,name=avg_round_interval,json=avgRoundInterval,proto3" json:"avg_round_interval,omitempty"`
 	Date             *timestamppb.Timestamp `protobuf:"bytes,4,opt,name=date,proto3" json:"date,omitempty"`
+	unknownFields    protoimpl.UnknownFields
+	sizeCache        protoimpl.SizeCache
 }
 
 func (x *RoundsPoint) Reset() {
 	*x = RoundsPoint{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_charts_proto_msgTypes[7]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_charts_proto_msgTypes[7]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *RoundsPoint) String() string {
@@ -478,7 +455,7 @@ func (*RoundsPoint) ProtoMessage() {}
 
 func (x *RoundsPoint) ProtoReflect() protoreflect.Message {
 	mi := &file_charts_proto_msgTypes[7]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -522,20 +499,17 @@ func (x *RoundsPoint) GetDate() *timestamppb.Timestamp {
 }
 
 type RoundsPoints struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Points        []*RoundsPoint         `protobuf:"bytes,1,rep,name=points,proto3" json:"points,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Points []*RoundsPoint `protobuf:"bytes,1,rep,name=points,proto3" json:"points,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *RoundsPoints) Reset() {
 	*x = RoundsPoints{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_charts_proto_msgTypes[8]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_charts_proto_msgTypes[8]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *RoundsPoints) String() string {
@@ -546,7 +520,7 @@ func (*RoundsPoints) ProtoMessage() {}
 
 func (x *RoundsPoints) ProtoReflect() protoreflect.Message {
 	mi := &file_charts_proto_msgTypes[8]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -569,22 +543,19 @@ func (x *RoundsPoints) GetPoints() []*RoundsPoint {
 }
 
 type GetMinerChartRequest struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Miner         string                 `protobuf:"bytes,1,opt,name=miner,proto3" json:"miner,omitempty"`
+	Solo          *bool                  `protobuf:"varint,2,opt,name=solo,proto3,oneof" json:"solo,omitempty"`
+	Period        ChartPeriod            `protobuf:"varint,3,opt,name=period,proto3,enum=charts.ChartPeriod" json:"period,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Miner  string      `protobuf:"bytes,1,opt,name=miner,proto3" json:"miner,omitempty"`
-	Solo   *bool       `protobuf:"varint,2,opt,name=solo,proto3,oneof" json:"solo,omitempty"`
-	Period ChartPeriod `protobuf:"varint,3,opt,name=period,proto3,enum=charts.ChartPeriod" json:"period,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *GetMinerChartRequest) Reset() {
 	*x = GetMinerChartRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_charts_proto_msgTypes[9]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_charts_proto_msgTypes[9]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *GetMinerChartRequest) String() string {
@@ -595,7 +566,7 @@ func (*GetMinerChartRequest) ProtoMessage() {}
 
 func (x *GetMinerChartRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_charts_proto_msgTypes[9]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -632,21 +603,18 @@ func (x *GetMinerChartRequest) GetPeriod() ChartPeriod {
 }
 
 type MinerProfitabilityPoint struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Balance       uint64                 `protobuf:"varint,1,opt,name=balance,proto3" json:"balance,omitempty"`
+	Date          *timestamppb.Timestamp `protobuf:"bytes,2,opt,name=date,proto3" json:"date,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Balance uint64                 `protobuf:"varint,1,opt,name=balance,proto3" json:"balance,omitempty"`
-	Date    *timestamppb.Timestamp `protobuf:"bytes,2,opt,name=date,proto3" json:"date,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *MinerProfitabilityPoint) Reset() {
 	*x = MinerProfitabilityPoint{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_charts_proto_msgTypes[10]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_charts_proto_msgTypes[10]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *MinerProfitabilityPoint) String() string {
@@ -657,7 +625,7 @@ func (*MinerProfitabilityPoint) ProtoMessage() {}
 
 func (x *MinerProfitabilityPoint) ProtoReflect() protoreflect.Message {
 	mi := &file_charts_proto_msgTypes[10]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -687,20 +655,17 @@ func (x *MinerProfitabilityPoint) GetDate() *timestamppb.Timestamp {
 }
 
 type MinerProfitabilitiesPoints struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState     `protogen:"open.v1"`
+	Points        []*MinerProfitabilityPoint `protobuf:"bytes,1,rep,name=points,proto3" json:"points,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Points []*MinerProfitabilityPoint `protobuf:"bytes,1,rep,name=points,proto3" json:"points,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *MinerProfitabilitiesPoints) Reset() {
 	*x = MinerProfitabilitiesPoints{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_charts_proto_msgTypes[11]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_charts_proto_msgTypes[11]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *MinerProfitabilitiesPoints) String() string {
@@ -711,7 +676,7 @@ func (*MinerProfitabilitiesPoints) ProtoMessage() {}
 
 func (x *MinerProfitabilitiesPoints) ProtoReflect() protoreflect.Message {
 	mi := &file_charts_proto_msgTypes[11]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -734,22 +699,19 @@ func (x *MinerProfitabilitiesPoints) GetPoints() []*MinerProfitabilityPoint {
 }
 
 type MinerHashratesPoint struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Hashrate      []byte                 `protobuf:"bytes,1,opt,name=hashrate,proto3" json:"hashrate,omitempty"`
+	AvgHashrate   []byte                 `protobuf:"bytes,2,opt,name=avg_hashrate,json=avgHashrate,proto3" json:"avg_hashrate,omitempty"`
+	Date          *timestamppb.Timestamp `protobuf:"bytes,3,opt,name=date,proto3" json:"date,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Hashrate    []byte                 `protobuf:"bytes,1,opt,name=hashrate,proto3" json:"hashrate,omitempty"`
-	AvgHashrate []byte                 `protobuf:"bytes,2,opt,name=avg_hashrate,json=avgHashrate,proto3" json:"avg_hashrate,omitempty"`
-	Date        *timestamppb.Timestamp `protobuf:"bytes,3,opt,name=date,proto3" json:"date,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *MinerHashratesPoint) Reset() {
 	*x = MinerHashratesPoint{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_charts_proto_msgTypes[12]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_charts_proto_msgTypes[12]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *MinerHashratesPoint) String() string {
@@ -760,7 +722,7 @@ func (*MinerHashratesPoint) ProtoMessage() {}
 
 func (x *MinerHashratesPoint) ProtoReflect() protoreflect.Message {
 	mi := &file_charts_proto_msgTypes[12]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -797,20 +759,17 @@ func (x *MinerHashratesPoint) GetDate() *timestamppb.Timestamp {
 }
 
 type MinerHashratesPoints struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Points        []*MinerHashratesPoint `protobuf:"bytes,1,rep,name=points,proto3" json:"points,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Points []*MinerHashratesPoint `protobuf:"bytes,1,rep,name=points,proto3" json:"points,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *MinerHashratesPoints) Reset() {
 	*x = MinerHashratesPoints{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_charts_proto_msgTypes[13]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_charts_proto_msgTypes[13]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *MinerHashratesPoints) String() string {
@@ -821,7 +780,7 @@ func (*MinerHashratesPoints) ProtoMessage() {}
 
 func (x *MinerHashratesPoints) ProtoReflect() protoreflect.Message {
 	mi := &file_charts_proto_msgTypes[13]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -844,22 +803,19 @@ func (x *MinerHashratesPoints) GetPoints() []*MinerHashratesPoint {
 }
 
 type GetMinerWorkerChartRequest struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Miner         string                 `protobuf:"bytes,1,opt,name=miner,proto3" json:"miner,omitempty"`
+	Worker        string                 `protobuf:"bytes,2,opt,name=worker,proto3" json:"worker,omitempty"`
+	Period        ChartPeriod            `protobuf:"varint,3,opt,name=period,proto3,enum=charts.ChartPeriod" json:"period,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Miner  string      `protobuf:"bytes,1,opt,name=miner,proto3" json:"miner,omitempty"`
-	Worker string      `protobuf:"bytes,2,opt,name=worker,proto3" json:"worker,omitempty"`
-	Period ChartPeriod `protobuf:"varint,3,opt,name=period,proto3,enum=charts.ChartPeriod" json:"period,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *GetMinerWorkerChartRequest) Reset() {
 	*x = GetMinerWorkerChartRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_charts_proto_msgTypes[14]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_charts_proto_msgTypes[14]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *GetMinerWorkerChartRequest) String() string {
@@ -870,7 +826,7 @@ func (*GetMinerWorkerChartRequest) ProtoMessage() {}
 
 func (x *GetMinerWorkerChartRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_charts_proto_msgTypes[14]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -907,24 +863,21 @@ func (x *GetMinerWorkerChartRequest) GetPeriod() ChartPeriod {
 }
 
 type MinerSharesPoint struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
+	state                 protoimpl.MessageState `protogen:"open.v1"`
 	AcceptedSharesCount   uint32                 `protobuf:"varint,1,opt,name=accepted_shares_count,json=acceptedSharesCount,proto3" json:"accepted_shares_count,omitempty"`
 	RejectedSharesCount   uint32                 `protobuf:"varint,2,opt,name=rejected_shares_count,json=rejectedSharesCount,proto3" json:"rejected_shares_count,omitempty"`
 	StaleSharesCount      uint32                 `protobuf:"varint,3,opt,name=stale_shares_count,json=staleSharesCount,proto3" json:"stale_shares_count,omitempty"`
 	ValidBlockSharesCount uint32                 `protobuf:"varint,4,opt,name=valid_block_shares_count,json=validBlockSharesCount,proto3" json:"valid_block_shares_count,omitempty"`
 	Date                  *timestamppb.Timestamp `protobuf:"bytes,5,opt,name=date,proto3" json:"date,omitempty"`
+	unknownFields         protoimpl.UnknownFields
+	sizeCache             protoimpl.SizeCache
 }
 
 func (x *MinerSharesPoint) Reset() {
 	*x = MinerSharesPoint{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_charts_proto_msgTypes[15]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_charts_proto_msgTypes[15]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *MinerSharesPoint) String() string {
@@ -935,7 +888,7 @@ func (*MinerSharesPoint) ProtoMessage() {}
 
 func (x *MinerSharesPoint) ProtoReflect() protoreflect.Message {
 	mi := &file_charts_proto_msgTypes[15]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -986,20 +939,17 @@ func (x *MinerSharesPoint) GetDate() *timestamppb.Timestamp {
 }
 
 type MinerSharesPoints struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Points        []*MinerSharesPoint    `protobuf:"bytes,1,rep,name=points,proto3" json:"points,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Points []*MinerSharesPoint `protobuf:"bytes,1,rep,name=points,proto3" json:"points,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *MinerSharesPoints) Reset() {
 	*x = MinerSharesPoints{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_charts_proto_msgTypes[16]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_charts_proto_msgTypes[16]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *MinerSharesPoints) String() string {
@@ -1010,7 +960,7 @@ func (*MinerSharesPoints) ProtoMessage() {}
 
 func (x *MinerSharesPoints) ProtoReflect() protoreflect.Message {
 	mi := &file_charts_proto_msgTypes[16]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1034,200 +984,96 @@ func (x *MinerSharesPoints) GetPoints() []*MinerSharesPoint {
 
 var File_charts_proto protoreflect.FileDescriptor
 
-var file_charts_proto_rawDesc = []byte{
-	0x0a, 0x0c, 0x63, 0x68, 0x61, 0x72, 0x74, 0x73, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12, 0x06,
-	0x63, 0x68, 0x61, 0x72, 0x74, 0x73, 0x1a, 0x1f, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2f, 0x70,
-	0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2f, 0x74, 0x69, 0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d,
-	0x70, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x22, 0x56, 0x0a, 0x13, 0x47, 0x65, 0x74, 0x50, 0x6f,
-	0x6f, 0x6c, 0x53, 0x74, 0x61, 0x74, 0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x2b,
-	0x0a, 0x06, 0x70, 0x65, 0x72, 0x69, 0x6f, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0e, 0x32, 0x13,
-	0x2e, 0x63, 0x68, 0x61, 0x72, 0x74, 0x73, 0x2e, 0x43, 0x68, 0x61, 0x72, 0x74, 0x50, 0x65, 0x72,
-	0x69, 0x6f, 0x64, 0x52, 0x06, 0x70, 0x65, 0x72, 0x69, 0x6f, 0x64, 0x12, 0x12, 0x0a, 0x04, 0x73,
-	0x6f, 0x6c, 0x6f, 0x18, 0x02, 0x20, 0x01, 0x28, 0x08, 0x52, 0x04, 0x73, 0x6f, 0x6c, 0x6f, 0x22,
-	0xa2, 0x01, 0x0a, 0x0e, 0x50, 0x6f, 0x6f, 0x6c, 0x53, 0x74, 0x61, 0x74, 0x73, 0x50, 0x6f, 0x69,
-	0x6e, 0x74, 0x12, 0x1a, 0x0a, 0x08, 0x68, 0x61, 0x73, 0x68, 0x72, 0x61, 0x74, 0x65, 0x18, 0x01,
-	0x20, 0x01, 0x28, 0x0c, 0x52, 0x08, 0x68, 0x61, 0x73, 0x68, 0x72, 0x61, 0x74, 0x65, 0x12, 0x21,
-	0x0a, 0x0c, 0x61, 0x76, 0x67, 0x5f, 0x68, 0x61, 0x73, 0x68, 0x72, 0x61, 0x74, 0x65, 0x18, 0x02,
-	0x20, 0x01, 0x28, 0x0c, 0x52, 0x0b, 0x61, 0x76, 0x67, 0x48, 0x61, 0x73, 0x68, 0x72, 0x61, 0x74,
-	0x65, 0x12, 0x21, 0x0a, 0x0c, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x73, 0x5f, 0x63, 0x6f, 0x75, 0x6e,
-	0x74, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x0b, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x73, 0x43,
-	0x6f, 0x75, 0x6e, 0x74, 0x12, 0x2e, 0x0a, 0x04, 0x64, 0x61, 0x74, 0x65, 0x18, 0x04, 0x20, 0x01,
-	0x28, 0x0b, 0x32, 0x1a, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74,
-	0x6f, 0x62, 0x75, 0x66, 0x2e, 0x54, 0x69, 0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x52, 0x04,
-	0x64, 0x61, 0x74, 0x65, 0x22, 0x41, 0x0a, 0x0f, 0x50, 0x6f, 0x6f, 0x6c, 0x53, 0x74, 0x61, 0x74,
-	0x73, 0x50, 0x6f, 0x69, 0x6e, 0x74, 0x73, 0x12, 0x2e, 0x0a, 0x06, 0x70, 0x6f, 0x69, 0x6e, 0x74,
-	0x73, 0x18, 0x01, 0x20, 0x03, 0x28, 0x0b, 0x32, 0x16, 0x2e, 0x63, 0x68, 0x61, 0x72, 0x74, 0x73,
-	0x2e, 0x50, 0x6f, 0x6f, 0x6c, 0x53, 0x74, 0x61, 0x74, 0x73, 0x50, 0x6f, 0x69, 0x6e, 0x74, 0x52,
-	0x06, 0x70, 0x6f, 0x69, 0x6e, 0x74, 0x73, 0x22, 0x4c, 0x0a, 0x1d, 0x47, 0x65, 0x74, 0x4e, 0x65,
-	0x74, 0x77, 0x6f, 0x72, 0x6b, 0x44, 0x69, 0x66, 0x66, 0x69, 0x63, 0x75, 0x6c, 0x74, 0x69, 0x65,
-	0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x2b, 0x0a, 0x06, 0x70, 0x65, 0x72, 0x69,
-	0x6f, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0e, 0x32, 0x13, 0x2e, 0x63, 0x68, 0x61, 0x72, 0x74,
-	0x73, 0x2e, 0x43, 0x68, 0x61, 0x72, 0x74, 0x50, 0x65, 0x72, 0x69, 0x6f, 0x64, 0x52, 0x06, 0x70,
-	0x65, 0x72, 0x69, 0x6f, 0x64, 0x22, 0x91, 0x01, 0x0a, 0x18, 0x4e, 0x65, 0x74, 0x77, 0x6f, 0x72,
-	0x6b, 0x44, 0x69, 0x66, 0x66, 0x69, 0x63, 0x75, 0x6c, 0x74, 0x69, 0x65, 0x73, 0x50, 0x6f, 0x69,
-	0x6e, 0x74, 0x12, 0x1e, 0x0a, 0x0a, 0x64, 0x69, 0x66, 0x66, 0x69, 0x63, 0x75, 0x6c, 0x74, 0x79,
-	0x18, 0x01, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x0a, 0x64, 0x69, 0x66, 0x66, 0x69, 0x63, 0x75, 0x6c,
-	0x74, 0x79, 0x12, 0x25, 0x0a, 0x0e, 0x61, 0x76, 0x67, 0x5f, 0x64, 0x69, 0x66, 0x66, 0x69, 0x63,
-	0x75, 0x6c, 0x74, 0x79, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x0d, 0x61, 0x76, 0x67, 0x44,
-	0x69, 0x66, 0x66, 0x69, 0x63, 0x75, 0x6c, 0x74, 0x79, 0x12, 0x2e, 0x0a, 0x04, 0x64, 0x61, 0x74,
-	0x65, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1a, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65,
-	0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x54, 0x69, 0x6d, 0x65, 0x73, 0x74,
-	0x61, 0x6d, 0x70, 0x52, 0x04, 0x64, 0x61, 0x74, 0x65, 0x22, 0x55, 0x0a, 0x19, 0x4e, 0x65, 0x74,
-	0x77, 0x6f, 0x72, 0x6b, 0x44, 0x69, 0x66, 0x66, 0x69, 0x63, 0x75, 0x6c, 0x74, 0x69, 0x65, 0x73,
-	0x50, 0x6f, 0x69, 0x6e, 0x74, 0x73, 0x12, 0x38, 0x0a, 0x06, 0x70, 0x6f, 0x69, 0x6e, 0x74, 0x73,
-	0x18, 0x01, 0x20, 0x03, 0x28, 0x0b, 0x32, 0x20, 0x2e, 0x63, 0x68, 0x61, 0x72, 0x74, 0x73, 0x2e,
-	0x4e, 0x65, 0x74, 0x77, 0x6f, 0x72, 0x6b, 0x44, 0x69, 0x66, 0x66, 0x69, 0x63, 0x75, 0x6c, 0x74,
-	0x69, 0x65, 0x73, 0x50, 0x6f, 0x69, 0x6e, 0x74, 0x52, 0x06, 0x70, 0x6f, 0x69, 0x6e, 0x74, 0x73,
-	0x22, 0x3f, 0x0a, 0x10, 0x47, 0x65, 0x74, 0x52, 0x6f, 0x75, 0x6e, 0x64, 0x73, 0x52, 0x65, 0x71,
-	0x75, 0x65, 0x73, 0x74, 0x12, 0x2b, 0x0a, 0x06, 0x70, 0x65, 0x72, 0x69, 0x6f, 0x64, 0x18, 0x01,
-	0x20, 0x01, 0x28, 0x0e, 0x32, 0x13, 0x2e, 0x63, 0x68, 0x61, 0x72, 0x74, 0x73, 0x2e, 0x43, 0x68,
-	0x61, 0x72, 0x74, 0x50, 0x65, 0x72, 0x69, 0x6f, 0x64, 0x52, 0x06, 0x70, 0x65, 0x72, 0x69, 0x6f,
-	0x64, 0x22, 0xb1, 0x01, 0x0a, 0x0b, 0x52, 0x6f, 0x75, 0x6e, 0x64, 0x73, 0x50, 0x6f, 0x69, 0x6e,
-	0x74, 0x12, 0x21, 0x0a, 0x0c, 0x72, 0x6f, 0x75, 0x6e, 0x64, 0x73, 0x5f, 0x63, 0x6f, 0x75, 0x6e,
-	0x74, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x0b, 0x72, 0x6f, 0x75, 0x6e, 0x64, 0x73, 0x43,
-	0x6f, 0x75, 0x6e, 0x74, 0x12, 0x21, 0x0a, 0x0c, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x73, 0x5f, 0x63,
-	0x6f, 0x75, 0x6e, 0x74, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x0b, 0x6d, 0x69, 0x6e, 0x65,
-	0x72, 0x73, 0x43, 0x6f, 0x75, 0x6e, 0x74, 0x12, 0x2c, 0x0a, 0x12, 0x61, 0x76, 0x67, 0x5f, 0x72,
-	0x6f, 0x75, 0x6e, 0x64, 0x5f, 0x69, 0x6e, 0x74, 0x65, 0x72, 0x76, 0x61, 0x6c, 0x18, 0x03, 0x20,
-	0x01, 0x28, 0x01, 0x52, 0x10, 0x61, 0x76, 0x67, 0x52, 0x6f, 0x75, 0x6e, 0x64, 0x49, 0x6e, 0x74,
-	0x65, 0x72, 0x76, 0x61, 0x6c, 0x12, 0x2e, 0x0a, 0x04, 0x64, 0x61, 0x74, 0x65, 0x18, 0x04, 0x20,
-	0x01, 0x28, 0x0b, 0x32, 0x1a, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f,
-	0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x54, 0x69, 0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x52,
-	0x04, 0x64, 0x61, 0x74, 0x65, 0x22, 0x3b, 0x0a, 0x0c, 0x52, 0x6f, 0x75, 0x6e, 0x64, 0x73, 0x50,
-	0x6f, 0x69, 0x6e, 0x74, 0x73, 0x12, 0x2b, 0x0a, 0x06, 0x70, 0x6f, 0x69, 0x6e, 0x74, 0x73, 0x18,
-	0x01, 0x20, 0x03, 0x28, 0x0b, 0x32, 0x13, 0x2e, 0x63, 0x68, 0x61, 0x72, 0x74, 0x73, 0x2e, 0x52,
-	0x6f, 0x75, 0x6e, 0x64, 0x73, 0x50, 0x6f, 0x69, 0x6e, 0x74, 0x52, 0x06, 0x70, 0x6f, 0x69, 0x6e,
-	0x74, 0x73, 0x22, 0x7b, 0x0a, 0x14, 0x47, 0x65, 0x74, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x43, 0x68,
-	0x61, 0x72, 0x74, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x14, 0x0a, 0x05, 0x6d, 0x69,
-	0x6e, 0x65, 0x72, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x05, 0x6d, 0x69, 0x6e, 0x65, 0x72,
-	0x12, 0x17, 0x0a, 0x04, 0x73, 0x6f, 0x6c, 0x6f, 0x18, 0x02, 0x20, 0x01, 0x28, 0x08, 0x48, 0x00,
-	0x52, 0x04, 0x73, 0x6f, 0x6c, 0x6f, 0x88, 0x01, 0x01, 0x12, 0x2b, 0x0a, 0x06, 0x70, 0x65, 0x72,
-	0x69, 0x6f, 0x64, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0e, 0x32, 0x13, 0x2e, 0x63, 0x68, 0x61, 0x72,
-	0x74, 0x73, 0x2e, 0x43, 0x68, 0x61, 0x72, 0x74, 0x50, 0x65, 0x72, 0x69, 0x6f, 0x64, 0x52, 0x06,
-	0x70, 0x65, 0x72, 0x69, 0x6f, 0x64, 0x42, 0x07, 0x0a, 0x05, 0x5f, 0x73, 0x6f, 0x6c, 0x6f, 0x22,
-	0x63, 0x0a, 0x17, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x50, 0x72, 0x6f, 0x66, 0x69, 0x74, 0x61, 0x62,
-	0x69, 0x6c, 0x69, 0x74, 0x79, 0x50, 0x6f, 0x69, 0x6e, 0x74, 0x12, 0x18, 0x0a, 0x07, 0x62, 0x61,
-	0x6c, 0x61, 0x6e, 0x63, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x04, 0x52, 0x07, 0x62, 0x61, 0x6c,
-	0x61, 0x6e, 0x63, 0x65, 0x12, 0x2e, 0x0a, 0x04, 0x64, 0x61, 0x74, 0x65, 0x18, 0x02, 0x20, 0x01,
-	0x28, 0x0b, 0x32, 0x1a, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74,
-	0x6f, 0x62, 0x75, 0x66, 0x2e, 0x54, 0x69, 0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x52, 0x04,
-	0x64, 0x61, 0x74, 0x65, 0x22, 0x55, 0x0a, 0x1a, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x50, 0x72, 0x6f,
-	0x66, 0x69, 0x74, 0x61, 0x62, 0x69, 0x6c, 0x69, 0x74, 0x69, 0x65, 0x73, 0x50, 0x6f, 0x69, 0x6e,
-	0x74, 0x73, 0x12, 0x37, 0x0a, 0x06, 0x70, 0x6f, 0x69, 0x6e, 0x74, 0x73, 0x18, 0x01, 0x20, 0x03,
-	0x28, 0x0b, 0x32, 0x1f, 0x2e, 0x63, 0x68, 0x61, 0x72, 0x74, 0x73, 0x2e, 0x4d, 0x69, 0x6e, 0x65,
-	0x72, 0x50, 0x72, 0x6f, 0x66, 0x69, 0x74, 0x61, 0x62, 0x69, 0x6c, 0x69, 0x74, 0x79, 0x50, 0x6f,
-	0x69, 0x6e, 0x74, 0x52, 0x06, 0x70, 0x6f, 0x69, 0x6e, 0x74, 0x73, 0x22, 0x84, 0x01, 0x0a, 0x13,
-	0x4d, 0x69, 0x6e, 0x65, 0x72, 0x48, 0x61, 0x73, 0x68, 0x72, 0x61, 0x74, 0x65, 0x73, 0x50, 0x6f,
-	0x69, 0x6e, 0x74, 0x12, 0x1a, 0x0a, 0x08, 0x68, 0x61, 0x73, 0x68, 0x72, 0x61, 0x74, 0x65, 0x18,
-	0x01, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x08, 0x68, 0x61, 0x73, 0x68, 0x72, 0x61, 0x74, 0x65, 0x12,
-	0x21, 0x0a, 0x0c, 0x61, 0x76, 0x67, 0x5f, 0x68, 0x61, 0x73, 0x68, 0x72, 0x61, 0x74, 0x65, 0x18,
-	0x02, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x0b, 0x61, 0x76, 0x67, 0x48, 0x61, 0x73, 0x68, 0x72, 0x61,
-	0x74, 0x65, 0x12, 0x2e, 0x0a, 0x04, 0x64, 0x61, 0x74, 0x65, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0b,
-	0x32, 0x1a, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62,
-	0x75, 0x66, 0x2e, 0x54, 0x69, 0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x52, 0x04, 0x64, 0x61,
-	0x74, 0x65, 0x22, 0x4b, 0x0a, 0x14, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x48, 0x61, 0x73, 0x68, 0x72,
-	0x61, 0x74, 0x65, 0x73, 0x50, 0x6f, 0x69, 0x6e, 0x74, 0x73, 0x12, 0x33, 0x0a, 0x06, 0x70, 0x6f,
-	0x69, 0x6e, 0x74, 0x73, 0x18, 0x01, 0x20, 0x03, 0x28, 0x0b, 0x32, 0x1b, 0x2e, 0x63, 0x68, 0x61,
-	0x72, 0x74, 0x73, 0x2e, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x48, 0x61, 0x73, 0x68, 0x72, 0x61, 0x74,
-	0x65, 0x73, 0x50, 0x6f, 0x69, 0x6e, 0x74, 0x52, 0x06, 0x70, 0x6f, 0x69, 0x6e, 0x74, 0x73, 0x22,
-	0x77, 0x0a, 0x1a, 0x47, 0x65, 0x74, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x57, 0x6f, 0x72, 0x6b, 0x65,
-	0x72, 0x43, 0x68, 0x61, 0x72, 0x74, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x14, 0x0a,
-	0x05, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x05, 0x6d, 0x69,
-	0x6e, 0x65, 0x72, 0x12, 0x16, 0x0a, 0x06, 0x77, 0x6f, 0x72, 0x6b, 0x65, 0x72, 0x18, 0x02, 0x20,
-	0x01, 0x28, 0x09, 0x52, 0x06, 0x77, 0x6f, 0x72, 0x6b, 0x65, 0x72, 0x12, 0x2b, 0x0a, 0x06, 0x70,
-	0x65, 0x72, 0x69, 0x6f, 0x64, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0e, 0x32, 0x13, 0x2e, 0x63, 0x68,
-	0x61, 0x72, 0x74, 0x73, 0x2e, 0x43, 0x68, 0x61, 0x72, 0x74, 0x50, 0x65, 0x72, 0x69, 0x6f, 0x64,
-	0x52, 0x06, 0x70, 0x65, 0x72, 0x69, 0x6f, 0x64, 0x22, 0x91, 0x02, 0x0a, 0x10, 0x4d, 0x69, 0x6e,
-	0x65, 0x72, 0x53, 0x68, 0x61, 0x72, 0x65, 0x73, 0x50, 0x6f, 0x69, 0x6e, 0x74, 0x12, 0x32, 0x0a,
-	0x15, 0x61, 0x63, 0x63, 0x65, 0x70, 0x74, 0x65, 0x64, 0x5f, 0x73, 0x68, 0x61, 0x72, 0x65, 0x73,
-	0x5f, 0x63, 0x6f, 0x75, 0x6e, 0x74, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x13, 0x61, 0x63,
-	0x63, 0x65, 0x70, 0x74, 0x65, 0x64, 0x53, 0x68, 0x61, 0x72, 0x65, 0x73, 0x43, 0x6f, 0x75, 0x6e,
-	0x74, 0x12, 0x32, 0x0a, 0x15, 0x72, 0x65, 0x6a, 0x65, 0x63, 0x74, 0x65, 0x64, 0x5f, 0x73, 0x68,
-	0x61, 0x72, 0x65, 0x73, 0x5f, 0x63, 0x6f, 0x75, 0x6e, 0x74, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0d,
-	0x52, 0x13, 0x72, 0x65, 0x6a, 0x65, 0x63, 0x74, 0x65, 0x64, 0x53, 0x68, 0x61, 0x72, 0x65, 0x73,
-	0x43, 0x6f, 0x75, 0x6e, 0x74, 0x12, 0x2c, 0x0a, 0x12, 0x73, 0x74, 0x61, 0x6c, 0x65, 0x5f, 0x73,
-	0x68, 0x61, 0x72, 0x65, 0x73, 0x5f, 0x63, 0x6f, 0x75, 0x6e, 0x74, 0x18, 0x03, 0x20, 0x01, 0x28,
-	0x0d, 0x52, 0x10, 0x73, 0x74, 0x61, 0x6c, 0x65, 0x53, 0x68, 0x61, 0x72, 0x65, 0x73, 0x43, 0x6f,
-	0x75, 0x6e, 0x74, 0x12, 0x37, 0x0a, 0x18, 0x76, 0x61, 0x6c, 0x69, 0x64, 0x5f, 0x62, 0x6c, 0x6f,
-	0x63, 0x6b, 0x5f, 0x73, 0x68, 0x61, 0x72, 0x65, 0x73, 0x5f, 0x63, 0x6f, 0x75, 0x6e, 0x74, 0x18,
-	0x04, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x15, 0x76, 0x61, 0x6c, 0x69, 0x64, 0x42, 0x6c, 0x6f, 0x63,
-	0x6b, 0x53, 0x68, 0x61, 0x72, 0x65, 0x73, 0x43, 0x6f, 0x75, 0x6e, 0x74, 0x12, 0x2e, 0x0a, 0x04,
-	0x64, 0x61, 0x74, 0x65, 0x18, 0x05, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1a, 0x2e, 0x67, 0x6f, 0x6f,
-	0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x54, 0x69, 0x6d,
-	0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x52, 0x04, 0x64, 0x61, 0x74, 0x65, 0x22, 0x45, 0x0a, 0x11,
-	0x4d, 0x69, 0x6e, 0x65, 0x72, 0x53, 0x68, 0x61, 0x72, 0x65, 0x73, 0x50, 0x6f, 0x69, 0x6e, 0x74,
-	0x73, 0x12, 0x30, 0x0a, 0x06, 0x70, 0x6f, 0x69, 0x6e, 0x74, 0x73, 0x18, 0x01, 0x20, 0x03, 0x28,
-	0x0b, 0x32, 0x18, 0x2e, 0x63, 0x68, 0x61, 0x72, 0x74, 0x73, 0x2e, 0x4d, 0x69, 0x6e, 0x65, 0x72,
-	0x53, 0x68, 0x61, 0x72, 0x65, 0x73, 0x50, 0x6f, 0x69, 0x6e, 0x74, 0x52, 0x06, 0x70, 0x6f, 0x69,
-	0x6e, 0x74, 0x73, 0x2a, 0x35, 0x0a, 0x0b, 0x43, 0x68, 0x61, 0x72, 0x74, 0x50, 0x65, 0x72, 0x69,
-	0x6f, 0x64, 0x12, 0x08, 0x0a, 0x04, 0x48, 0x6f, 0x75, 0x72, 0x10, 0x00, 0x12, 0x07, 0x0a, 0x03,
-	0x44, 0x61, 0x79, 0x10, 0x01, 0x12, 0x08, 0x0a, 0x04, 0x57, 0x65, 0x65, 0x6b, 0x10, 0x02, 0x12,
-	0x09, 0x0a, 0x05, 0x4d, 0x6f, 0x6e, 0x74, 0x68, 0x10, 0x03, 0x32, 0xa3, 0x05, 0x0a, 0x0d, 0x43,
-	0x68, 0x61, 0x72, 0x74, 0x73, 0x53, 0x65, 0x72, 0x76, 0x69, 0x63, 0x65, 0x12, 0x44, 0x0a, 0x0c,
-	0x47, 0x65, 0x74, 0x50, 0x6f, 0x6f, 0x6c, 0x53, 0x74, 0x61, 0x74, 0x73, 0x12, 0x1b, 0x2e, 0x63,
-	0x68, 0x61, 0x72, 0x74, 0x73, 0x2e, 0x47, 0x65, 0x74, 0x50, 0x6f, 0x6f, 0x6c, 0x53, 0x74, 0x61,
-	0x74, 0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x17, 0x2e, 0x63, 0x68, 0x61, 0x72,
-	0x74, 0x73, 0x2e, 0x50, 0x6f, 0x6f, 0x6c, 0x53, 0x74, 0x61, 0x74, 0x73, 0x50, 0x6f, 0x69, 0x6e,
-	0x74, 0x73, 0x12, 0x62, 0x0a, 0x16, 0x47, 0x65, 0x74, 0x4e, 0x65, 0x74, 0x77, 0x6f, 0x72, 0x6b,
-	0x44, 0x69, 0x66, 0x66, 0x69, 0x63, 0x75, 0x6c, 0x74, 0x69, 0x65, 0x73, 0x12, 0x25, 0x2e, 0x63,
-	0x68, 0x61, 0x72, 0x74, 0x73, 0x2e, 0x47, 0x65, 0x74, 0x4e, 0x65, 0x74, 0x77, 0x6f, 0x72, 0x6b,
-	0x44, 0x69, 0x66, 0x66, 0x69, 0x63, 0x75, 0x6c, 0x74, 0x69, 0x65, 0x73, 0x52, 0x65, 0x71, 0x75,
-	0x65, 0x73, 0x74, 0x1a, 0x21, 0x2e, 0x63, 0x68, 0x61, 0x72, 0x74, 0x73, 0x2e, 0x4e, 0x65, 0x74,
-	0x77, 0x6f, 0x72, 0x6b, 0x44, 0x69, 0x66, 0x66, 0x69, 0x63, 0x75, 0x6c, 0x74, 0x69, 0x65, 0x73,
-	0x50, 0x6f, 0x69, 0x6e, 0x74, 0x73, 0x12, 0x3b, 0x0a, 0x09, 0x47, 0x65, 0x74, 0x52, 0x6f, 0x75,
-	0x6e, 0x64, 0x73, 0x12, 0x18, 0x2e, 0x63, 0x68, 0x61, 0x72, 0x74, 0x73, 0x2e, 0x47, 0x65, 0x74,
-	0x52, 0x6f, 0x75, 0x6e, 0x64, 0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x14, 0x2e,
-	0x63, 0x68, 0x61, 0x72, 0x74, 0x73, 0x2e, 0x52, 0x6f, 0x75, 0x6e, 0x64, 0x73, 0x50, 0x6f, 0x69,
-	0x6e, 0x74, 0x73, 0x12, 0x5b, 0x0a, 0x17, 0x47, 0x65, 0x74, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x50,
-	0x72, 0x6f, 0x66, 0x69, 0x74, 0x61, 0x62, 0x69, 0x6c, 0x69, 0x74, 0x69, 0x65, 0x73, 0x12, 0x1c,
-	0x2e, 0x63, 0x68, 0x61, 0x72, 0x74, 0x73, 0x2e, 0x47, 0x65, 0x74, 0x4d, 0x69, 0x6e, 0x65, 0x72,
-	0x43, 0x68, 0x61, 0x72, 0x74, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x22, 0x2e, 0x63,
-	0x68, 0x61, 0x72, 0x74, 0x73, 0x2e, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x50, 0x72, 0x6f, 0x66, 0x69,
-	0x74, 0x61, 0x62, 0x69, 0x6c, 0x69, 0x74, 0x69, 0x65, 0x73, 0x50, 0x6f, 0x69, 0x6e, 0x74, 0x73,
-	0x12, 0x4f, 0x0a, 0x11, 0x47, 0x65, 0x74, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x48, 0x61, 0x73, 0x68,
-	0x72, 0x61, 0x74, 0x65, 0x73, 0x12, 0x1c, 0x2e, 0x63, 0x68, 0x61, 0x72, 0x74, 0x73, 0x2e, 0x47,
-	0x65, 0x74, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x43, 0x68, 0x61, 0x72, 0x74, 0x52, 0x65, 0x71, 0x75,
-	0x65, 0x73, 0x74, 0x1a, 0x1c, 0x2e, 0x63, 0x68, 0x61, 0x72, 0x74, 0x73, 0x2e, 0x4d, 0x69, 0x6e,
-	0x65, 0x72, 0x48, 0x61, 0x73, 0x68, 0x72, 0x61, 0x74, 0x65, 0x73, 0x50, 0x6f, 0x69, 0x6e, 0x74,
-	0x73, 0x12, 0x5b, 0x0a, 0x17, 0x47, 0x65, 0x74, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x57, 0x6f, 0x72,
-	0x6b, 0x65, 0x72, 0x48, 0x61, 0x73, 0x68, 0x72, 0x61, 0x74, 0x65, 0x73, 0x12, 0x22, 0x2e, 0x63,
-	0x68, 0x61, 0x72, 0x74, 0x73, 0x2e, 0x47, 0x65, 0x74, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x57, 0x6f,
-	0x72, 0x6b, 0x65, 0x72, 0x43, 0x68, 0x61, 0x72, 0x74, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74,
-	0x1a, 0x1c, 0x2e, 0x63, 0x68, 0x61, 0x72, 0x74, 0x73, 0x2e, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x48,
-	0x61, 0x73, 0x68, 0x72, 0x61, 0x74, 0x65, 0x73, 0x50, 0x6f, 0x69, 0x6e, 0x74, 0x73, 0x12, 0x49,
-	0x0a, 0x0e, 0x47, 0x65, 0x74, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x53, 0x68, 0x61, 0x72, 0x65, 0x73,
-	0x12, 0x1c, 0x2e, 0x63, 0x68, 0x61, 0x72, 0x74, 0x73, 0x2e, 0x47, 0x65, 0x74, 0x4d, 0x69, 0x6e,
-	0x65, 0x72, 0x43, 0x68, 0x61, 0x72, 0x74, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x19,
-	0x2e, 0x63, 0x68, 0x61, 0x72, 0x74, 0x73, 0x2e, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x53, 0x68, 0x61,
-	0x72, 0x65, 0x73, 0x50, 0x6f, 0x69, 0x6e, 0x74, 0x73, 0x12, 0x55, 0x0a, 0x14, 0x47, 0x65, 0x74,
-	0x4d, 0x69, 0x6e, 0x65, 0x72, 0x57, 0x6f, 0x72, 0x6b, 0x65, 0x72, 0x53, 0x68, 0x61, 0x72, 0x65,
-	0x73, 0x12, 0x22, 0x2e, 0x63, 0x68, 0x61, 0x72, 0x74, 0x73, 0x2e, 0x47, 0x65, 0x74, 0x4d, 0x69,
-	0x6e, 0x65, 0x72, 0x57, 0x6f, 0x72, 0x6b, 0x65, 0x72, 0x43, 0x68, 0x61, 0x72, 0x74, 0x52, 0x65,
-	0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x19, 0x2e, 0x63, 0x68, 0x61, 0x72, 0x74, 0x73, 0x2e, 0x4d,
-	0x69, 0x6e, 0x65, 0x72, 0x53, 0x68, 0x61, 0x72, 0x65, 0x73, 0x50, 0x6f, 0x69, 0x6e, 0x74, 0x73,
-	0x42, 0x48, 0x5a, 0x46, 0x67, 0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d, 0x2f, 0x67,
-	0x72, 0x61, 0x6e, 0x64, 0x6d, 0x69, 0x6e, 0x69, 0x6e, 0x67, 0x70, 0x6f, 0x6f, 0x6c, 0x2f, 0x70,
-	0x6f, 0x6f, 0x6c, 0x2d, 0x61, 0x70, 0x69, 0x2d, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2f, 0x67, 0x65,
-	0x6e, 0x65, 0x72, 0x61, 0x74, 0x65, 0x64, 0x2f, 0x63, 0x68, 0x61, 0x72, 0x74, 0x73, 0x3b, 0x63,
-	0x68, 0x61, 0x72, 0x74, 0x73, 0x50, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74,
-	0x6f, 0x33,
-}
+const file_charts_proto_rawDesc = "" +
+	"\n" +
+	"\fcharts.proto\x12\x06charts\x1a\x1fgoogle/protobuf/timestamp.proto\"V\n" +
+	"\x13GetPoolStatsRequest\x12+\n" +
+	"\x06period\x18\x01 \x01(\x0e2\x13.charts.ChartPeriodR\x06period\x12\x12\n" +
+	"\x04solo\x18\x02 \x01(\bR\x04solo\"\xa2\x01\n" +
+	"\x0ePoolStatsPoint\x12\x1a\n" +
+	"\bhashrate\x18\x01 \x01(\fR\bhashrate\x12!\n" +
+	"\favg_hashrate\x18\x02 \x01(\fR\vavgHashrate\x12!\n" +
+	"\fminers_count\x18\x03 \x01(\rR\vminersCount\x12.\n" +
+	"\x04date\x18\x04 \x01(\v2\x1a.google.protobuf.TimestampR\x04date\"A\n" +
+	"\x0fPoolStatsPoints\x12.\n" +
+	"\x06points\x18\x01 \x03(\v2\x16.charts.PoolStatsPointR\x06points\"L\n" +
+	"\x1dGetNetworkDifficultiesRequest\x12+\n" +
+	"\x06period\x18\x01 \x01(\x0e2\x13.charts.ChartPeriodR\x06period\"\x91\x01\n" +
+	"\x18NetworkDifficultiesPoint\x12\x1e\n" +
+	"\n" +
+	"difficulty\x18\x01 \x01(\fR\n" +
+	"difficulty\x12%\n" +
+	"\x0eavg_difficulty\x18\x02 \x01(\fR\ravgDifficulty\x12.\n" +
+	"\x04date\x18\x03 \x01(\v2\x1a.google.protobuf.TimestampR\x04date\"U\n" +
+	"\x19NetworkDifficultiesPoints\x128\n" +
+	"\x06points\x18\x01 \x03(\v2 .charts.NetworkDifficultiesPointR\x06points\"?\n" +
+	"\x10GetRoundsRequest\x12+\n" +
+	"\x06period\x18\x01 \x01(\x0e2\x13.charts.ChartPeriodR\x06period\"\xb1\x01\n" +
+	"\vRoundsPoint\x12!\n" +
+	"\frounds_count\x18\x01 \x01(\rR\vroundsCount\x12!\n" +
+	"\fminers_count\x18\x02 \x01(\rR\vminersCount\x12,\n" +
+	"\x12avg_round_interval\x18\x03 \x01(\x01R\x10avgRoundInterval\x12.\n" +
+	"\x04date\x18\x04 \x01(\v2\x1a.google.protobuf.TimestampR\x04date\";\n" +
+	"\fRoundsPoints\x12+\n" +
+	"\x06points\x18\x01 \x03(\v2\x13.charts.RoundsPointR\x06points\"{\n" +
+	"\x14GetMinerChartRequest\x12\x14\n" +
+	"\x05miner\x18\x01 \x01(\tR\x05miner\x12\x17\n" +
+	"\x04solo\x18\x02 \x01(\bH\x00R\x04solo\x88\x01\x01\x12+\n" +
+	"\x06period\x18\x03 \x01(\x0e2\x13.charts.ChartPeriodR\x06periodB\a\n" +
+	"\x05_solo\"c\n" +
+	"\x17MinerProfitabilityPoint\x12\x18\n" +
+	"\abalance\x18\x01 \x01(\x04R\abalance\x12.\n" +
+	"\x04date\x18\x02 \x01(\v2\x1a.google.protobuf.TimestampR\x04date\"U\n" +
+	"\x1aMinerProfitabilitiesPoints\x127\n" +
+	"\x06points\x18\x01 \x03(\v2\x1f.charts.MinerProfitabilityPointR\x06points\"\x84\x01\n" +
+	"\x13MinerHashratesPoint\x12\x1a\n" +
+	"\bhashrate\x18\x01 \x01(\fR\bhashrate\x12!\n" +
+	"\favg_hashrate\x18\x02 \x01(\fR\vavgHashrate\x12.\n" +
+	"\x04date\x18\x03 \x01(\v2\x1a.google.protobuf.TimestampR\x04date\"K\n" +
+	"\x14MinerHashratesPoints\x123\n" +
+	"\x06points\x18\x01 \x03(\v2\x1b.charts.MinerHashratesPointR\x06points\"w\n" +
+	"\x1aGetMinerWorkerChartRequest\x12\x14\n" +
+	"\x05miner\x18\x01 \x01(\tR\x05miner\x12\x16\n" +
+	"\x06worker\x18\x02 \x01(\tR\x06worker\x12+\n" +
+	"\x06period\x18\x03 \x01(\x0e2\x13.charts.ChartPeriodR\x06period\"\x91\x02\n" +
+	"\x10MinerSharesPoint\x122\n" +
+	"\x15accepted_shares_count\x18\x01 \x01(\rR\x13acceptedSharesCount\x122\n" +
+	"\x15rejected_shares_count\x18\x02 \x01(\rR\x13rejectedSharesCount\x12,\n" +
+	"\x12stale_shares_count\x18\x03 \x01(\rR\x10staleSharesCount\x127\n" +
+	"\x18valid_block_shares_count\x18\x04 \x01(\rR\x15validBlockSharesCount\x12.\n" +
+	"\x04date\x18\x05 \x01(\v2\x1a.google.protobuf.TimestampR\x04date\"E\n" +
+	"\x11MinerSharesPoints\x120\n" +
+	"\x06points\x18\x01 \x03(\v2\x18.charts.MinerSharesPointR\x06points*5\n" +
+	"\vChartPeriod\x12\b\n" +
+	"\x04Hour\x10\x00\x12\a\n" +
+	"\x03Day\x10\x01\x12\b\n" +
+	"\x04Week\x10\x02\x12\t\n" +
+	"\x05Month\x10\x032\xa3\x05\n" +
+	"\rChartsService\x12D\n" +
+	"\fGetPoolStats\x12\x1b.charts.GetPoolStatsRequest\x1a\x17.charts.PoolStatsPoints\x12b\n" +
+	"\x16GetNetworkDifficulties\x12%.charts.GetNetworkDifficultiesRequest\x1a!.charts.NetworkDifficultiesPoints\x12;\n" +
+	"\tGetRounds\x12\x18.charts.GetRoundsRequest\x1a\x14.charts.RoundsPoints\x12[\n" +
+	"\x17GetMinerProfitabilities\x12\x1c.charts.GetMinerChartRequest\x1a\".charts.MinerProfitabilitiesPoints\x12O\n" +
+	"\x11GetMinerHashrates\x12\x1c.charts.GetMinerChartRequest\x1a\x1c.charts.MinerHashratesPoints\x12[\n" +
+	"\x17GetMinerWorkerHashrates\x12\".charts.GetMinerWorkerChartRequest\x1a\x1c.charts.MinerHashratesPoints\x12I\n" +
+	"\x0eGetMinerShares\x12\x1c.charts.GetMinerChartRequest\x1a\x19.charts.MinerSharesPoints\x12U\n" +
+	"\x14GetMinerWorkerShares\x12\".charts.GetMinerWorkerChartRequest\x1a\x19.charts.MinerSharesPointsBCZAgithub.com/grandminingpool/pool-api-proto/generated/charts;chartsb\x06proto3"
 
 var (
 	file_charts_proto_rawDescOnce sync.Once
-	file_charts_proto_rawDescData = file_charts_proto_rawDesc
+	file_charts_proto_rawDescData []byte
 )
 
 func file_charts_proto_rawDescGZIP() []byte {
 	file_charts_proto_rawDescOnce.Do(func() {
-		file_charts_proto_rawDescData = protoimpl.X.CompressGZIP(file_charts_proto_rawDescData)
+		file_charts_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_charts_proto_rawDesc), len(file_charts_proto_rawDesc)))
 	})
 	return file_charts_proto_rawDescData
 }
 
 var file_charts_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
 var file_charts_proto_msgTypes = make([]protoimpl.MessageInfo, 17)
-var file_charts_proto_goTypes = []interface{}{
+var file_charts_proto_goTypes = []any{
 	(ChartPeriod)(0),                      // 0: charts.ChartPeriod
 	(*GetPoolStatsRequest)(nil),           // 1: charts.GetPoolStatsRequest
 	(*PoolStatsPoint)(nil),                // 2: charts.PoolStatsPoint
@@ -1294,218 +1140,12 @@ func file_charts_proto_init() {
 	if File_charts_proto != nil {
 		return
 	}
-	if !protoimpl.UnsafeEnabled {
-		file_charts_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*GetPoolStatsRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_charts_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*PoolStatsPoint); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_charts_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*PoolStatsPoints); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_charts_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*GetNetworkDifficultiesRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_charts_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*NetworkDifficultiesPoint); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_charts_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*NetworkDifficultiesPoints); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_charts_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*GetRoundsRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_charts_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*RoundsPoint); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_charts_proto_msgTypes[8].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*RoundsPoints); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_charts_proto_msgTypes[9].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*GetMinerChartRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_charts_proto_msgTypes[10].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*MinerProfitabilityPoint); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_charts_proto_msgTypes[11].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*MinerProfitabilitiesPoints); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_charts_proto_msgTypes[12].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*MinerHashratesPoint); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_charts_proto_msgTypes[13].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*MinerHashratesPoints); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_charts_proto_msgTypes[14].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*GetMinerWorkerChartRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_charts_proto_msgTypes[15].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*MinerSharesPoint); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_charts_proto_msgTypes[16].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*MinerSharesPoints); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-	}
-	file_charts_proto_msgTypes[9].OneofWrappers = []interface{}{}
+	file_charts_proto_msgTypes[9].OneofWrappers = []any{}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
-			RawDescriptor: file_charts_proto_rawDesc,
+			RawDescriptor: unsafe.Slice(unsafe.StringData(file_charts_proto_rawDesc), len(file_charts_proto_rawDesc)),
 			NumEnums:      1,
 			NumMessages:   17,
 			NumExtensions: 0,
@@ -1517,7 +1157,6 @@ func file_charts_proto_init() {
 		MessageInfos:      file_charts_proto_msgTypes,
 	}.Build()
 	File_charts_proto = out.File
-	file_charts_proto_rawDesc = nil
 	file_charts_proto_goTypes = nil
 	file_charts_proto_depIdxs = nil
 }
