@@ -1,10 +1,10 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.28.1
-// 	protoc        v5.26.1
+// 	protoc-gen-go v1.36.6
+// 	protoc        v5.29.3
 // source: pool.proto
 
-package poolProto
+package pool
 
 import (
 	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
@@ -13,6 +13,7 @@ import (
 	timestamppb "google.golang.org/protobuf/types/known/timestamppb"
 	reflect "reflect"
 	sync "sync"
+	unsafe "unsafe"
 )
 
 const (
@@ -115,22 +116,19 @@ func (AddressFormat) EnumDescriptor() ([]byte, []int) {
 }
 
 type PayoutsInfo struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Interval      uint32                 `protobuf:"varint,1,opt,name=interval,proto3" json:"interval,omitempty"`
+	MinPayout     uint64                 `protobuf:"varint,2,opt,name=min_payout,json=minPayout,proto3" json:"min_payout,omitempty"`
+	MaxPayout     *uint64                `protobuf:"varint,3,opt,name=max_payout,json=maxPayout,proto3,oneof" json:"max_payout,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Interval  uint32  `protobuf:"varint,1,opt,name=interval,proto3" json:"interval,omitempty"`
-	MinPayout uint64  `protobuf:"varint,2,opt,name=min_payout,json=minPayout,proto3" json:"min_payout,omitempty"`
-	MaxPayout *uint64 `protobuf:"varint,3,opt,name=max_payout,json=maxPayout,proto3,oneof" json:"max_payout,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *PayoutsInfo) Reset() {
 	*x = PayoutsInfo{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *PayoutsInfo) String() string {
@@ -141,7 +139,7 @@ func (*PayoutsInfo) ProtoMessage() {}
 
 func (x *PayoutsInfo) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -178,21 +176,18 @@ func (x *PayoutsInfo) GetMaxPayout() uint64 {
 }
 
 type PoolFee struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Fee           float64                `protobuf:"fixed64,1,opt,name=fee,proto3" json:"fee,omitempty"`
+	SoloFee       *float64               `protobuf:"fixed64,2,opt,name=solo_fee,json=soloFee,proto3,oneof" json:"solo_fee,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Fee     float64  `protobuf:"fixed64,1,opt,name=fee,proto3" json:"fee,omitempty"`
-	SoloFee *float64 `protobuf:"fixed64,2,opt,name=solo_fee,json=soloFee,proto3,oneof" json:"solo_fee,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *PoolFee) Reset() {
 	*x = PoolFee{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *PoolFee) String() string {
@@ -203,7 +198,7 @@ func (*PoolFee) ProtoMessage() {}
 
 func (x *PoolFee) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -233,24 +228,21 @@ func (x *PoolFee) GetSoloFee() float64 {
 }
 
 type PoolSlave struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Region        string                 `protobuf:"bytes,1,opt,name=region,proto3" json:"region,omitempty"`
+	Host          string                 `protobuf:"bytes,2,opt,name=host,proto3" json:"host,omitempty"`
+	TcpPort       uint32                 `protobuf:"varint,3,opt,name=tcp_port,json=tcpPort,proto3" json:"tcp_port,omitempty"`
+	SslPort       uint32                 `protobuf:"varint,4,opt,name=ssl_port,json=sslPort,proto3" json:"ssl_port,omitempty"`
+	ConnectedAt   *timestamppb.Timestamp `protobuf:"bytes,6,opt,name=connected_at,json=connectedAt,proto3" json:"connected_at,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Region      string                 `protobuf:"bytes,1,opt,name=region,proto3" json:"region,omitempty"`
-	Host        string                 `protobuf:"bytes,2,opt,name=host,proto3" json:"host,omitempty"`
-	TcpPort     uint32                 `protobuf:"varint,3,opt,name=tcp_port,json=tcpPort,proto3" json:"tcp_port,omitempty"`
-	SslPort     uint32                 `protobuf:"varint,4,opt,name=ssl_port,json=sslPort,proto3" json:"ssl_port,omitempty"`
-	ConnectedAt *timestamppb.Timestamp `protobuf:"bytes,6,opt,name=connected_at,json=connectedAt,proto3" json:"connected_at,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *PoolSlave) Reset() {
 	*x = PoolSlave{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_proto_msgTypes[2]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_proto_msgTypes[2]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *PoolSlave) String() string {
@@ -261,7 +253,7 @@ func (*PoolSlave) ProtoMessage() {}
 
 func (x *PoolSlave) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_proto_msgTypes[2]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -312,28 +304,25 @@ func (x *PoolSlave) GetConnectedAt() *timestamppb.Timestamp {
 }
 
 type PoolInfo struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Blockchain    string                 `protobuf:"bytes,1,opt,name=blockchain,proto3" json:"blockchain,omitempty"`
+	AddressFormat AddressFormat          `protobuf:"varint,2,opt,name=address_format,json=addressFormat,proto3,enum=pool.AddressFormat" json:"address_format,omitempty"`
+	Host          string                 `protobuf:"bytes,3,opt,name=host,proto3" json:"host,omitempty"`
+	Algos         []string               `protobuf:"bytes,4,rep,name=algos,proto3" json:"algos,omitempty"`
+	PayoutMode    PayoutMode             `protobuf:"varint,5,opt,name=payout_mode,json=payoutMode,proto3,enum=pool.PayoutMode" json:"payout_mode,omitempty"`
+	Solo          bool                   `protobuf:"varint,6,opt,name=solo,proto3" json:"solo,omitempty"`
+	Fee           *PoolFee               `protobuf:"bytes,7,opt,name=fee,proto3" json:"fee,omitempty"`
+	PayoutsInfo   *PayoutsInfo           `protobuf:"bytes,8,opt,name=payouts_info,json=payoutsInfo,proto3" json:"payouts_info,omitempty"`
+	Agents        []string               `protobuf:"bytes,9,rep,name=agents,proto3" json:"agents,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Blockchain    string        `protobuf:"bytes,1,opt,name=blockchain,proto3" json:"blockchain,omitempty"`
-	AddressFormat AddressFormat `protobuf:"varint,2,opt,name=address_format,json=addressFormat,proto3,enum=pool.AddressFormat" json:"address_format,omitempty"`
-	Host          string        `protobuf:"bytes,3,opt,name=host,proto3" json:"host,omitempty"`
-	Algos         []string      `protobuf:"bytes,4,rep,name=algos,proto3" json:"algos,omitempty"`
-	PayoutMode    PayoutMode    `protobuf:"varint,5,opt,name=payout_mode,json=payoutMode,proto3,enum=pool.PayoutMode" json:"payout_mode,omitempty"`
-	Solo          bool          `protobuf:"varint,6,opt,name=solo,proto3" json:"solo,omitempty"`
-	Fee           *PoolFee      `protobuf:"bytes,7,opt,name=fee,proto3" json:"fee,omitempty"`
-	PayoutsInfo   *PayoutsInfo  `protobuf:"bytes,8,opt,name=payouts_info,json=payoutsInfo,proto3" json:"payouts_info,omitempty"`
-	Agents        []string      `protobuf:"bytes,9,rep,name=agents,proto3" json:"agents,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *PoolInfo) Reset() {
 	*x = PoolInfo{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_proto_msgTypes[3]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_proto_msgTypes[3]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *PoolInfo) String() string {
@@ -344,7 +333,7 @@ func (*PoolInfo) ProtoMessage() {}
 
 func (x *PoolInfo) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_proto_msgTypes[3]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -423,22 +412,19 @@ func (x *PoolInfo) GetAgents() []string {
 }
 
 type NetworkInfo struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Height        uint64                 `protobuf:"varint,1,opt,name=height,proto3" json:"height,omitempty"`
+	BlockReward   uint64                 `protobuf:"varint,2,opt,name=block_reward,json=blockReward,proto3" json:"block_reward,omitempty"`
+	Difficulty    []byte                 `protobuf:"bytes,3,opt,name=difficulty,proto3" json:"difficulty,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Height      uint64 `protobuf:"varint,1,opt,name=height,proto3" json:"height,omitempty"`
-	BlockReward uint64 `protobuf:"varint,2,opt,name=block_reward,json=blockReward,proto3" json:"block_reward,omitempty"`
-	Difficulty  []byte `protobuf:"bytes,3,opt,name=difficulty,proto3" json:"difficulty,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *NetworkInfo) Reset() {
 	*x = NetworkInfo{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_proto_msgTypes[4]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_proto_msgTypes[4]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *NetworkInfo) String() string {
@@ -449,7 +435,7 @@ func (*NetworkInfo) ProtoMessage() {}
 
 func (x *NetworkInfo) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_proto_msgTypes[4]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -486,20 +472,17 @@ func (x *NetworkInfo) GetDifficulty() []byte {
 }
 
 type PoolSlaves struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Slaves        []*PoolSlave           `protobuf:"bytes,1,rep,name=slaves,proto3" json:"slaves,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Slaves []*PoolSlave `protobuf:"bytes,1,rep,name=slaves,proto3" json:"slaves,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *PoolSlaves) Reset() {
 	*x = PoolSlaves{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_proto_msgTypes[5]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_proto_msgTypes[5]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *PoolSlaves) String() string {
@@ -510,7 +493,7 @@ func (*PoolSlaves) ProtoMessage() {}
 
 func (x *PoolSlaves) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_proto_msgTypes[5]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -533,20 +516,17 @@ func (x *PoolSlaves) GetSlaves() []*PoolSlave {
 }
 
 type GetPoolAssetRequest struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Solo          bool                   `protobuf:"varint,1,opt,name=solo,proto3" json:"solo,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Solo bool `protobuf:"varint,1,opt,name=solo,proto3" json:"solo,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *GetPoolAssetRequest) Reset() {
 	*x = GetPoolAssetRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_proto_msgTypes[6]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_proto_msgTypes[6]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *GetPoolAssetRequest) String() string {
@@ -557,7 +537,7 @@ func (*GetPoolAssetRequest) ProtoMessage() {}
 
 func (x *GetPoolAssetRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_proto_msgTypes[6]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -580,24 +560,21 @@ func (x *GetPoolAssetRequest) GetSolo() bool {
 }
 
 type PoolStats struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	MinersCount     uint32 `protobuf:"varint,1,opt,name=miners_count,json=minersCount,proto3" json:"miners_count,omitempty"`
-	WorkersCount    uint32 `protobuf:"varint,2,opt,name=workers_count,json=workersCount,proto3" json:"workers_count,omitempty"`
-	Hashrate        []byte `protobuf:"bytes,3,opt,name=hashrate,proto3" json:"hashrate,omitempty"`
-	AvgHashrate     []byte `protobuf:"bytes,4,opt,name=avg_hashrate,json=avgHashrate,proto3" json:"avg_hashrate,omitempty"`
-	BlocksCount_24H uint32 `protobuf:"varint,5,opt,name=blocks_count_24h,json=blocksCount24h,proto3" json:"blocks_count_24h,omitempty"`
+	state           protoimpl.MessageState `protogen:"open.v1"`
+	MinersCount     uint32                 `protobuf:"varint,1,opt,name=miners_count,json=minersCount,proto3" json:"miners_count,omitempty"`
+	WorkersCount    uint32                 `protobuf:"varint,2,opt,name=workers_count,json=workersCount,proto3" json:"workers_count,omitempty"`
+	Hashrate        []byte                 `protobuf:"bytes,3,opt,name=hashrate,proto3" json:"hashrate,omitempty"`
+	AvgHashrate     []byte                 `protobuf:"bytes,4,opt,name=avg_hashrate,json=avgHashrate,proto3" json:"avg_hashrate,omitempty"`
+	BlocksCount_24H uint32                 `protobuf:"varint,5,opt,name=blocks_count_24h,json=blocksCount24h,proto3" json:"blocks_count_24h,omitempty"`
+	unknownFields   protoimpl.UnknownFields
+	sizeCache       protoimpl.SizeCache
 }
 
 func (x *PoolStats) Reset() {
 	*x = PoolStats{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_proto_msgTypes[7]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_proto_msgTypes[7]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *PoolStats) String() string {
@@ -608,7 +585,7 @@ func (*PoolStats) ProtoMessage() {}
 
 func (x *PoolStats) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_proto_msgTypes[7]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -660,123 +637,86 @@ func (x *PoolStats) GetBlocksCount_24H() uint32 {
 
 var File_pool_proto protoreflect.FileDescriptor
 
-var file_pool_proto_rawDesc = []byte{
-	0x0a, 0x0a, 0x70, 0x6f, 0x6f, 0x6c, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12, 0x04, 0x70, 0x6f,
-	0x6f, 0x6c, 0x1a, 0x1b, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2f, 0x70, 0x72, 0x6f, 0x74, 0x6f,
-	0x62, 0x75, 0x66, 0x2f, 0x65, 0x6d, 0x70, 0x74, 0x79, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x1a,
-	0x1f, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2f, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66,
-	0x2f, 0x74, 0x69, 0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f,
-	0x22, 0x7b, 0x0a, 0x0b, 0x50, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73, 0x49, 0x6e, 0x66, 0x6f, 0x12,
-	0x1a, 0x0a, 0x08, 0x69, 0x6e, 0x74, 0x65, 0x72, 0x76, 0x61, 0x6c, 0x18, 0x01, 0x20, 0x01, 0x28,
-	0x0d, 0x52, 0x08, 0x69, 0x6e, 0x74, 0x65, 0x72, 0x76, 0x61, 0x6c, 0x12, 0x1d, 0x0a, 0x0a, 0x6d,
-	0x69, 0x6e, 0x5f, 0x70, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x18, 0x02, 0x20, 0x01, 0x28, 0x04, 0x52,
-	0x09, 0x6d, 0x69, 0x6e, 0x50, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x12, 0x22, 0x0a, 0x0a, 0x6d, 0x61,
-	0x78, 0x5f, 0x70, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x18, 0x03, 0x20, 0x01, 0x28, 0x04, 0x48, 0x00,
-	0x52, 0x09, 0x6d, 0x61, 0x78, 0x50, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x88, 0x01, 0x01, 0x42, 0x0d,
-	0x0a, 0x0b, 0x5f, 0x6d, 0x61, 0x78, 0x5f, 0x70, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x22, 0x48, 0x0a,
-	0x07, 0x50, 0x6f, 0x6f, 0x6c, 0x46, 0x65, 0x65, 0x12, 0x10, 0x0a, 0x03, 0x66, 0x65, 0x65, 0x18,
-	0x01, 0x20, 0x01, 0x28, 0x01, 0x52, 0x03, 0x66, 0x65, 0x65, 0x12, 0x1e, 0x0a, 0x08, 0x73, 0x6f,
-	0x6c, 0x6f, 0x5f, 0x66, 0x65, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x01, 0x48, 0x00, 0x52, 0x07,
-	0x73, 0x6f, 0x6c, 0x6f, 0x46, 0x65, 0x65, 0x88, 0x01, 0x01, 0x42, 0x0b, 0x0a, 0x09, 0x5f, 0x73,
-	0x6f, 0x6c, 0x6f, 0x5f, 0x66, 0x65, 0x65, 0x22, 0xac, 0x01, 0x0a, 0x09, 0x50, 0x6f, 0x6f, 0x6c,
-	0x53, 0x6c, 0x61, 0x76, 0x65, 0x12, 0x16, 0x0a, 0x06, 0x72, 0x65, 0x67, 0x69, 0x6f, 0x6e, 0x18,
-	0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x06, 0x72, 0x65, 0x67, 0x69, 0x6f, 0x6e, 0x12, 0x12, 0x0a,
-	0x04, 0x68, 0x6f, 0x73, 0x74, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x04, 0x68, 0x6f, 0x73,
-	0x74, 0x12, 0x19, 0x0a, 0x08, 0x74, 0x63, 0x70, 0x5f, 0x70, 0x6f, 0x72, 0x74, 0x18, 0x03, 0x20,
-	0x01, 0x28, 0x0d, 0x52, 0x07, 0x74, 0x63, 0x70, 0x50, 0x6f, 0x72, 0x74, 0x12, 0x19, 0x0a, 0x08,
-	0x73, 0x73, 0x6c, 0x5f, 0x70, 0x6f, 0x72, 0x74, 0x18, 0x04, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x07,
-	0x73, 0x73, 0x6c, 0x50, 0x6f, 0x72, 0x74, 0x12, 0x3d, 0x0a, 0x0c, 0x63, 0x6f, 0x6e, 0x6e, 0x65,
-	0x63, 0x74, 0x65, 0x64, 0x5f, 0x61, 0x74, 0x18, 0x06, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1a, 0x2e,
-	0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e,
-	0x54, 0x69, 0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x52, 0x0b, 0x63, 0x6f, 0x6e, 0x6e, 0x65,
-	0x63, 0x74, 0x65, 0x64, 0x41, 0x74, 0x22, 0xc6, 0x02, 0x0a, 0x08, 0x50, 0x6f, 0x6f, 0x6c, 0x49,
-	0x6e, 0x66, 0x6f, 0x12, 0x1e, 0x0a, 0x0a, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x63, 0x68, 0x61, 0x69,
-	0x6e, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x0a, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x63, 0x68,
-	0x61, 0x69, 0x6e, 0x12, 0x3a, 0x0a, 0x0e, 0x61, 0x64, 0x64, 0x72, 0x65, 0x73, 0x73, 0x5f, 0x66,
-	0x6f, 0x72, 0x6d, 0x61, 0x74, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0e, 0x32, 0x13, 0x2e, 0x70, 0x6f,
-	0x6f, 0x6c, 0x2e, 0x41, 0x64, 0x64, 0x72, 0x65, 0x73, 0x73, 0x46, 0x6f, 0x72, 0x6d, 0x61, 0x74,
-	0x52, 0x0d, 0x61, 0x64, 0x64, 0x72, 0x65, 0x73, 0x73, 0x46, 0x6f, 0x72, 0x6d, 0x61, 0x74, 0x12,
-	0x12, 0x0a, 0x04, 0x68, 0x6f, 0x73, 0x74, 0x18, 0x03, 0x20, 0x01, 0x28, 0x09, 0x52, 0x04, 0x68,
-	0x6f, 0x73, 0x74, 0x12, 0x14, 0x0a, 0x05, 0x61, 0x6c, 0x67, 0x6f, 0x73, 0x18, 0x04, 0x20, 0x03,
-	0x28, 0x09, 0x52, 0x05, 0x61, 0x6c, 0x67, 0x6f, 0x73, 0x12, 0x31, 0x0a, 0x0b, 0x70, 0x61, 0x79,
-	0x6f, 0x75, 0x74, 0x5f, 0x6d, 0x6f, 0x64, 0x65, 0x18, 0x05, 0x20, 0x01, 0x28, 0x0e, 0x32, 0x10,
-	0x2e, 0x70, 0x6f, 0x6f, 0x6c, 0x2e, 0x50, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x4d, 0x6f, 0x64, 0x65,
-	0x52, 0x0a, 0x70, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x4d, 0x6f, 0x64, 0x65, 0x12, 0x12, 0x0a, 0x04,
-	0x73, 0x6f, 0x6c, 0x6f, 0x18, 0x06, 0x20, 0x01, 0x28, 0x08, 0x52, 0x04, 0x73, 0x6f, 0x6c, 0x6f,
-	0x12, 0x1f, 0x0a, 0x03, 0x66, 0x65, 0x65, 0x18, 0x07, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x0d, 0x2e,
-	0x70, 0x6f, 0x6f, 0x6c, 0x2e, 0x50, 0x6f, 0x6f, 0x6c, 0x46, 0x65, 0x65, 0x52, 0x03, 0x66, 0x65,
-	0x65, 0x12, 0x34, 0x0a, 0x0c, 0x70, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73, 0x5f, 0x69, 0x6e, 0x66,
-	0x6f, 0x18, 0x08, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x11, 0x2e, 0x70, 0x6f, 0x6f, 0x6c, 0x2e, 0x50,
-	0x61, 0x79, 0x6f, 0x75, 0x74, 0x73, 0x49, 0x6e, 0x66, 0x6f, 0x52, 0x0b, 0x70, 0x61, 0x79, 0x6f,
-	0x75, 0x74, 0x73, 0x49, 0x6e, 0x66, 0x6f, 0x12, 0x16, 0x0a, 0x06, 0x61, 0x67, 0x65, 0x6e, 0x74,
-	0x73, 0x18, 0x09, 0x20, 0x03, 0x28, 0x09, 0x52, 0x06, 0x61, 0x67, 0x65, 0x6e, 0x74, 0x73, 0x22,
-	0x68, 0x0a, 0x0b, 0x4e, 0x65, 0x74, 0x77, 0x6f, 0x72, 0x6b, 0x49, 0x6e, 0x66, 0x6f, 0x12, 0x16,
-	0x0a, 0x06, 0x68, 0x65, 0x69, 0x67, 0x68, 0x74, 0x18, 0x01, 0x20, 0x01, 0x28, 0x04, 0x52, 0x06,
-	0x68, 0x65, 0x69, 0x67, 0x68, 0x74, 0x12, 0x21, 0x0a, 0x0c, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x5f,
-	0x72, 0x65, 0x77, 0x61, 0x72, 0x64, 0x18, 0x02, 0x20, 0x01, 0x28, 0x04, 0x52, 0x0b, 0x62, 0x6c,
-	0x6f, 0x63, 0x6b, 0x52, 0x65, 0x77, 0x61, 0x72, 0x64, 0x12, 0x1e, 0x0a, 0x0a, 0x64, 0x69, 0x66,
-	0x66, 0x69, 0x63, 0x75, 0x6c, 0x74, 0x79, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x0a, 0x64,
-	0x69, 0x66, 0x66, 0x69, 0x63, 0x75, 0x6c, 0x74, 0x79, 0x22, 0x35, 0x0a, 0x0a, 0x50, 0x6f, 0x6f,
-	0x6c, 0x53, 0x6c, 0x61, 0x76, 0x65, 0x73, 0x12, 0x27, 0x0a, 0x06, 0x73, 0x6c, 0x61, 0x76, 0x65,
-	0x73, 0x18, 0x01, 0x20, 0x03, 0x28, 0x0b, 0x32, 0x0f, 0x2e, 0x70, 0x6f, 0x6f, 0x6c, 0x2e, 0x50,
-	0x6f, 0x6f, 0x6c, 0x53, 0x6c, 0x61, 0x76, 0x65, 0x52, 0x06, 0x73, 0x6c, 0x61, 0x76, 0x65, 0x73,
-	0x22, 0x29, 0x0a, 0x13, 0x47, 0x65, 0x74, 0x50, 0x6f, 0x6f, 0x6c, 0x41, 0x73, 0x73, 0x65, 0x74,
-	0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x12, 0x0a, 0x04, 0x73, 0x6f, 0x6c, 0x6f, 0x18,
-	0x01, 0x20, 0x01, 0x28, 0x08, 0x52, 0x04, 0x73, 0x6f, 0x6c, 0x6f, 0x22, 0xbc, 0x01, 0x0a, 0x09,
-	0x50, 0x6f, 0x6f, 0x6c, 0x53, 0x74, 0x61, 0x74, 0x73, 0x12, 0x21, 0x0a, 0x0c, 0x6d, 0x69, 0x6e,
-	0x65, 0x72, 0x73, 0x5f, 0x63, 0x6f, 0x75, 0x6e, 0x74, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0d, 0x52,
-	0x0b, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x73, 0x43, 0x6f, 0x75, 0x6e, 0x74, 0x12, 0x23, 0x0a, 0x0d,
-	0x77, 0x6f, 0x72, 0x6b, 0x65, 0x72, 0x73, 0x5f, 0x63, 0x6f, 0x75, 0x6e, 0x74, 0x18, 0x02, 0x20,
-	0x01, 0x28, 0x0d, 0x52, 0x0c, 0x77, 0x6f, 0x72, 0x6b, 0x65, 0x72, 0x73, 0x43, 0x6f, 0x75, 0x6e,
-	0x74, 0x12, 0x1a, 0x0a, 0x08, 0x68, 0x61, 0x73, 0x68, 0x72, 0x61, 0x74, 0x65, 0x18, 0x03, 0x20,
-	0x01, 0x28, 0x0c, 0x52, 0x08, 0x68, 0x61, 0x73, 0x68, 0x72, 0x61, 0x74, 0x65, 0x12, 0x21, 0x0a,
-	0x0c, 0x61, 0x76, 0x67, 0x5f, 0x68, 0x61, 0x73, 0x68, 0x72, 0x61, 0x74, 0x65, 0x18, 0x04, 0x20,
-	0x01, 0x28, 0x0c, 0x52, 0x0b, 0x61, 0x76, 0x67, 0x48, 0x61, 0x73, 0x68, 0x72, 0x61, 0x74, 0x65,
-	0x12, 0x28, 0x0a, 0x10, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x5f, 0x63, 0x6f, 0x75, 0x6e, 0x74,
-	0x5f, 0x32, 0x34, 0x68, 0x18, 0x05, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x0e, 0x62, 0x6c, 0x6f, 0x63,
-	0x6b, 0x73, 0x43, 0x6f, 0x75, 0x6e, 0x74, 0x32, 0x34, 0x68, 0x2a, 0x21, 0x0a, 0x0a, 0x50, 0x61,
-	0x79, 0x6f, 0x75, 0x74, 0x4d, 0x6f, 0x64, 0x65, 0x12, 0x09, 0x0a, 0x05, 0x50, 0x50, 0x4c, 0x4e,
-	0x53, 0x10, 0x00, 0x12, 0x08, 0x0a, 0x04, 0x50, 0x52, 0x4f, 0x50, 0x10, 0x01, 0x2a, 0x25, 0x0a,
-	0x0d, 0x41, 0x64, 0x64, 0x72, 0x65, 0x73, 0x73, 0x46, 0x6f, 0x72, 0x6d, 0x61, 0x74, 0x12, 0x0a,
-	0x0a, 0x06, 0x4e, 0x6f, 0x72, 0x6d, 0x61, 0x6c, 0x10, 0x00, 0x12, 0x08, 0x0a, 0x04, 0x4d, 0x65,
-	0x6d, 0x6f, 0x10, 0x01, 0x32, 0xfb, 0x01, 0x0a, 0x0b, 0x50, 0x6f, 0x6f, 0x6c, 0x53, 0x65, 0x72,
-	0x76, 0x69, 0x63, 0x65, 0x12, 0x35, 0x0a, 0x0b, 0x47, 0x65, 0x74, 0x50, 0x6f, 0x6f, 0x6c, 0x49,
-	0x6e, 0x66, 0x6f, 0x12, 0x16, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f,
-	0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x45, 0x6d, 0x70, 0x74, 0x79, 0x1a, 0x0e, 0x2e, 0x70, 0x6f,
-	0x6f, 0x6c, 0x2e, 0x50, 0x6f, 0x6f, 0x6c, 0x49, 0x6e, 0x66, 0x6f, 0x12, 0x3b, 0x0a, 0x0e, 0x47,
-	0x65, 0x74, 0x4e, 0x65, 0x74, 0x77, 0x6f, 0x72, 0x6b, 0x49, 0x6e, 0x66, 0x6f, 0x12, 0x16, 0x2e,
-	0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e,
-	0x45, 0x6d, 0x70, 0x74, 0x79, 0x1a, 0x11, 0x2e, 0x70, 0x6f, 0x6f, 0x6c, 0x2e, 0x4e, 0x65, 0x74,
-	0x77, 0x6f, 0x72, 0x6b, 0x49, 0x6e, 0x66, 0x6f, 0x12, 0x3c, 0x0a, 0x0d, 0x47, 0x65, 0x74, 0x50,
-	0x6f, 0x6f, 0x6c, 0x53, 0x6c, 0x61, 0x76, 0x65, 0x73, 0x12, 0x19, 0x2e, 0x70, 0x6f, 0x6f, 0x6c,
-	0x2e, 0x47, 0x65, 0x74, 0x50, 0x6f, 0x6f, 0x6c, 0x41, 0x73, 0x73, 0x65, 0x74, 0x52, 0x65, 0x71,
-	0x75, 0x65, 0x73, 0x74, 0x1a, 0x10, 0x2e, 0x70, 0x6f, 0x6f, 0x6c, 0x2e, 0x50, 0x6f, 0x6f, 0x6c,
-	0x53, 0x6c, 0x61, 0x76, 0x65, 0x73, 0x12, 0x3a, 0x0a, 0x0c, 0x47, 0x65, 0x74, 0x50, 0x6f, 0x6f,
-	0x6c, 0x53, 0x74, 0x61, 0x74, 0x73, 0x12, 0x19, 0x2e, 0x70, 0x6f, 0x6f, 0x6c, 0x2e, 0x47, 0x65,
-	0x74, 0x50, 0x6f, 0x6f, 0x6c, 0x41, 0x73, 0x73, 0x65, 0x74, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73,
-	0x74, 0x1a, 0x0f, 0x2e, 0x70, 0x6f, 0x6f, 0x6c, 0x2e, 0x50, 0x6f, 0x6f, 0x6c, 0x53, 0x74, 0x61,
-	0x74, 0x73, 0x42, 0x44, 0x5a, 0x42, 0x67, 0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d,
-	0x2f, 0x67, 0x72, 0x61, 0x6e, 0x64, 0x6d, 0x69, 0x6e, 0x69, 0x6e, 0x67, 0x70, 0x6f, 0x6f, 0x6c,
-	0x2f, 0x70, 0x6f, 0x6f, 0x6c, 0x2d, 0x61, 0x70, 0x69, 0x2d, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2f,
-	0x67, 0x65, 0x6e, 0x65, 0x72, 0x61, 0x74, 0x65, 0x64, 0x2f, 0x70, 0x6f, 0x6f, 0x6c, 0x3b, 0x70,
-	0x6f, 0x6f, 0x6c, 0x50, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
-}
+const file_pool_proto_rawDesc = "" +
+	"\n" +
+	"\n" +
+	"pool.proto\x12\x04pool\x1a\x1bgoogle/protobuf/empty.proto\x1a\x1fgoogle/protobuf/timestamp.proto\"{\n" +
+	"\vPayoutsInfo\x12\x1a\n" +
+	"\binterval\x18\x01 \x01(\rR\binterval\x12\x1d\n" +
+	"\n" +
+	"min_payout\x18\x02 \x01(\x04R\tminPayout\x12\"\n" +
+	"\n" +
+	"max_payout\x18\x03 \x01(\x04H\x00R\tmaxPayout\x88\x01\x01B\r\n" +
+	"\v_max_payout\"H\n" +
+	"\aPoolFee\x12\x10\n" +
+	"\x03fee\x18\x01 \x01(\x01R\x03fee\x12\x1e\n" +
+	"\bsolo_fee\x18\x02 \x01(\x01H\x00R\asoloFee\x88\x01\x01B\v\n" +
+	"\t_solo_fee\"\xac\x01\n" +
+	"\tPoolSlave\x12\x16\n" +
+	"\x06region\x18\x01 \x01(\tR\x06region\x12\x12\n" +
+	"\x04host\x18\x02 \x01(\tR\x04host\x12\x19\n" +
+	"\btcp_port\x18\x03 \x01(\rR\atcpPort\x12\x19\n" +
+	"\bssl_port\x18\x04 \x01(\rR\asslPort\x12=\n" +
+	"\fconnected_at\x18\x06 \x01(\v2\x1a.google.protobuf.TimestampR\vconnectedAt\"\xc6\x02\n" +
+	"\bPoolInfo\x12\x1e\n" +
+	"\n" +
+	"blockchain\x18\x01 \x01(\tR\n" +
+	"blockchain\x12:\n" +
+	"\x0eaddress_format\x18\x02 \x01(\x0e2\x13.pool.AddressFormatR\raddressFormat\x12\x12\n" +
+	"\x04host\x18\x03 \x01(\tR\x04host\x12\x14\n" +
+	"\x05algos\x18\x04 \x03(\tR\x05algos\x121\n" +
+	"\vpayout_mode\x18\x05 \x01(\x0e2\x10.pool.PayoutModeR\n" +
+	"payoutMode\x12\x12\n" +
+	"\x04solo\x18\x06 \x01(\bR\x04solo\x12\x1f\n" +
+	"\x03fee\x18\a \x01(\v2\r.pool.PoolFeeR\x03fee\x124\n" +
+	"\fpayouts_info\x18\b \x01(\v2\x11.pool.PayoutsInfoR\vpayoutsInfo\x12\x16\n" +
+	"\x06agents\x18\t \x03(\tR\x06agents\"h\n" +
+	"\vNetworkInfo\x12\x16\n" +
+	"\x06height\x18\x01 \x01(\x04R\x06height\x12!\n" +
+	"\fblock_reward\x18\x02 \x01(\x04R\vblockReward\x12\x1e\n" +
+	"\n" +
+	"difficulty\x18\x03 \x01(\fR\n" +
+	"difficulty\"5\n" +
+	"\n" +
+	"PoolSlaves\x12'\n" +
+	"\x06slaves\x18\x01 \x03(\v2\x0f.pool.PoolSlaveR\x06slaves\")\n" +
+	"\x13GetPoolAssetRequest\x12\x12\n" +
+	"\x04solo\x18\x01 \x01(\bR\x04solo\"\xbc\x01\n" +
+	"\tPoolStats\x12!\n" +
+	"\fminers_count\x18\x01 \x01(\rR\vminersCount\x12#\n" +
+	"\rworkers_count\x18\x02 \x01(\rR\fworkersCount\x12\x1a\n" +
+	"\bhashrate\x18\x03 \x01(\fR\bhashrate\x12!\n" +
+	"\favg_hashrate\x18\x04 \x01(\fR\vavgHashrate\x12(\n" +
+	"\x10blocks_count_24h\x18\x05 \x01(\rR\x0eblocksCount24h*!\n" +
+	"\n" +
+	"PayoutMode\x12\t\n" +
+	"\x05PPLNS\x10\x00\x12\b\n" +
+	"\x04PROP\x10\x01*%\n" +
+	"\rAddressFormat\x12\n" +
+	"\n" +
+	"\x06Normal\x10\x00\x12\b\n" +
+	"\x04Memo\x10\x012\xfb\x01\n" +
+	"\vPoolService\x125\n" +
+	"\vGetPoolInfo\x12\x16.google.protobuf.Empty\x1a\x0e.pool.PoolInfo\x12;\n" +
+	"\x0eGetNetworkInfo\x12\x16.google.protobuf.Empty\x1a\x11.pool.NetworkInfo\x12<\n" +
+	"\rGetPoolSlaves\x12\x19.pool.GetPoolAssetRequest\x1a\x10.pool.PoolSlaves\x12:\n" +
+	"\fGetPoolStats\x12\x19.pool.GetPoolAssetRequest\x1a\x0f.pool.PoolStatsB?Z=github.com/grandminingpool/pool-api-proto/generated/pool;poolb\x06proto3"
 
 var (
 	file_pool_proto_rawDescOnce sync.Once
-	file_pool_proto_rawDescData = file_pool_proto_rawDesc
+	file_pool_proto_rawDescData []byte
 )
 
 func file_pool_proto_rawDescGZIP() []byte {
 	file_pool_proto_rawDescOnce.Do(func() {
-		file_pool_proto_rawDescData = protoimpl.X.CompressGZIP(file_pool_proto_rawDescData)
+		file_pool_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_pool_proto_rawDesc), len(file_pool_proto_rawDesc)))
 	})
 	return file_pool_proto_rawDescData
 }
 
 var file_pool_proto_enumTypes = make([]protoimpl.EnumInfo, 2)
 var file_pool_proto_msgTypes = make([]protoimpl.MessageInfo, 8)
-var file_pool_proto_goTypes = []interface{}{
+var file_pool_proto_goTypes = []any{
 	(PayoutMode)(0),               // 0: pool.PayoutMode
 	(AddressFormat)(0),            // 1: pool.AddressFormat
 	(*PayoutsInfo)(nil),           // 2: pool.PayoutsInfo
@@ -817,111 +757,13 @@ func file_pool_proto_init() {
 	if File_pool_proto != nil {
 		return
 	}
-	if !protoimpl.UnsafeEnabled {
-		file_pool_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*PayoutsInfo); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_pool_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*PoolFee); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_pool_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*PoolSlave); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_pool_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*PoolInfo); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_pool_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*NetworkInfo); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_pool_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*PoolSlaves); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_pool_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*GetPoolAssetRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_pool_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*PoolStats); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-	}
-	file_pool_proto_msgTypes[0].OneofWrappers = []interface{}{}
-	file_pool_proto_msgTypes[1].OneofWrappers = []interface{}{}
+	file_pool_proto_msgTypes[0].OneofWrappers = []any{}
+	file_pool_proto_msgTypes[1].OneofWrappers = []any{}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
-			RawDescriptor: file_pool_proto_rawDesc,
+			RawDescriptor: unsafe.Slice(unsafe.StringData(file_pool_proto_rawDesc), len(file_pool_proto_rawDesc)),
 			NumEnums:      2,
 			NumMessages:   8,
 			NumExtensions: 0,
@@ -933,7 +775,6 @@ func file_pool_proto_init() {
 		MessageInfos:      file_pool_proto_msgTypes,
 	}.Build()
 	File_pool_proto = out.File
-	file_pool_proto_rawDesc = nil
 	file_pool_proto_goTypes = nil
 	file_pool_proto_depIdxs = nil
 }
