@@ -1,10 +1,10 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.28.1
-// 	protoc        v5.26.1
+// 	protoc-gen-go v1.36.6
+// 	protoc        v5.29.3
 // source: pool_miners.proto
 
-package poolMinersProto
+package pool_miners
 
 import (
 	filters "github.com/grandminingpool/pool-api-proto/generated/utils/filters"
@@ -15,6 +15,7 @@ import (
 	timestamppb "google.golang.org/protobuf/types/known/timestamppb"
 	reflect "reflect"
 	sync "sync"
+	unsafe "unsafe"
 )
 
 const (
@@ -25,20 +26,17 @@ const (
 )
 
 type MinerAddressRequest struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Address       string                 `protobuf:"bytes,1,opt,name=address,proto3" json:"address,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Address string `protobuf:"bytes,1,opt,name=address,proto3" json:"address,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *MinerAddressRequest) Reset() {
 	*x = MinerAddressRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_miners_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_miners_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *MinerAddressRequest) String() string {
@@ -49,7 +47,7 @@ func (*MinerAddressRequest) ProtoMessage() {}
 
 func (x *MinerAddressRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_miners_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -72,20 +70,17 @@ func (x *MinerAddressRequest) GetAddress() string {
 }
 
 type MinerAddressesRequest struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Addresses     []string               `protobuf:"bytes,1,rep,name=addresses,proto3" json:"addresses,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Addresses []string `protobuf:"bytes,1,rep,name=addresses,proto3" json:"addresses,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *MinerAddressesRequest) Reset() {
 	*x = MinerAddressesRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_miners_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_miners_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *MinerAddressesRequest) String() string {
@@ -96,7 +91,7 @@ func (*MinerAddressesRequest) ProtoMessage() {}
 
 func (x *MinerAddressesRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_miners_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -119,20 +114,17 @@ func (x *MinerAddressesRequest) GetAddresses() []string {
 }
 
 type ValidateAddressResponse struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Valid         bool                   `protobuf:"varint,1,opt,name=valid,proto3" json:"valid,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Valid bool `protobuf:"varint,1,opt,name=valid,proto3" json:"valid,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *ValidateAddressResponse) Reset() {
 	*x = ValidateAddressResponse{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_miners_proto_msgTypes[2]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_miners_proto_msgTypes[2]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *ValidateAddressResponse) String() string {
@@ -143,7 +135,7 @@ func (*ValidateAddressResponse) ProtoMessage() {}
 
 func (x *ValidateAddressResponse) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_miners_proto_msgTypes[2]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -166,25 +158,22 @@ func (x *ValidateAddressResponse) GetValid() bool {
 }
 
 type MinersFilters struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
+	state           protoimpl.MessageState       `protogen:"open.v1"`
 	Address         *string                      `protobuf:"bytes,1,opt,name=address,proto3,oneof" json:"address,omitempty"`
 	Hashrate        *filters.BigIntRangeFilter   `protobuf:"bytes,2,opt,name=hashrate,proto3,oneof" json:"hashrate,omitempty"`
 	WorkersCount    *filters.UInt32Filter        `protobuf:"bytes,3,opt,name=workers_count,json=workersCount,proto3,oneof" json:"workers_count,omitempty"`
 	BlocksCount     *filters.UInt32Filter        `protobuf:"bytes,4,opt,name=blocks_count,json=blocksCount,proto3,oneof" json:"blocks_count,omitempty"`
 	SoloBlocksCount *filters.UInt32Filter        `protobuf:"bytes,5,opt,name=solo_blocks_count,json=soloBlocksCount,proto3,oneof" json:"solo_blocks_count,omitempty"`
 	JoinedAt        *filters.DateTimeRangeFilter `protobuf:"bytes,6,opt,name=joined_at,json=joinedAt,proto3,oneof" json:"joined_at,omitempty"`
+	unknownFields   protoimpl.UnknownFields
+	sizeCache       protoimpl.SizeCache
 }
 
 func (x *MinersFilters) Reset() {
 	*x = MinersFilters{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_miners_proto_msgTypes[3]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_miners_proto_msgTypes[3]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *MinersFilters) String() string {
@@ -195,7 +184,7 @@ func (*MinersFilters) ProtoMessage() {}
 
 func (x *MinersFilters) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_miners_proto_msgTypes[3]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -253,25 +242,22 @@ func (x *MinersFilters) GetJoinedAt() *filters.DateTimeRangeFilter {
 }
 
 type MinersSorts struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	Address         *sorts.SortOrder `protobuf:"bytes,1,opt,name=address,proto3,oneof" json:"address,omitempty"`
-	Hashrate        *sorts.SortOrder `protobuf:"bytes,2,opt,name=hashrate,proto3,oneof" json:"hashrate,omitempty"`
-	WorkersCount    *sorts.SortOrder `protobuf:"bytes,3,opt,name=workers_count,json=workersCount,proto3,oneof" json:"workers_count,omitempty"`
-	BlocksCount     *sorts.SortOrder `protobuf:"bytes,4,opt,name=blocks_count,json=blocksCount,proto3,oneof" json:"blocks_count,omitempty"`
-	SoloBlocksCount *sorts.SortOrder `protobuf:"bytes,5,opt,name=solo_blocks_count,json=soloBlocksCount,proto3,oneof" json:"solo_blocks_count,omitempty"`
-	JoinedAt        *sorts.SortOrder `protobuf:"bytes,6,opt,name=joined_at,json=joinedAt,proto3,oneof" json:"joined_at,omitempty"`
+	state           protoimpl.MessageState `protogen:"open.v1"`
+	Address         *sorts.SortOrder       `protobuf:"bytes,1,opt,name=address,proto3,oneof" json:"address,omitempty"`
+	Hashrate        *sorts.SortOrder       `protobuf:"bytes,2,opt,name=hashrate,proto3,oneof" json:"hashrate,omitempty"`
+	WorkersCount    *sorts.SortOrder       `protobuf:"bytes,3,opt,name=workers_count,json=workersCount,proto3,oneof" json:"workers_count,omitempty"`
+	BlocksCount     *sorts.SortOrder       `protobuf:"bytes,4,opt,name=blocks_count,json=blocksCount,proto3,oneof" json:"blocks_count,omitempty"`
+	SoloBlocksCount *sorts.SortOrder       `protobuf:"bytes,5,opt,name=solo_blocks_count,json=soloBlocksCount,proto3,oneof" json:"solo_blocks_count,omitempty"`
+	JoinedAt        *sorts.SortOrder       `protobuf:"bytes,6,opt,name=joined_at,json=joinedAt,proto3,oneof" json:"joined_at,omitempty"`
+	unknownFields   protoimpl.UnknownFields
+	sizeCache       protoimpl.SizeCache
 }
 
 func (x *MinersSorts) Reset() {
 	*x = MinersSorts{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_miners_proto_msgTypes[4]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_miners_proto_msgTypes[4]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *MinersSorts) String() string {
@@ -282,7 +268,7 @@ func (*MinersSorts) ProtoMessage() {}
 
 func (x *MinersSorts) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_miners_proto_msgTypes[4]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -340,22 +326,19 @@ func (x *MinersSorts) GetJoinedAt() *sorts.SortOrder {
 }
 
 type GetMinersRequest struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState        `protogen:"open.v1"`
+	Filters       *MinersFilters                `protobuf:"bytes,1,opt,name=filters,proto3" json:"filters,omitempty"`
+	Sorts         *MinersSorts                  `protobuf:"bytes,2,opt,name=sorts,proto3" json:"sorts,omitempty"`
+	Pagination    *pagination.PaginationRequest `protobuf:"bytes,3,opt,name=pagination,proto3" json:"pagination,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Filters    *MinersFilters                `protobuf:"bytes,1,opt,name=filters,proto3" json:"filters,omitempty"`
-	Sorts      *MinersSorts                  `protobuf:"bytes,2,opt,name=sorts,proto3" json:"sorts,omitempty"`
-	Pagination *pagination.PaginationRequest `protobuf:"bytes,3,opt,name=pagination,proto3" json:"pagination,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *GetMinersRequest) Reset() {
 	*x = GetMinersRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_miners_proto_msgTypes[5]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_miners_proto_msgTypes[5]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *GetMinersRequest) String() string {
@@ -366,7 +349,7 @@ func (*GetMinersRequest) ProtoMessage() {}
 
 func (x *GetMinersRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_miners_proto_msgTypes[5]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -403,10 +386,7 @@ func (x *GetMinersRequest) GetPagination() *pagination.PaginationRequest {
 }
 
 type Miner struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
+	state           protoimpl.MessageState `protogen:"open.v1"`
 	Address         string                 `protobuf:"bytes,1,opt,name=address,proto3" json:"address,omitempty"`
 	Hashrate        []byte                 `protobuf:"bytes,2,opt,name=hashrate,proto3" json:"hashrate,omitempty"`
 	WorkersCount    uint32                 `protobuf:"varint,3,opt,name=workers_count,json=workersCount,proto3" json:"workers_count,omitempty"`
@@ -414,15 +394,15 @@ type Miner struct {
 	SoloBlocksCount uint32                 `protobuf:"varint,5,opt,name=solo_blocks_count,json=soloBlocksCount,proto3" json:"solo_blocks_count,omitempty"`
 	JoinedAt        *timestamppb.Timestamp `protobuf:"bytes,6,opt,name=joined_at,json=joinedAt,proto3" json:"joined_at,omitempty"`
 	LastActivity    *timestamppb.Timestamp `protobuf:"bytes,7,opt,name=last_activity,json=lastActivity,proto3" json:"last_activity,omitempty"`
+	unknownFields   protoimpl.UnknownFields
+	sizeCache       protoimpl.SizeCache
 }
 
 func (x *Miner) Reset() {
 	*x = Miner{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_miners_proto_msgTypes[6]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_miners_proto_msgTypes[6]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Miner) String() string {
@@ -433,7 +413,7 @@ func (*Miner) ProtoMessage() {}
 
 func (x *Miner) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_miners_proto_msgTypes[6]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -498,21 +478,18 @@ func (x *Miner) GetLastActivity() *timestamppb.Timestamp {
 }
 
 type MinersList struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState         `protogen:"open.v1"`
+	Miners        []*Miner                       `protobuf:"bytes,1,rep,name=miners,proto3" json:"miners,omitempty"`
+	Pagination    *pagination.PaginationResponse `protobuf:"bytes,2,opt,name=pagination,proto3" json:"pagination,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Miners     []*Miner                       `protobuf:"bytes,1,rep,name=miners,proto3" json:"miners,omitempty"`
-	Pagination *pagination.PaginationResponse `protobuf:"bytes,2,opt,name=pagination,proto3" json:"pagination,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *MinersList) Reset() {
 	*x = MinersList{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_miners_proto_msgTypes[7]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_miners_proto_msgTypes[7]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *MinersList) String() string {
@@ -523,7 +500,7 @@ func (*MinersList) ProtoMessage() {}
 
 func (x *MinersList) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_miners_proto_msgTypes[7]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -553,25 +530,22 @@ func (x *MinersList) GetPagination() *pagination.PaginationResponse {
 }
 
 type MinerWorker struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Worker        string                 `protobuf:"bytes,1,opt,name=worker,proto3" json:"worker,omitempty"`
+	Region        string                 `protobuf:"bytes,2,opt,name=region,proto3" json:"region,omitempty"`
+	Agent         string                 `protobuf:"bytes,3,opt,name=agent,proto3" json:"agent,omitempty"`
+	Solo          bool                   `protobuf:"varint,4,opt,name=solo,proto3" json:"solo,omitempty"`
+	Hashrate      []byte                 `protobuf:"bytes,5,opt,name=hashrate,proto3" json:"hashrate,omitempty"`
+	ConnectedAt   *timestamppb.Timestamp `protobuf:"bytes,6,opt,name=connected_at,json=connectedAt,proto3" json:"connected_at,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Worker      string                 `protobuf:"bytes,1,opt,name=worker,proto3" json:"worker,omitempty"`
-	Region      string                 `protobuf:"bytes,2,opt,name=region,proto3" json:"region,omitempty"`
-	Agent       string                 `protobuf:"bytes,3,opt,name=agent,proto3" json:"agent,omitempty"`
-	Solo        bool                   `protobuf:"varint,4,opt,name=solo,proto3" json:"solo,omitempty"`
-	Hashrate    []byte                 `protobuf:"bytes,5,opt,name=hashrate,proto3" json:"hashrate,omitempty"`
-	ConnectedAt *timestamppb.Timestamp `protobuf:"bytes,6,opt,name=connected_at,json=connectedAt,proto3" json:"connected_at,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *MinerWorker) Reset() {
 	*x = MinerWorker{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_miners_proto_msgTypes[8]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_miners_proto_msgTypes[8]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *MinerWorker) String() string {
@@ -582,7 +556,7 @@ func (*MinerWorker) ProtoMessage() {}
 
 func (x *MinerWorker) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_miners_proto_msgTypes[8]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -640,20 +614,17 @@ func (x *MinerWorker) GetConnectedAt() *timestamppb.Timestamp {
 }
 
 type MinerWorkers struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Workers       []*MinerWorker         `protobuf:"bytes,1,rep,name=workers,proto3" json:"workers,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Workers []*MinerWorker `protobuf:"bytes,1,rep,name=workers,proto3" json:"workers,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *MinerWorkers) Reset() {
 	*x = MinerWorkers{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_miners_proto_msgTypes[9]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_miners_proto_msgTypes[9]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *MinerWorkers) String() string {
@@ -664,7 +635,7 @@ func (*MinerWorkers) ProtoMessage() {}
 
 func (x *MinerWorkers) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_miners_proto_msgTypes[9]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -687,20 +658,17 @@ func (x *MinerWorkers) GetWorkers() []*MinerWorker {
 }
 
 type MinersWorkersMap struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState   `protogen:"open.v1"`
+	Workers       map[string]*MinerWorkers `protobuf:"bytes,1,rep,name=workers,proto3" json:"workers,omitempty" protobuf_key:"bytes,1,opt,name=key" protobuf_val:"bytes,2,opt,name=value"`
 	unknownFields protoimpl.UnknownFields
-
-	Workers map[string]*MinerWorkers `protobuf:"bytes,1,rep,name=workers,proto3" json:"workers,omitempty" protobuf_key:"bytes,1,opt,name=key,proto3" protobuf_val:"bytes,2,opt,name=value,proto3"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *MinersWorkersMap) Reset() {
 	*x = MinersWorkersMap{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_miners_proto_msgTypes[10]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_miners_proto_msgTypes[10]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *MinersWorkersMap) String() string {
@@ -711,7 +679,7 @@ func (*MinersWorkersMap) ProtoMessage() {}
 
 func (x *MinersWorkersMap) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_miners_proto_msgTypes[10]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -734,25 +702,22 @@ func (x *MinersWorkersMap) GetWorkers() map[string]*MinerWorkers {
 }
 
 type MinerWorkersSorts struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Worker        *sorts.SortOrder       `protobuf:"bytes,1,opt,name=worker,proto3,oneof" json:"worker,omitempty"`
+	Region        *sorts.SortOrder       `protobuf:"bytes,2,opt,name=region,proto3,oneof" json:"region,omitempty"`
+	Agent         *sorts.SortOrder       `protobuf:"bytes,3,opt,name=agent,proto3,oneof" json:"agent,omitempty"`
+	Solo          *sorts.SortOrder       `protobuf:"bytes,4,opt,name=solo,proto3,oneof" json:"solo,omitempty"`
+	Hashrate      *sorts.SortOrder       `protobuf:"bytes,5,opt,name=hashrate,proto3,oneof" json:"hashrate,omitempty"`
+	ConnectedAt   *sorts.SortOrder       `protobuf:"bytes,6,opt,name=connected_at,json=connectedAt,proto3,oneof" json:"connected_at,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Worker      *sorts.SortOrder `protobuf:"bytes,1,opt,name=worker,proto3,oneof" json:"worker,omitempty"`
-	Region      *sorts.SortOrder `protobuf:"bytes,2,opt,name=region,proto3,oneof" json:"region,omitempty"`
-	Agent       *sorts.SortOrder `protobuf:"bytes,3,opt,name=agent,proto3,oneof" json:"agent,omitempty"`
-	Solo        *sorts.SortOrder `protobuf:"bytes,4,opt,name=solo,proto3,oneof" json:"solo,omitempty"`
-	Hashrate    *sorts.SortOrder `protobuf:"bytes,5,opt,name=hashrate,proto3,oneof" json:"hashrate,omitempty"`
-	ConnectedAt *sorts.SortOrder `protobuf:"bytes,6,opt,name=connected_at,json=connectedAt,proto3,oneof" json:"connected_at,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *MinerWorkersSorts) Reset() {
 	*x = MinerWorkersSorts{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_miners_proto_msgTypes[11]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_miners_proto_msgTypes[11]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *MinerWorkersSorts) String() string {
@@ -763,7 +728,7 @@ func (*MinerWorkersSorts) ProtoMessage() {}
 
 func (x *MinerWorkersSorts) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_miners_proto_msgTypes[11]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -821,25 +786,22 @@ func (x *MinerWorkersSorts) GetConnectedAt() *sorts.SortOrder {
 }
 
 type MinerWorkersFilters struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState       `protogen:"open.v1"`
+	Worker        *string                      `protobuf:"bytes,1,opt,name=worker,proto3,oneof" json:"worker,omitempty"`
+	Region        *string                      `protobuf:"bytes,2,opt,name=region,proto3,oneof" json:"region,omitempty"`
+	Agent         *string                      `protobuf:"bytes,3,opt,name=agent,proto3,oneof" json:"agent,omitempty"`
+	Solo          *bool                        `protobuf:"varint,4,opt,name=solo,proto3,oneof" json:"solo,omitempty"`
+	Hashrate      *filters.BigIntRangeFilter   `protobuf:"bytes,5,opt,name=hashrate,proto3,oneof" json:"hashrate,omitempty"`
+	ConnectedAt   *filters.DateTimeRangeFilter `protobuf:"bytes,6,opt,name=connected_at,json=connectedAt,proto3,oneof" json:"connected_at,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Worker      *string                      `protobuf:"bytes,1,opt,name=worker,proto3,oneof" json:"worker,omitempty"`
-	Region      *string                      `protobuf:"bytes,2,opt,name=region,proto3,oneof" json:"region,omitempty"`
-	Agent       *string                      `protobuf:"bytes,3,opt,name=agent,proto3,oneof" json:"agent,omitempty"`
-	Solo        *bool                        `protobuf:"varint,4,opt,name=solo,proto3,oneof" json:"solo,omitempty"`
-	Hashrate    *filters.BigIntRangeFilter   `protobuf:"bytes,5,opt,name=hashrate,proto3,oneof" json:"hashrate,omitempty"`
-	ConnectedAt *filters.DateTimeRangeFilter `protobuf:"bytes,6,opt,name=connected_at,json=connectedAt,proto3,oneof" json:"connected_at,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *MinerWorkersFilters) Reset() {
 	*x = MinerWorkersFilters{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_miners_proto_msgTypes[12]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_miners_proto_msgTypes[12]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *MinerWorkersFilters) String() string {
@@ -850,7 +812,7 @@ func (*MinerWorkersFilters) ProtoMessage() {}
 
 func (x *MinerWorkersFilters) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_miners_proto_msgTypes[12]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -908,23 +870,20 @@ func (x *MinerWorkersFilters) GetConnectedAt() *filters.DateTimeRangeFilter {
 }
 
 type GetMinerWorkersRequest struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState        `protogen:"open.v1"`
+	Address       string                        `protobuf:"bytes,1,opt,name=address,proto3" json:"address,omitempty"`
+	Sorts         *MinerWorkersSorts            `protobuf:"bytes,2,opt,name=sorts,proto3" json:"sorts,omitempty"`
+	Filters       *MinerWorkersFilters          `protobuf:"bytes,3,opt,name=filters,proto3" json:"filters,omitempty"`
+	Pagination    *pagination.PaginationRequest `protobuf:"bytes,4,opt,name=pagination,proto3" json:"pagination,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Address    string                        `protobuf:"bytes,1,opt,name=address,proto3" json:"address,omitempty"`
-	Sorts      *MinerWorkersSorts            `protobuf:"bytes,2,opt,name=sorts,proto3" json:"sorts,omitempty"`
-	Filters    *MinerWorkersFilters          `protobuf:"bytes,3,opt,name=filters,proto3" json:"filters,omitempty"`
-	Pagination *pagination.PaginationRequest `protobuf:"bytes,4,opt,name=pagination,proto3" json:"pagination,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *GetMinerWorkersRequest) Reset() {
 	*x = GetMinerWorkersRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_miners_proto_msgTypes[13]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_miners_proto_msgTypes[13]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *GetMinerWorkersRequest) String() string {
@@ -935,7 +894,7 @@ func (*GetMinerWorkersRequest) ProtoMessage() {}
 
 func (x *GetMinerWorkersRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_miners_proto_msgTypes[13]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -979,21 +938,18 @@ func (x *GetMinerWorkersRequest) GetPagination() *pagination.PaginationRequest {
 }
 
 type MinerWorkersList struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState         `protogen:"open.v1"`
+	Workers       *MinerWorkers                  `protobuf:"bytes,1,opt,name=workers,proto3" json:"workers,omitempty"`
+	Pagination    *pagination.PaginationResponse `protobuf:"bytes,2,opt,name=pagination,proto3" json:"pagination,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Workers    *MinerWorkers                  `protobuf:"bytes,1,opt,name=workers,proto3" json:"workers,omitempty"`
-	Pagination *pagination.PaginationResponse `protobuf:"bytes,2,opt,name=pagination,proto3" json:"pagination,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *MinerWorkersList) Reset() {
 	*x = MinerWorkersList{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_miners_proto_msgTypes[14]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_miners_proto_msgTypes[14]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *MinerWorkersList) String() string {
@@ -1004,7 +960,7 @@ func (*MinerWorkersList) ProtoMessage() {}
 
 func (x *MinerWorkersList) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_miners_proto_msgTypes[14]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1035,265 +991,138 @@ func (x *MinerWorkersList) GetPagination() *pagination.PaginationResponse {
 
 var File_pool_miners_proto protoreflect.FileDescriptor
 
-var file_pool_miners_proto_rawDesc = []byte{
-	0x0a, 0x11, 0x70, 0x6f, 0x6f, 0x6c, 0x5f, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x73, 0x2e, 0x70, 0x72,
-	0x6f, 0x74, 0x6f, 0x12, 0x0b, 0x70, 0x6f, 0x6f, 0x6c, 0x5f, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x73,
-	0x1a, 0x1f, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2f, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75,
-	0x66, 0x2f, 0x74, 0x69, 0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x2e, 0x70, 0x72, 0x6f, 0x74,
-	0x6f, 0x1a, 0x16, 0x75, 0x74, 0x69, 0x6c, 0x73, 0x2f, 0x70, 0x61, 0x67, 0x69, 0x6e, 0x61, 0x74,
-	0x69, 0x6f, 0x6e, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x1a, 0x11, 0x75, 0x74, 0x69, 0x6c, 0x73,
-	0x2f, 0x73, 0x6f, 0x72, 0x74, 0x73, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x1a, 0x13, 0x75, 0x74,
-	0x69, 0x6c, 0x73, 0x2f, 0x66, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x73, 0x2e, 0x70, 0x72, 0x6f, 0x74,
-	0x6f, 0x22, 0x2f, 0x0a, 0x13, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x41, 0x64, 0x64, 0x72, 0x65, 0x73,
-	0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x18, 0x0a, 0x07, 0x61, 0x64, 0x64, 0x72,
-	0x65, 0x73, 0x73, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x07, 0x61, 0x64, 0x64, 0x72, 0x65,
-	0x73, 0x73, 0x22, 0x35, 0x0a, 0x15, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x41, 0x64, 0x64, 0x72, 0x65,
-	0x73, 0x73, 0x65, 0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x1c, 0x0a, 0x09, 0x61,
-	0x64, 0x64, 0x72, 0x65, 0x73, 0x73, 0x65, 0x73, 0x18, 0x01, 0x20, 0x03, 0x28, 0x09, 0x52, 0x09,
-	0x61, 0x64, 0x64, 0x72, 0x65, 0x73, 0x73, 0x65, 0x73, 0x22, 0x2f, 0x0a, 0x17, 0x56, 0x61, 0x6c,
-	0x69, 0x64, 0x61, 0x74, 0x65, 0x41, 0x64, 0x64, 0x72, 0x65, 0x73, 0x73, 0x52, 0x65, 0x73, 0x70,
-	0x6f, 0x6e, 0x73, 0x65, 0x12, 0x14, 0x0a, 0x05, 0x76, 0x61, 0x6c, 0x69, 0x64, 0x18, 0x01, 0x20,
-	0x01, 0x28, 0x08, 0x52, 0x05, 0x76, 0x61, 0x6c, 0x69, 0x64, 0x22, 0xd3, 0x03, 0x0a, 0x0d, 0x4d,
-	0x69, 0x6e, 0x65, 0x72, 0x73, 0x46, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x73, 0x12, 0x1d, 0x0a, 0x07,
-	0x61, 0x64, 0x64, 0x72, 0x65, 0x73, 0x73, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x48, 0x00, 0x52,
-	0x07, 0x61, 0x64, 0x64, 0x72, 0x65, 0x73, 0x73, 0x88, 0x01, 0x01, 0x12, 0x3b, 0x0a, 0x08, 0x68,
-	0x61, 0x73, 0x68, 0x72, 0x61, 0x74, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1a, 0x2e,
-	0x66, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x73, 0x2e, 0x42, 0x69, 0x67, 0x49, 0x6e, 0x74, 0x52, 0x61,
-	0x6e, 0x67, 0x65, 0x46, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x48, 0x01, 0x52, 0x08, 0x68, 0x61, 0x73,
-	0x68, 0x72, 0x61, 0x74, 0x65, 0x88, 0x01, 0x01, 0x12, 0x3f, 0x0a, 0x0d, 0x77, 0x6f, 0x72, 0x6b,
-	0x65, 0x72, 0x73, 0x5f, 0x63, 0x6f, 0x75, 0x6e, 0x74, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0b, 0x32,
-	0x15, 0x2e, 0x66, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x73, 0x2e, 0x55, 0x49, 0x6e, 0x74, 0x33, 0x32,
-	0x46, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x48, 0x02, 0x52, 0x0c, 0x77, 0x6f, 0x72, 0x6b, 0x65, 0x72,
-	0x73, 0x43, 0x6f, 0x75, 0x6e, 0x74, 0x88, 0x01, 0x01, 0x12, 0x3d, 0x0a, 0x0c, 0x62, 0x6c, 0x6f,
-	0x63, 0x6b, 0x73, 0x5f, 0x63, 0x6f, 0x75, 0x6e, 0x74, 0x18, 0x04, 0x20, 0x01, 0x28, 0x0b, 0x32,
-	0x15, 0x2e, 0x66, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x73, 0x2e, 0x55, 0x49, 0x6e, 0x74, 0x33, 0x32,
-	0x46, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x48, 0x03, 0x52, 0x0b, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x73,
-	0x43, 0x6f, 0x75, 0x6e, 0x74, 0x88, 0x01, 0x01, 0x12, 0x46, 0x0a, 0x11, 0x73, 0x6f, 0x6c, 0x6f,
-	0x5f, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x5f, 0x63, 0x6f, 0x75, 0x6e, 0x74, 0x18, 0x05, 0x20,
-	0x01, 0x28, 0x0b, 0x32, 0x15, 0x2e, 0x66, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x73, 0x2e, 0x55, 0x49,
-	0x6e, 0x74, 0x33, 0x32, 0x46, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x48, 0x04, 0x52, 0x0f, 0x73, 0x6f,
-	0x6c, 0x6f, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x43, 0x6f, 0x75, 0x6e, 0x74, 0x88, 0x01, 0x01,
-	0x12, 0x3e, 0x0a, 0x09, 0x6a, 0x6f, 0x69, 0x6e, 0x65, 0x64, 0x5f, 0x61, 0x74, 0x18, 0x06, 0x20,
-	0x01, 0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x66, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x73, 0x2e, 0x44, 0x61,
-	0x74, 0x65, 0x54, 0x69, 0x6d, 0x65, 0x52, 0x61, 0x6e, 0x67, 0x65, 0x46, 0x69, 0x6c, 0x74, 0x65,
-	0x72, 0x48, 0x05, 0x52, 0x08, 0x6a, 0x6f, 0x69, 0x6e, 0x65, 0x64, 0x41, 0x74, 0x88, 0x01, 0x01,
-	0x42, 0x0a, 0x0a, 0x08, 0x5f, 0x61, 0x64, 0x64, 0x72, 0x65, 0x73, 0x73, 0x42, 0x0b, 0x0a, 0x09,
-	0x5f, 0x68, 0x61, 0x73, 0x68, 0x72, 0x61, 0x74, 0x65, 0x42, 0x10, 0x0a, 0x0e, 0x5f, 0x77, 0x6f,
-	0x72, 0x6b, 0x65, 0x72, 0x73, 0x5f, 0x63, 0x6f, 0x75, 0x6e, 0x74, 0x42, 0x0f, 0x0a, 0x0d, 0x5f,
-	0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x5f, 0x63, 0x6f, 0x75, 0x6e, 0x74, 0x42, 0x14, 0x0a, 0x12,
-	0x5f, 0x73, 0x6f, 0x6c, 0x6f, 0x5f, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x5f, 0x63, 0x6f, 0x75,
-	0x6e, 0x74, 0x42, 0x0c, 0x0a, 0x0a, 0x5f, 0x6a, 0x6f, 0x69, 0x6e, 0x65, 0x64, 0x5f, 0x61, 0x74,
-	0x22, 0xbe, 0x03, 0x0a, 0x0b, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x73, 0x53, 0x6f, 0x72, 0x74, 0x73,
-	0x12, 0x2f, 0x0a, 0x07, 0x61, 0x64, 0x64, 0x72, 0x65, 0x73, 0x73, 0x18, 0x01, 0x20, 0x01, 0x28,
-	0x0b, 0x32, 0x10, 0x2e, 0x73, 0x6f, 0x72, 0x74, 0x73, 0x2e, 0x53, 0x6f, 0x72, 0x74, 0x4f, 0x72,
-	0x64, 0x65, 0x72, 0x48, 0x00, 0x52, 0x07, 0x61, 0x64, 0x64, 0x72, 0x65, 0x73, 0x73, 0x88, 0x01,
-	0x01, 0x12, 0x31, 0x0a, 0x08, 0x68, 0x61, 0x73, 0x68, 0x72, 0x61, 0x74, 0x65, 0x18, 0x02, 0x20,
-	0x01, 0x28, 0x0b, 0x32, 0x10, 0x2e, 0x73, 0x6f, 0x72, 0x74, 0x73, 0x2e, 0x53, 0x6f, 0x72, 0x74,
-	0x4f, 0x72, 0x64, 0x65, 0x72, 0x48, 0x01, 0x52, 0x08, 0x68, 0x61, 0x73, 0x68, 0x72, 0x61, 0x74,
-	0x65, 0x88, 0x01, 0x01, 0x12, 0x3a, 0x0a, 0x0d, 0x77, 0x6f, 0x72, 0x6b, 0x65, 0x72, 0x73, 0x5f,
-	0x63, 0x6f, 0x75, 0x6e, 0x74, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x10, 0x2e, 0x73, 0x6f,
-	0x72, 0x74, 0x73, 0x2e, 0x53, 0x6f, 0x72, 0x74, 0x4f, 0x72, 0x64, 0x65, 0x72, 0x48, 0x02, 0x52,
-	0x0c, 0x77, 0x6f, 0x72, 0x6b, 0x65, 0x72, 0x73, 0x43, 0x6f, 0x75, 0x6e, 0x74, 0x88, 0x01, 0x01,
-	0x12, 0x38, 0x0a, 0x0c, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x5f, 0x63, 0x6f, 0x75, 0x6e, 0x74,
-	0x18, 0x04, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x10, 0x2e, 0x73, 0x6f, 0x72, 0x74, 0x73, 0x2e, 0x53,
-	0x6f, 0x72, 0x74, 0x4f, 0x72, 0x64, 0x65, 0x72, 0x48, 0x03, 0x52, 0x0b, 0x62, 0x6c, 0x6f, 0x63,
-	0x6b, 0x73, 0x43, 0x6f, 0x75, 0x6e, 0x74, 0x88, 0x01, 0x01, 0x12, 0x41, 0x0a, 0x11, 0x73, 0x6f,
-	0x6c, 0x6f, 0x5f, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x5f, 0x63, 0x6f, 0x75, 0x6e, 0x74, 0x18,
-	0x05, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x10, 0x2e, 0x73, 0x6f, 0x72, 0x74, 0x73, 0x2e, 0x53, 0x6f,
-	0x72, 0x74, 0x4f, 0x72, 0x64, 0x65, 0x72, 0x48, 0x04, 0x52, 0x0f, 0x73, 0x6f, 0x6c, 0x6f, 0x42,
-	0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x43, 0x6f, 0x75, 0x6e, 0x74, 0x88, 0x01, 0x01, 0x12, 0x32, 0x0a,
-	0x09, 0x6a, 0x6f, 0x69, 0x6e, 0x65, 0x64, 0x5f, 0x61, 0x74, 0x18, 0x06, 0x20, 0x01, 0x28, 0x0b,
-	0x32, 0x10, 0x2e, 0x73, 0x6f, 0x72, 0x74, 0x73, 0x2e, 0x53, 0x6f, 0x72, 0x74, 0x4f, 0x72, 0x64,
-	0x65, 0x72, 0x48, 0x05, 0x52, 0x08, 0x6a, 0x6f, 0x69, 0x6e, 0x65, 0x64, 0x41, 0x74, 0x88, 0x01,
-	0x01, 0x42, 0x0a, 0x0a, 0x08, 0x5f, 0x61, 0x64, 0x64, 0x72, 0x65, 0x73, 0x73, 0x42, 0x0b, 0x0a,
-	0x09, 0x5f, 0x68, 0x61, 0x73, 0x68, 0x72, 0x61, 0x74, 0x65, 0x42, 0x10, 0x0a, 0x0e, 0x5f, 0x77,
-	0x6f, 0x72, 0x6b, 0x65, 0x72, 0x73, 0x5f, 0x63, 0x6f, 0x75, 0x6e, 0x74, 0x42, 0x0f, 0x0a, 0x0d,
-	0x5f, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x5f, 0x63, 0x6f, 0x75, 0x6e, 0x74, 0x42, 0x14, 0x0a,
-	0x12, 0x5f, 0x73, 0x6f, 0x6c, 0x6f, 0x5f, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x5f, 0x63, 0x6f,
-	0x75, 0x6e, 0x74, 0x42, 0x0c, 0x0a, 0x0a, 0x5f, 0x6a, 0x6f, 0x69, 0x6e, 0x65, 0x64, 0x5f, 0x61,
-	0x74, 0x22, 0xb7, 0x01, 0x0a, 0x10, 0x47, 0x65, 0x74, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x73, 0x52,
-	0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x34, 0x0a, 0x07, 0x66, 0x69, 0x6c, 0x74, 0x65, 0x72,
-	0x73, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1a, 0x2e, 0x70, 0x6f, 0x6f, 0x6c, 0x5f, 0x6d,
-	0x69, 0x6e, 0x65, 0x72, 0x73, 0x2e, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x73, 0x46, 0x69, 0x6c, 0x74,
-	0x65, 0x72, 0x73, 0x52, 0x07, 0x66, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x73, 0x12, 0x2e, 0x0a, 0x05,
-	0x73, 0x6f, 0x72, 0x74, 0x73, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x18, 0x2e, 0x70, 0x6f,
-	0x6f, 0x6c, 0x5f, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x73, 0x2e, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x73,
-	0x53, 0x6f, 0x72, 0x74, 0x73, 0x52, 0x05, 0x73, 0x6f, 0x72, 0x74, 0x73, 0x12, 0x3d, 0x0a, 0x0a,
-	0x70, 0x61, 0x67, 0x69, 0x6e, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0b,
-	0x32, 0x1d, 0x2e, 0x70, 0x61, 0x67, 0x69, 0x6e, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x2e, 0x50, 0x61,
-	0x67, 0x69, 0x6e, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x52,
-	0x0a, 0x70, 0x61, 0x67, 0x69, 0x6e, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x22, 0xab, 0x02, 0x0a, 0x05,
-	0x4d, 0x69, 0x6e, 0x65, 0x72, 0x12, 0x18, 0x0a, 0x07, 0x61, 0x64, 0x64, 0x72, 0x65, 0x73, 0x73,
-	0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x07, 0x61, 0x64, 0x64, 0x72, 0x65, 0x73, 0x73, 0x12,
-	0x1a, 0x0a, 0x08, 0x68, 0x61, 0x73, 0x68, 0x72, 0x61, 0x74, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28,
-	0x0c, 0x52, 0x08, 0x68, 0x61, 0x73, 0x68, 0x72, 0x61, 0x74, 0x65, 0x12, 0x23, 0x0a, 0x0d, 0x77,
-	0x6f, 0x72, 0x6b, 0x65, 0x72, 0x73, 0x5f, 0x63, 0x6f, 0x75, 0x6e, 0x74, 0x18, 0x03, 0x20, 0x01,
-	0x28, 0x0d, 0x52, 0x0c, 0x77, 0x6f, 0x72, 0x6b, 0x65, 0x72, 0x73, 0x43, 0x6f, 0x75, 0x6e, 0x74,
-	0x12, 0x21, 0x0a, 0x0c, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x5f, 0x63, 0x6f, 0x75, 0x6e, 0x74,
-	0x18, 0x04, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x0b, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x43, 0x6f,
-	0x75, 0x6e, 0x74, 0x12, 0x2a, 0x0a, 0x11, 0x73, 0x6f, 0x6c, 0x6f, 0x5f, 0x62, 0x6c, 0x6f, 0x63,
-	0x6b, 0x73, 0x5f, 0x63, 0x6f, 0x75, 0x6e, 0x74, 0x18, 0x05, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x0f,
-	0x73, 0x6f, 0x6c, 0x6f, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x43, 0x6f, 0x75, 0x6e, 0x74, 0x12,
-	0x37, 0x0a, 0x09, 0x6a, 0x6f, 0x69, 0x6e, 0x65, 0x64, 0x5f, 0x61, 0x74, 0x18, 0x06, 0x20, 0x01,
-	0x28, 0x0b, 0x32, 0x1a, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74,
-	0x6f, 0x62, 0x75, 0x66, 0x2e, 0x54, 0x69, 0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x52, 0x08,
-	0x6a, 0x6f, 0x69, 0x6e, 0x65, 0x64, 0x41, 0x74, 0x12, 0x3f, 0x0a, 0x0d, 0x6c, 0x61, 0x73, 0x74,
-	0x5f, 0x61, 0x63, 0x74, 0x69, 0x76, 0x69, 0x74, 0x79, 0x18, 0x07, 0x20, 0x01, 0x28, 0x0b, 0x32,
-	0x1a, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75,
-	0x66, 0x2e, 0x54, 0x69, 0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x52, 0x0c, 0x6c, 0x61, 0x73,
-	0x74, 0x41, 0x63, 0x74, 0x69, 0x76, 0x69, 0x74, 0x79, 0x22, 0x78, 0x0a, 0x0a, 0x4d, 0x69, 0x6e,
-	0x65, 0x72, 0x73, 0x4c, 0x69, 0x73, 0x74, 0x12, 0x2a, 0x0a, 0x06, 0x6d, 0x69, 0x6e, 0x65, 0x72,
-	0x73, 0x18, 0x01, 0x20, 0x03, 0x28, 0x0b, 0x32, 0x12, 0x2e, 0x70, 0x6f, 0x6f, 0x6c, 0x5f, 0x6d,
-	0x69, 0x6e, 0x65, 0x72, 0x73, 0x2e, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x52, 0x06, 0x6d, 0x69, 0x6e,
-	0x65, 0x72, 0x73, 0x12, 0x3e, 0x0a, 0x0a, 0x70, 0x61, 0x67, 0x69, 0x6e, 0x61, 0x74, 0x69, 0x6f,
-	0x6e, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1e, 0x2e, 0x70, 0x61, 0x67, 0x69, 0x6e, 0x61,
-	0x74, 0x69, 0x6f, 0x6e, 0x2e, 0x50, 0x61, 0x67, 0x69, 0x6e, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x52,
-	0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x52, 0x0a, 0x70, 0x61, 0x67, 0x69, 0x6e, 0x61, 0x74,
-	0x69, 0x6f, 0x6e, 0x22, 0xc2, 0x01, 0x0a, 0x0b, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x57, 0x6f, 0x72,
-	0x6b, 0x65, 0x72, 0x12, 0x16, 0x0a, 0x06, 0x77, 0x6f, 0x72, 0x6b, 0x65, 0x72, 0x18, 0x01, 0x20,
-	0x01, 0x28, 0x09, 0x52, 0x06, 0x77, 0x6f, 0x72, 0x6b, 0x65, 0x72, 0x12, 0x16, 0x0a, 0x06, 0x72,
-	0x65, 0x67, 0x69, 0x6f, 0x6e, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x06, 0x72, 0x65, 0x67,
-	0x69, 0x6f, 0x6e, 0x12, 0x14, 0x0a, 0x05, 0x61, 0x67, 0x65, 0x6e, 0x74, 0x18, 0x03, 0x20, 0x01,
-	0x28, 0x09, 0x52, 0x05, 0x61, 0x67, 0x65, 0x6e, 0x74, 0x12, 0x12, 0x0a, 0x04, 0x73, 0x6f, 0x6c,
-	0x6f, 0x18, 0x04, 0x20, 0x01, 0x28, 0x08, 0x52, 0x04, 0x73, 0x6f, 0x6c, 0x6f, 0x12, 0x1a, 0x0a,
-	0x08, 0x68, 0x61, 0x73, 0x68, 0x72, 0x61, 0x74, 0x65, 0x18, 0x05, 0x20, 0x01, 0x28, 0x0c, 0x52,
-	0x08, 0x68, 0x61, 0x73, 0x68, 0x72, 0x61, 0x74, 0x65, 0x12, 0x3d, 0x0a, 0x0c, 0x63, 0x6f, 0x6e,
-	0x6e, 0x65, 0x63, 0x74, 0x65, 0x64, 0x5f, 0x61, 0x74, 0x18, 0x06, 0x20, 0x01, 0x28, 0x0b, 0x32,
-	0x1a, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75,
-	0x66, 0x2e, 0x54, 0x69, 0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x52, 0x0b, 0x63, 0x6f, 0x6e,
-	0x6e, 0x65, 0x63, 0x74, 0x65, 0x64, 0x41, 0x74, 0x22, 0x42, 0x0a, 0x0c, 0x4d, 0x69, 0x6e, 0x65,
-	0x72, 0x57, 0x6f, 0x72, 0x6b, 0x65, 0x72, 0x73, 0x12, 0x32, 0x0a, 0x07, 0x77, 0x6f, 0x72, 0x6b,
-	0x65, 0x72, 0x73, 0x18, 0x01, 0x20, 0x03, 0x28, 0x0b, 0x32, 0x18, 0x2e, 0x70, 0x6f, 0x6f, 0x6c,
-	0x5f, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x73, 0x2e, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x57, 0x6f, 0x72,
-	0x6b, 0x65, 0x72, 0x52, 0x07, 0x77, 0x6f, 0x72, 0x6b, 0x65, 0x72, 0x73, 0x22, 0xaf, 0x01, 0x0a,
-	0x10, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x73, 0x57, 0x6f, 0x72, 0x6b, 0x65, 0x72, 0x73, 0x4d, 0x61,
-	0x70, 0x12, 0x44, 0x0a, 0x07, 0x77, 0x6f, 0x72, 0x6b, 0x65, 0x72, 0x73, 0x18, 0x01, 0x20, 0x03,
-	0x28, 0x0b, 0x32, 0x2a, 0x2e, 0x70, 0x6f, 0x6f, 0x6c, 0x5f, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x73,
-	0x2e, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x73, 0x57, 0x6f, 0x72, 0x6b, 0x65, 0x72, 0x73, 0x4d, 0x61,
-	0x70, 0x2e, 0x57, 0x6f, 0x72, 0x6b, 0x65, 0x72, 0x73, 0x45, 0x6e, 0x74, 0x72, 0x79, 0x52, 0x07,
-	0x77, 0x6f, 0x72, 0x6b, 0x65, 0x72, 0x73, 0x1a, 0x55, 0x0a, 0x0c, 0x57, 0x6f, 0x72, 0x6b, 0x65,
-	0x72, 0x73, 0x45, 0x6e, 0x74, 0x72, 0x79, 0x12, 0x10, 0x0a, 0x03, 0x6b, 0x65, 0x79, 0x18, 0x01,
-	0x20, 0x01, 0x28, 0x09, 0x52, 0x03, 0x6b, 0x65, 0x79, 0x12, 0x2f, 0x0a, 0x05, 0x76, 0x61, 0x6c,
-	0x75, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x19, 0x2e, 0x70, 0x6f, 0x6f, 0x6c, 0x5f,
-	0x6d, 0x69, 0x6e, 0x65, 0x72, 0x73, 0x2e, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x57, 0x6f, 0x72, 0x6b,
-	0x65, 0x72, 0x73, 0x52, 0x05, 0x76, 0x61, 0x6c, 0x75, 0x65, 0x3a, 0x02, 0x38, 0x01, 0x22, 0xfd,
-	0x02, 0x0a, 0x11, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x57, 0x6f, 0x72, 0x6b, 0x65, 0x72, 0x73, 0x53,
-	0x6f, 0x72, 0x74, 0x73, 0x12, 0x2d, 0x0a, 0x06, 0x77, 0x6f, 0x72, 0x6b, 0x65, 0x72, 0x18, 0x01,
-	0x20, 0x01, 0x28, 0x0b, 0x32, 0x10, 0x2e, 0x73, 0x6f, 0x72, 0x74, 0x73, 0x2e, 0x53, 0x6f, 0x72,
-	0x74, 0x4f, 0x72, 0x64, 0x65, 0x72, 0x48, 0x00, 0x52, 0x06, 0x77, 0x6f, 0x72, 0x6b, 0x65, 0x72,
-	0x88, 0x01, 0x01, 0x12, 0x2d, 0x0a, 0x06, 0x72, 0x65, 0x67, 0x69, 0x6f, 0x6e, 0x18, 0x02, 0x20,
-	0x01, 0x28, 0x0b, 0x32, 0x10, 0x2e, 0x73, 0x6f, 0x72, 0x74, 0x73, 0x2e, 0x53, 0x6f, 0x72, 0x74,
-	0x4f, 0x72, 0x64, 0x65, 0x72, 0x48, 0x01, 0x52, 0x06, 0x72, 0x65, 0x67, 0x69, 0x6f, 0x6e, 0x88,
-	0x01, 0x01, 0x12, 0x2b, 0x0a, 0x05, 0x61, 0x67, 0x65, 0x6e, 0x74, 0x18, 0x03, 0x20, 0x01, 0x28,
-	0x0b, 0x32, 0x10, 0x2e, 0x73, 0x6f, 0x72, 0x74, 0x73, 0x2e, 0x53, 0x6f, 0x72, 0x74, 0x4f, 0x72,
-	0x64, 0x65, 0x72, 0x48, 0x02, 0x52, 0x05, 0x61, 0x67, 0x65, 0x6e, 0x74, 0x88, 0x01, 0x01, 0x12,
-	0x29, 0x0a, 0x04, 0x73, 0x6f, 0x6c, 0x6f, 0x18, 0x04, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x10, 0x2e,
-	0x73, 0x6f, 0x72, 0x74, 0x73, 0x2e, 0x53, 0x6f, 0x72, 0x74, 0x4f, 0x72, 0x64, 0x65, 0x72, 0x48,
-	0x03, 0x52, 0x04, 0x73, 0x6f, 0x6c, 0x6f, 0x88, 0x01, 0x01, 0x12, 0x31, 0x0a, 0x08, 0x68, 0x61,
-	0x73, 0x68, 0x72, 0x61, 0x74, 0x65, 0x18, 0x05, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x10, 0x2e, 0x73,
-	0x6f, 0x72, 0x74, 0x73, 0x2e, 0x53, 0x6f, 0x72, 0x74, 0x4f, 0x72, 0x64, 0x65, 0x72, 0x48, 0x04,
-	0x52, 0x08, 0x68, 0x61, 0x73, 0x68, 0x72, 0x61, 0x74, 0x65, 0x88, 0x01, 0x01, 0x12, 0x38, 0x0a,
-	0x0c, 0x63, 0x6f, 0x6e, 0x6e, 0x65, 0x63, 0x74, 0x65, 0x64, 0x5f, 0x61, 0x74, 0x18, 0x06, 0x20,
-	0x01, 0x28, 0x0b, 0x32, 0x10, 0x2e, 0x73, 0x6f, 0x72, 0x74, 0x73, 0x2e, 0x53, 0x6f, 0x72, 0x74,
-	0x4f, 0x72, 0x64, 0x65, 0x72, 0x48, 0x05, 0x52, 0x0b, 0x63, 0x6f, 0x6e, 0x6e, 0x65, 0x63, 0x74,
-	0x65, 0x64, 0x41, 0x74, 0x88, 0x01, 0x01, 0x42, 0x09, 0x0a, 0x07, 0x5f, 0x77, 0x6f, 0x72, 0x6b,
-	0x65, 0x72, 0x42, 0x09, 0x0a, 0x07, 0x5f, 0x72, 0x65, 0x67, 0x69, 0x6f, 0x6e, 0x42, 0x08, 0x0a,
-	0x06, 0x5f, 0x61, 0x67, 0x65, 0x6e, 0x74, 0x42, 0x07, 0x0a, 0x05, 0x5f, 0x73, 0x6f, 0x6c, 0x6f,
-	0x42, 0x0b, 0x0a, 0x09, 0x5f, 0x68, 0x61, 0x73, 0x68, 0x72, 0x61, 0x74, 0x65, 0x42, 0x0f, 0x0a,
-	0x0d, 0x5f, 0x63, 0x6f, 0x6e, 0x6e, 0x65, 0x63, 0x74, 0x65, 0x64, 0x5f, 0x61, 0x74, 0x22, 0xcd,
-	0x02, 0x0a, 0x13, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x57, 0x6f, 0x72, 0x6b, 0x65, 0x72, 0x73, 0x46,
-	0x69, 0x6c, 0x74, 0x65, 0x72, 0x73, 0x12, 0x1b, 0x0a, 0x06, 0x77, 0x6f, 0x72, 0x6b, 0x65, 0x72,
-	0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x48, 0x00, 0x52, 0x06, 0x77, 0x6f, 0x72, 0x6b, 0x65, 0x72,
-	0x88, 0x01, 0x01, 0x12, 0x1b, 0x0a, 0x06, 0x72, 0x65, 0x67, 0x69, 0x6f, 0x6e, 0x18, 0x02, 0x20,
-	0x01, 0x28, 0x09, 0x48, 0x01, 0x52, 0x06, 0x72, 0x65, 0x67, 0x69, 0x6f, 0x6e, 0x88, 0x01, 0x01,
-	0x12, 0x19, 0x0a, 0x05, 0x61, 0x67, 0x65, 0x6e, 0x74, 0x18, 0x03, 0x20, 0x01, 0x28, 0x09, 0x48,
-	0x02, 0x52, 0x05, 0x61, 0x67, 0x65, 0x6e, 0x74, 0x88, 0x01, 0x01, 0x12, 0x17, 0x0a, 0x04, 0x73,
-	0x6f, 0x6c, 0x6f, 0x18, 0x04, 0x20, 0x01, 0x28, 0x08, 0x48, 0x03, 0x52, 0x04, 0x73, 0x6f, 0x6c,
-	0x6f, 0x88, 0x01, 0x01, 0x12, 0x3b, 0x0a, 0x08, 0x68, 0x61, 0x73, 0x68, 0x72, 0x61, 0x74, 0x65,
-	0x18, 0x05, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1a, 0x2e, 0x66, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x73,
-	0x2e, 0x42, 0x69, 0x67, 0x49, 0x6e, 0x74, 0x52, 0x61, 0x6e, 0x67, 0x65, 0x46, 0x69, 0x6c, 0x74,
-	0x65, 0x72, 0x48, 0x04, 0x52, 0x08, 0x68, 0x61, 0x73, 0x68, 0x72, 0x61, 0x74, 0x65, 0x88, 0x01,
-	0x01, 0x12, 0x44, 0x0a, 0x0c, 0x63, 0x6f, 0x6e, 0x6e, 0x65, 0x63, 0x74, 0x65, 0x64, 0x5f, 0x61,
-	0x74, 0x18, 0x06, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x66, 0x69, 0x6c, 0x74, 0x65, 0x72,
-	0x73, 0x2e, 0x44, 0x61, 0x74, 0x65, 0x54, 0x69, 0x6d, 0x65, 0x52, 0x61, 0x6e, 0x67, 0x65, 0x46,
-	0x69, 0x6c, 0x74, 0x65, 0x72, 0x48, 0x05, 0x52, 0x0b, 0x63, 0x6f, 0x6e, 0x6e, 0x65, 0x63, 0x74,
-	0x65, 0x64, 0x41, 0x74, 0x88, 0x01, 0x01, 0x42, 0x09, 0x0a, 0x07, 0x5f, 0x77, 0x6f, 0x72, 0x6b,
-	0x65, 0x72, 0x42, 0x09, 0x0a, 0x07, 0x5f, 0x72, 0x65, 0x67, 0x69, 0x6f, 0x6e, 0x42, 0x08, 0x0a,
-	0x06, 0x5f, 0x61, 0x67, 0x65, 0x6e, 0x74, 0x42, 0x07, 0x0a, 0x05, 0x5f, 0x73, 0x6f, 0x6c, 0x6f,
-	0x42, 0x0b, 0x0a, 0x09, 0x5f, 0x68, 0x61, 0x73, 0x68, 0x72, 0x61, 0x74, 0x65, 0x42, 0x0f, 0x0a,
-	0x0d, 0x5f, 0x63, 0x6f, 0x6e, 0x6e, 0x65, 0x63, 0x74, 0x65, 0x64, 0x5f, 0x61, 0x74, 0x22, 0xe3,
-	0x01, 0x0a, 0x16, 0x47, 0x65, 0x74, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x57, 0x6f, 0x72, 0x6b, 0x65,
-	0x72, 0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x18, 0x0a, 0x07, 0x61, 0x64, 0x64,
-	0x72, 0x65, 0x73, 0x73, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x07, 0x61, 0x64, 0x64, 0x72,
-	0x65, 0x73, 0x73, 0x12, 0x34, 0x0a, 0x05, 0x73, 0x6f, 0x72, 0x74, 0x73, 0x18, 0x02, 0x20, 0x01,
-	0x28, 0x0b, 0x32, 0x1e, 0x2e, 0x70, 0x6f, 0x6f, 0x6c, 0x5f, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x73,
-	0x2e, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x57, 0x6f, 0x72, 0x6b, 0x65, 0x72, 0x73, 0x53, 0x6f, 0x72,
-	0x74, 0x73, 0x52, 0x05, 0x73, 0x6f, 0x72, 0x74, 0x73, 0x12, 0x3a, 0x0a, 0x07, 0x66, 0x69, 0x6c,
-	0x74, 0x65, 0x72, 0x73, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x20, 0x2e, 0x70, 0x6f, 0x6f,
-	0x6c, 0x5f, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x73, 0x2e, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x57, 0x6f,
-	0x72, 0x6b, 0x65, 0x72, 0x73, 0x46, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x73, 0x52, 0x07, 0x66, 0x69,
-	0x6c, 0x74, 0x65, 0x72, 0x73, 0x12, 0x3d, 0x0a, 0x0a, 0x70, 0x61, 0x67, 0x69, 0x6e, 0x61, 0x74,
-	0x69, 0x6f, 0x6e, 0x18, 0x04, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1d, 0x2e, 0x70, 0x61, 0x67, 0x69,
-	0x6e, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x2e, 0x50, 0x61, 0x67, 0x69, 0x6e, 0x61, 0x74, 0x69, 0x6f,
-	0x6e, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x52, 0x0a, 0x70, 0x61, 0x67, 0x69, 0x6e, 0x61,
-	0x74, 0x69, 0x6f, 0x6e, 0x22, 0x87, 0x01, 0x0a, 0x10, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x57, 0x6f,
-	0x72, 0x6b, 0x65, 0x72, 0x73, 0x4c, 0x69, 0x73, 0x74, 0x12, 0x33, 0x0a, 0x07, 0x77, 0x6f, 0x72,
-	0x6b, 0x65, 0x72, 0x73, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x19, 0x2e, 0x70, 0x6f, 0x6f,
-	0x6c, 0x5f, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x73, 0x2e, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x57, 0x6f,
-	0x72, 0x6b, 0x65, 0x72, 0x73, 0x52, 0x07, 0x77, 0x6f, 0x72, 0x6b, 0x65, 0x72, 0x73, 0x12, 0x3e,
-	0x0a, 0x0a, 0x70, 0x61, 0x67, 0x69, 0x6e, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x18, 0x02, 0x20, 0x01,
-	0x28, 0x0b, 0x32, 0x1e, 0x2e, 0x70, 0x61, 0x67, 0x69, 0x6e, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x2e,
-	0x50, 0x61, 0x67, 0x69, 0x6e, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e,
-	0x73, 0x65, 0x52, 0x0a, 0x70, 0x61, 0x67, 0x69, 0x6e, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x32, 0xab,
-	0x03, 0x0a, 0x11, 0x50, 0x6f, 0x6f, 0x6c, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x73, 0x53, 0x65, 0x72,
-	0x76, 0x69, 0x63, 0x65, 0x12, 0x59, 0x0a, 0x0f, 0x56, 0x61, 0x6c, 0x69, 0x64, 0x61, 0x74, 0x65,
-	0x41, 0x64, 0x64, 0x72, 0x65, 0x73, 0x73, 0x12, 0x20, 0x2e, 0x70, 0x6f, 0x6f, 0x6c, 0x5f, 0x6d,
-	0x69, 0x6e, 0x65, 0x72, 0x73, 0x2e, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x41, 0x64, 0x64, 0x72, 0x65,
-	0x73, 0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x24, 0x2e, 0x70, 0x6f, 0x6f, 0x6c,
-	0x5f, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x73, 0x2e, 0x56, 0x61, 0x6c, 0x69, 0x64, 0x61, 0x74, 0x65,
-	0x41, 0x64, 0x64, 0x72, 0x65, 0x73, 0x73, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12,
-	0x43, 0x0a, 0x09, 0x47, 0x65, 0x74, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x73, 0x12, 0x1d, 0x2e, 0x70,
-	0x6f, 0x6f, 0x6c, 0x5f, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x73, 0x2e, 0x47, 0x65, 0x74, 0x4d, 0x69,
-	0x6e, 0x65, 0x72, 0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x17, 0x2e, 0x70, 0x6f,
-	0x6f, 0x6c, 0x5f, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x73, 0x2e, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x73,
-	0x4c, 0x69, 0x73, 0x74, 0x12, 0x40, 0x0a, 0x08, 0x47, 0x65, 0x74, 0x4d, 0x69, 0x6e, 0x65, 0x72,
-	0x12, 0x20, 0x2e, 0x70, 0x6f, 0x6f, 0x6c, 0x5f, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x73, 0x2e, 0x4d,
-	0x69, 0x6e, 0x65, 0x72, 0x41, 0x64, 0x64, 0x72, 0x65, 0x73, 0x73, 0x52, 0x65, 0x71, 0x75, 0x65,
-	0x73, 0x74, 0x1a, 0x12, 0x2e, 0x70, 0x6f, 0x6f, 0x6c, 0x5f, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x73,
-	0x2e, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x12, 0x55, 0x0a, 0x0f, 0x47, 0x65, 0x74, 0x4d, 0x69, 0x6e,
-	0x65, 0x72, 0x57, 0x6f, 0x72, 0x6b, 0x65, 0x72, 0x73, 0x12, 0x23, 0x2e, 0x70, 0x6f, 0x6f, 0x6c,
-	0x5f, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x73, 0x2e, 0x47, 0x65, 0x74, 0x4d, 0x69, 0x6e, 0x65, 0x72,
-	0x57, 0x6f, 0x72, 0x6b, 0x65, 0x72, 0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x1d,
-	0x2e, 0x70, 0x6f, 0x6f, 0x6c, 0x5f, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x73, 0x2e, 0x4d, 0x69, 0x6e,
-	0x65, 0x72, 0x57, 0x6f, 0x72, 0x6b, 0x65, 0x72, 0x73, 0x4c, 0x69, 0x73, 0x74, 0x12, 0x5d, 0x0a,
-	0x18, 0x47, 0x65, 0x74, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x73, 0x57, 0x6f, 0x72, 0x6b, 0x65, 0x72,
-	0x73, 0x46, 0x72, 0x6f, 0x6d, 0x4c, 0x69, 0x73, 0x74, 0x12, 0x22, 0x2e, 0x70, 0x6f, 0x6f, 0x6c,
-	0x5f, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x73, 0x2e, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x41, 0x64, 0x64,
-	0x72, 0x65, 0x73, 0x73, 0x65, 0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x1d, 0x2e,
-	0x70, 0x6f, 0x6f, 0x6c, 0x5f, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x73, 0x2e, 0x4d, 0x69, 0x6e, 0x65,
-	0x72, 0x73, 0x57, 0x6f, 0x72, 0x6b, 0x65, 0x72, 0x73, 0x4d, 0x61, 0x70, 0x42, 0x51, 0x5a, 0x4f,
-	0x67, 0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d, 0x2f, 0x67, 0x72, 0x61, 0x6e, 0x64,
-	0x6d, 0x69, 0x6e, 0x69, 0x6e, 0x67, 0x70, 0x6f, 0x6f, 0x6c, 0x2f, 0x70, 0x6f, 0x6f, 0x6c, 0x2d,
-	0x61, 0x70, 0x69, 0x2d, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2f, 0x67, 0x65, 0x6e, 0x65, 0x72, 0x61,
-	0x74, 0x65, 0x64, 0x2f, 0x70, 0x6f, 0x6f, 0x6c, 0x5f, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x73, 0x3b,
-	0x70, 0x6f, 0x6f, 0x6c, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x73, 0x50, 0x72, 0x6f, 0x74, 0x6f, 0x62,
-	0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
-}
+const file_pool_miners_proto_rawDesc = "" +
+	"\n" +
+	"\x11pool_miners.proto\x12\vpool_miners\x1a\x1fgoogle/protobuf/timestamp.proto\x1a\x16utils/pagination.proto\x1a\x11utils/sorts.proto\x1a\x13utils/filters.proto\"/\n" +
+	"\x13MinerAddressRequest\x12\x18\n" +
+	"\aaddress\x18\x01 \x01(\tR\aaddress\"5\n" +
+	"\x15MinerAddressesRequest\x12\x1c\n" +
+	"\taddresses\x18\x01 \x03(\tR\taddresses\"/\n" +
+	"\x17ValidateAddressResponse\x12\x14\n" +
+	"\x05valid\x18\x01 \x01(\bR\x05valid\"\xd3\x03\n" +
+	"\rMinersFilters\x12\x1d\n" +
+	"\aaddress\x18\x01 \x01(\tH\x00R\aaddress\x88\x01\x01\x12;\n" +
+	"\bhashrate\x18\x02 \x01(\v2\x1a.filters.BigIntRangeFilterH\x01R\bhashrate\x88\x01\x01\x12?\n" +
+	"\rworkers_count\x18\x03 \x01(\v2\x15.filters.UInt32FilterH\x02R\fworkersCount\x88\x01\x01\x12=\n" +
+	"\fblocks_count\x18\x04 \x01(\v2\x15.filters.UInt32FilterH\x03R\vblocksCount\x88\x01\x01\x12F\n" +
+	"\x11solo_blocks_count\x18\x05 \x01(\v2\x15.filters.UInt32FilterH\x04R\x0fsoloBlocksCount\x88\x01\x01\x12>\n" +
+	"\tjoined_at\x18\x06 \x01(\v2\x1c.filters.DateTimeRangeFilterH\x05R\bjoinedAt\x88\x01\x01B\n" +
+	"\n" +
+	"\b_addressB\v\n" +
+	"\t_hashrateB\x10\n" +
+	"\x0e_workers_countB\x0f\n" +
+	"\r_blocks_countB\x14\n" +
+	"\x12_solo_blocks_countB\f\n" +
+	"\n" +
+	"_joined_at\"\xbe\x03\n" +
+	"\vMinersSorts\x12/\n" +
+	"\aaddress\x18\x01 \x01(\v2\x10.sorts.SortOrderH\x00R\aaddress\x88\x01\x01\x121\n" +
+	"\bhashrate\x18\x02 \x01(\v2\x10.sorts.SortOrderH\x01R\bhashrate\x88\x01\x01\x12:\n" +
+	"\rworkers_count\x18\x03 \x01(\v2\x10.sorts.SortOrderH\x02R\fworkersCount\x88\x01\x01\x128\n" +
+	"\fblocks_count\x18\x04 \x01(\v2\x10.sorts.SortOrderH\x03R\vblocksCount\x88\x01\x01\x12A\n" +
+	"\x11solo_blocks_count\x18\x05 \x01(\v2\x10.sorts.SortOrderH\x04R\x0fsoloBlocksCount\x88\x01\x01\x122\n" +
+	"\tjoined_at\x18\x06 \x01(\v2\x10.sorts.SortOrderH\x05R\bjoinedAt\x88\x01\x01B\n" +
+	"\n" +
+	"\b_addressB\v\n" +
+	"\t_hashrateB\x10\n" +
+	"\x0e_workers_countB\x0f\n" +
+	"\r_blocks_countB\x14\n" +
+	"\x12_solo_blocks_countB\f\n" +
+	"\n" +
+	"_joined_at\"\xb7\x01\n" +
+	"\x10GetMinersRequest\x124\n" +
+	"\afilters\x18\x01 \x01(\v2\x1a.pool_miners.MinersFiltersR\afilters\x12.\n" +
+	"\x05sorts\x18\x02 \x01(\v2\x18.pool_miners.MinersSortsR\x05sorts\x12=\n" +
+	"\n" +
+	"pagination\x18\x03 \x01(\v2\x1d.pagination.PaginationRequestR\n" +
+	"pagination\"\xab\x02\n" +
+	"\x05Miner\x12\x18\n" +
+	"\aaddress\x18\x01 \x01(\tR\aaddress\x12\x1a\n" +
+	"\bhashrate\x18\x02 \x01(\fR\bhashrate\x12#\n" +
+	"\rworkers_count\x18\x03 \x01(\rR\fworkersCount\x12!\n" +
+	"\fblocks_count\x18\x04 \x01(\rR\vblocksCount\x12*\n" +
+	"\x11solo_blocks_count\x18\x05 \x01(\rR\x0fsoloBlocksCount\x127\n" +
+	"\tjoined_at\x18\x06 \x01(\v2\x1a.google.protobuf.TimestampR\bjoinedAt\x12?\n" +
+	"\rlast_activity\x18\a \x01(\v2\x1a.google.protobuf.TimestampR\flastActivity\"x\n" +
+	"\n" +
+	"MinersList\x12*\n" +
+	"\x06miners\x18\x01 \x03(\v2\x12.pool_miners.MinerR\x06miners\x12>\n" +
+	"\n" +
+	"pagination\x18\x02 \x01(\v2\x1e.pagination.PaginationResponseR\n" +
+	"pagination\"\xc2\x01\n" +
+	"\vMinerWorker\x12\x16\n" +
+	"\x06worker\x18\x01 \x01(\tR\x06worker\x12\x16\n" +
+	"\x06region\x18\x02 \x01(\tR\x06region\x12\x14\n" +
+	"\x05agent\x18\x03 \x01(\tR\x05agent\x12\x12\n" +
+	"\x04solo\x18\x04 \x01(\bR\x04solo\x12\x1a\n" +
+	"\bhashrate\x18\x05 \x01(\fR\bhashrate\x12=\n" +
+	"\fconnected_at\x18\x06 \x01(\v2\x1a.google.protobuf.TimestampR\vconnectedAt\"B\n" +
+	"\fMinerWorkers\x122\n" +
+	"\aworkers\x18\x01 \x03(\v2\x18.pool_miners.MinerWorkerR\aworkers\"\xaf\x01\n" +
+	"\x10MinersWorkersMap\x12D\n" +
+	"\aworkers\x18\x01 \x03(\v2*.pool_miners.MinersWorkersMap.WorkersEntryR\aworkers\x1aU\n" +
+	"\fWorkersEntry\x12\x10\n" +
+	"\x03key\x18\x01 \x01(\tR\x03key\x12/\n" +
+	"\x05value\x18\x02 \x01(\v2\x19.pool_miners.MinerWorkersR\x05value:\x028\x01\"\xfd\x02\n" +
+	"\x11MinerWorkersSorts\x12-\n" +
+	"\x06worker\x18\x01 \x01(\v2\x10.sorts.SortOrderH\x00R\x06worker\x88\x01\x01\x12-\n" +
+	"\x06region\x18\x02 \x01(\v2\x10.sorts.SortOrderH\x01R\x06region\x88\x01\x01\x12+\n" +
+	"\x05agent\x18\x03 \x01(\v2\x10.sorts.SortOrderH\x02R\x05agent\x88\x01\x01\x12)\n" +
+	"\x04solo\x18\x04 \x01(\v2\x10.sorts.SortOrderH\x03R\x04solo\x88\x01\x01\x121\n" +
+	"\bhashrate\x18\x05 \x01(\v2\x10.sorts.SortOrderH\x04R\bhashrate\x88\x01\x01\x128\n" +
+	"\fconnected_at\x18\x06 \x01(\v2\x10.sorts.SortOrderH\x05R\vconnectedAt\x88\x01\x01B\t\n" +
+	"\a_workerB\t\n" +
+	"\a_regionB\b\n" +
+	"\x06_agentB\a\n" +
+	"\x05_soloB\v\n" +
+	"\t_hashrateB\x0f\n" +
+	"\r_connected_at\"\xcd\x02\n" +
+	"\x13MinerWorkersFilters\x12\x1b\n" +
+	"\x06worker\x18\x01 \x01(\tH\x00R\x06worker\x88\x01\x01\x12\x1b\n" +
+	"\x06region\x18\x02 \x01(\tH\x01R\x06region\x88\x01\x01\x12\x19\n" +
+	"\x05agent\x18\x03 \x01(\tH\x02R\x05agent\x88\x01\x01\x12\x17\n" +
+	"\x04solo\x18\x04 \x01(\bH\x03R\x04solo\x88\x01\x01\x12;\n" +
+	"\bhashrate\x18\x05 \x01(\v2\x1a.filters.BigIntRangeFilterH\x04R\bhashrate\x88\x01\x01\x12D\n" +
+	"\fconnected_at\x18\x06 \x01(\v2\x1c.filters.DateTimeRangeFilterH\x05R\vconnectedAt\x88\x01\x01B\t\n" +
+	"\a_workerB\t\n" +
+	"\a_regionB\b\n" +
+	"\x06_agentB\a\n" +
+	"\x05_soloB\v\n" +
+	"\t_hashrateB\x0f\n" +
+	"\r_connected_at\"\xe3\x01\n" +
+	"\x16GetMinerWorkersRequest\x12\x18\n" +
+	"\aaddress\x18\x01 \x01(\tR\aaddress\x124\n" +
+	"\x05sorts\x18\x02 \x01(\v2\x1e.pool_miners.MinerWorkersSortsR\x05sorts\x12:\n" +
+	"\afilters\x18\x03 \x01(\v2 .pool_miners.MinerWorkersFiltersR\afilters\x12=\n" +
+	"\n" +
+	"pagination\x18\x04 \x01(\v2\x1d.pagination.PaginationRequestR\n" +
+	"pagination\"\x87\x01\n" +
+	"\x10MinerWorkersList\x123\n" +
+	"\aworkers\x18\x01 \x01(\v2\x19.pool_miners.MinerWorkersR\aworkers\x12>\n" +
+	"\n" +
+	"pagination\x18\x02 \x01(\v2\x1e.pagination.PaginationResponseR\n" +
+	"pagination2\xab\x03\n" +
+	"\x11PoolMinersService\x12Y\n" +
+	"\x0fValidateAddress\x12 .pool_miners.MinerAddressRequest\x1a$.pool_miners.ValidateAddressResponse\x12C\n" +
+	"\tGetMiners\x12\x1d.pool_miners.GetMinersRequest\x1a\x17.pool_miners.MinersList\x12@\n" +
+	"\bGetMiner\x12 .pool_miners.MinerAddressRequest\x1a\x12.pool_miners.Miner\x12U\n" +
+	"\x0fGetMinerWorkers\x12#.pool_miners.GetMinerWorkersRequest\x1a\x1d.pool_miners.MinerWorkersList\x12]\n" +
+	"\x18GetMinersWorkersFromList\x12\".pool_miners.MinerAddressesRequest\x1a\x1d.pool_miners.MinersWorkersMapBMZKgithub.com/grandminingpool/pool-api-proto/generated/pool_miners;pool_minersb\x06proto3"
 
 var (
 	file_pool_miners_proto_rawDescOnce sync.Once
-	file_pool_miners_proto_rawDescData = file_pool_miners_proto_rawDesc
+	file_pool_miners_proto_rawDescData []byte
 )
 
 func file_pool_miners_proto_rawDescGZIP() []byte {
 	file_pool_miners_proto_rawDescOnce.Do(func() {
-		file_pool_miners_proto_rawDescData = protoimpl.X.CompressGZIP(file_pool_miners_proto_rawDescData)
+		file_pool_miners_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_pool_miners_proto_rawDesc), len(file_pool_miners_proto_rawDesc)))
 	})
 	return file_pool_miners_proto_rawDescData
 }
 
 var file_pool_miners_proto_msgTypes = make([]protoimpl.MessageInfo, 16)
-var file_pool_miners_proto_goTypes = []interface{}{
+var file_pool_miners_proto_goTypes = []any{
 	(*MinerAddressRequest)(nil),           // 0: pool_miners.MinerAddressRequest
 	(*MinerAddressesRequest)(nil),         // 1: pool_miners.MinerAddressesRequest
 	(*ValidateAddressResponse)(nil),       // 2: pool_miners.ValidateAddressResponse
@@ -1376,197 +1205,15 @@ func file_pool_miners_proto_init() {
 	if File_pool_miners_proto != nil {
 		return
 	}
-	if !protoimpl.UnsafeEnabled {
-		file_pool_miners_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*MinerAddressRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_pool_miners_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*MinerAddressesRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_pool_miners_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*ValidateAddressResponse); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_pool_miners_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*MinersFilters); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_pool_miners_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*MinersSorts); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_pool_miners_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*GetMinersRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_pool_miners_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Miner); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_pool_miners_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*MinersList); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_pool_miners_proto_msgTypes[8].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*MinerWorker); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_pool_miners_proto_msgTypes[9].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*MinerWorkers); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_pool_miners_proto_msgTypes[10].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*MinersWorkersMap); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_pool_miners_proto_msgTypes[11].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*MinerWorkersSorts); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_pool_miners_proto_msgTypes[12].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*MinerWorkersFilters); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_pool_miners_proto_msgTypes[13].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*GetMinerWorkersRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_pool_miners_proto_msgTypes[14].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*MinerWorkersList); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-	}
-	file_pool_miners_proto_msgTypes[3].OneofWrappers = []interface{}{}
-	file_pool_miners_proto_msgTypes[4].OneofWrappers = []interface{}{}
-	file_pool_miners_proto_msgTypes[11].OneofWrappers = []interface{}{}
-	file_pool_miners_proto_msgTypes[12].OneofWrappers = []interface{}{}
+	file_pool_miners_proto_msgTypes[3].OneofWrappers = []any{}
+	file_pool_miners_proto_msgTypes[4].OneofWrappers = []any{}
+	file_pool_miners_proto_msgTypes[11].OneofWrappers = []any{}
+	file_pool_miners_proto_msgTypes[12].OneofWrappers = []any{}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
-			RawDescriptor: file_pool_miners_proto_rawDesc,
+			RawDescriptor: unsafe.Slice(unsafe.StringData(file_pool_miners_proto_rawDesc), len(file_pool_miners_proto_rawDesc)),
 			NumEnums:      0,
 			NumMessages:   16,
 			NumExtensions: 0,
@@ -1577,7 +1224,6 @@ func file_pool_miners_proto_init() {
 		MessageInfos:      file_pool_miners_proto_msgTypes,
 	}.Build()
 	File_pool_miners_proto = out.File
-	file_pool_miners_proto_rawDesc = nil
 	file_pool_miners_proto_goTypes = nil
 	file_pool_miners_proto_depIdxs = nil
 }
