@@ -1,10 +1,10 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.28.1
-// 	protoc        v5.26.1
+// 	protoc-gen-go v1.36.6
+// 	protoc        v5.29.3
 // source: pool_payouts.proto
 
-package poolPayoutsProto
+package pool_payouts
 
 import (
 	pool_miners "github.com/grandminingpool/pool-api-proto/generated/pool_miners"
@@ -16,6 +16,7 @@ import (
 	timestamppb "google.golang.org/protobuf/types/known/timestamppb"
 	reflect "reflect"
 	sync "sync"
+	unsafe "unsafe"
 )
 
 const (
@@ -26,20 +27,17 @@ const (
 )
 
 type MinerBalance struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Balance       uint64                 `protobuf:"varint,1,opt,name=balance,proto3" json:"balance,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Balance uint64 `protobuf:"varint,1,opt,name=balance,proto3" json:"balance,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *MinerBalance) Reset() {
 	*x = MinerBalance{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_payouts_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_payouts_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *MinerBalance) String() string {
@@ -50,7 +48,7 @@ func (*MinerBalance) ProtoMessage() {}
 
 func (x *MinerBalance) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_payouts_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -73,20 +71,17 @@ func (x *MinerBalance) GetBalance() uint64 {
 }
 
 type MinersBalancesMap struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState   `protogen:"open.v1"`
+	Balances      map[string]*MinerBalance `protobuf:"bytes,1,rep,name=balances,proto3" json:"balances,omitempty" protobuf_key:"bytes,1,opt,name=key" protobuf_val:"bytes,2,opt,name=value"`
 	unknownFields protoimpl.UnknownFields
-
-	Balances map[string]*MinerBalance `protobuf:"bytes,1,rep,name=balances,proto3" json:"balances,omitempty" protobuf_key:"bytes,1,opt,name=key,proto3" protobuf_val:"bytes,2,opt,name=value,proto3"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *MinersBalancesMap) Reset() {
 	*x = MinersBalancesMap{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_payouts_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_payouts_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *MinersBalancesMap) String() string {
@@ -97,7 +92,7 @@ func (*MinersBalancesMap) ProtoMessage() {}
 
 func (x *MinersBalancesMap) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_payouts_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -120,23 +115,20 @@ func (x *MinersBalancesMap) GetBalances() map[string]*MinerBalance {
 }
 
 type PayoutsFilters struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState       `protogen:"open.v1"`
+	Miner         *string                      `protobuf:"bytes,1,opt,name=miner,proto3,oneof" json:"miner,omitempty"`
+	TxHash        *string                      `protobuf:"bytes,2,opt,name=tx_hash,json=txHash,proto3,oneof" json:"tx_hash,omitempty"`
+	Amount        *filters.UInt64RangeFilter   `protobuf:"bytes,3,opt,name=amount,proto3,oneof" json:"amount,omitempty"`
+	PaidAt        *filters.DateTimeRangeFilter `protobuf:"bytes,4,opt,name=paid_at,json=paidAt,proto3,oneof" json:"paid_at,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Miner  *string                      `protobuf:"bytes,1,opt,name=miner,proto3,oneof" json:"miner,omitempty"`
-	TxHash *string                      `protobuf:"bytes,2,opt,name=tx_hash,json=txHash,proto3,oneof" json:"tx_hash,omitempty"`
-	Amount *filters.UInt64RangeFilter   `protobuf:"bytes,3,opt,name=amount,proto3,oneof" json:"amount,omitempty"`
-	PaidAt *filters.DateTimeRangeFilter `protobuf:"bytes,4,opt,name=paid_at,json=paidAt,proto3,oneof" json:"paid_at,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *PayoutsFilters) Reset() {
 	*x = PayoutsFilters{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_payouts_proto_msgTypes[2]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_payouts_proto_msgTypes[2]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *PayoutsFilters) String() string {
@@ -147,7 +139,7 @@ func (*PayoutsFilters) ProtoMessage() {}
 
 func (x *PayoutsFilters) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_payouts_proto_msgTypes[2]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -191,23 +183,20 @@ func (x *PayoutsFilters) GetPaidAt() *filters.DateTimeRangeFilter {
 }
 
 type PayoutsSorts struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Miner         *sorts.SortOrder       `protobuf:"bytes,1,opt,name=miner,proto3,oneof" json:"miner,omitempty"`
+	TxHash        *sorts.SortOrder       `protobuf:"bytes,2,opt,name=tx_hash,json=txHash,proto3,oneof" json:"tx_hash,omitempty"`
+	Amount        *sorts.SortOrder       `protobuf:"bytes,3,opt,name=amount,proto3,oneof" json:"amount,omitempty"`
+	PaidAt        *sorts.SortOrder       `protobuf:"bytes,4,opt,name=paid_at,json=paidAt,proto3,oneof" json:"paid_at,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Miner  *sorts.SortOrder `protobuf:"bytes,1,opt,name=miner,proto3,oneof" json:"miner,omitempty"`
-	TxHash *sorts.SortOrder `protobuf:"bytes,2,opt,name=tx_hash,json=txHash,proto3,oneof" json:"tx_hash,omitempty"`
-	Amount *sorts.SortOrder `protobuf:"bytes,3,opt,name=amount,proto3,oneof" json:"amount,omitempty"`
-	PaidAt *sorts.SortOrder `protobuf:"bytes,4,opt,name=paid_at,json=paidAt,proto3,oneof" json:"paid_at,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *PayoutsSorts) Reset() {
 	*x = PayoutsSorts{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_payouts_proto_msgTypes[3]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_payouts_proto_msgTypes[3]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *PayoutsSorts) String() string {
@@ -218,7 +207,7 @@ func (*PayoutsSorts) ProtoMessage() {}
 
 func (x *PayoutsSorts) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_payouts_proto_msgTypes[3]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -262,22 +251,19 @@ func (x *PayoutsSorts) GetPaidAt() *sorts.SortOrder {
 }
 
 type GetPayoutsRequest struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState        `protogen:"open.v1"`
+	Sorts         *PayoutsSorts                 `protobuf:"bytes,1,opt,name=sorts,proto3" json:"sorts,omitempty"`
+	Filters       *PayoutsFilters               `protobuf:"bytes,2,opt,name=filters,proto3" json:"filters,omitempty"`
+	Pagination    *pagination.PaginationRequest `protobuf:"bytes,3,opt,name=pagination,proto3" json:"pagination,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Sorts      *PayoutsSorts                 `protobuf:"bytes,1,opt,name=sorts,proto3" json:"sorts,omitempty"`
-	Filters    *PayoutsFilters               `protobuf:"bytes,2,opt,name=filters,proto3" json:"filters,omitempty"`
-	Pagination *pagination.PaginationRequest `protobuf:"bytes,3,opt,name=pagination,proto3" json:"pagination,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *GetPayoutsRequest) Reset() {
 	*x = GetPayoutsRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_payouts_proto_msgTypes[4]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_payouts_proto_msgTypes[4]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *GetPayoutsRequest) String() string {
@@ -288,7 +274,7 @@ func (*GetPayoutsRequest) ProtoMessage() {}
 
 func (x *GetPayoutsRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_payouts_proto_msgTypes[4]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -325,21 +311,18 @@ func (x *GetPayoutsRequest) GetPagination() *pagination.PaginationRequest {
 }
 
 type GetPayoutsFromListRequest struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Miners        []string               `protobuf:"bytes,1,rep,name=miners,proto3" json:"miners,omitempty"`
+	Filters       *PayoutsFilters        `protobuf:"bytes,2,opt,name=filters,proto3" json:"filters,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Miners  []string        `protobuf:"bytes,1,rep,name=miners,proto3" json:"miners,omitempty"`
-	Filters *PayoutsFilters `protobuf:"bytes,2,opt,name=filters,proto3" json:"filters,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *GetPayoutsFromListRequest) Reset() {
 	*x = GetPayoutsFromListRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_payouts_proto_msgTypes[5]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_payouts_proto_msgTypes[5]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *GetPayoutsFromListRequest) String() string {
@@ -350,7 +333,7 @@ func (*GetPayoutsFromListRequest) ProtoMessage() {}
 
 func (x *GetPayoutsFromListRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_payouts_proto_msgTypes[5]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -380,23 +363,20 @@ func (x *GetPayoutsFromListRequest) GetFilters() *PayoutsFilters {
 }
 
 type Payout struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Miner         string                 `protobuf:"bytes,1,opt,name=miner,proto3" json:"miner,omitempty"`
+	TxHash        string                 `protobuf:"bytes,2,opt,name=tx_hash,json=txHash,proto3" json:"tx_hash,omitempty"`
+	Amount        uint64                 `protobuf:"varint,3,opt,name=amount,proto3" json:"amount,omitempty"`
+	PaidAt        *timestamppb.Timestamp `protobuf:"bytes,4,opt,name=paid_at,json=paidAt,proto3" json:"paid_at,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Miner  string                 `protobuf:"bytes,1,opt,name=miner,proto3" json:"miner,omitempty"`
-	TxHash string                 `protobuf:"bytes,2,opt,name=tx_hash,json=txHash,proto3" json:"tx_hash,omitempty"`
-	Amount uint64                 `protobuf:"varint,3,opt,name=amount,proto3" json:"amount,omitempty"`
-	PaidAt *timestamppb.Timestamp `protobuf:"bytes,4,opt,name=paid_at,json=paidAt,proto3" json:"paid_at,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *Payout) Reset() {
 	*x = Payout{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_payouts_proto_msgTypes[6]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_payouts_proto_msgTypes[6]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Payout) String() string {
@@ -407,7 +387,7 @@ func (*Payout) ProtoMessage() {}
 
 func (x *Payout) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_payouts_proto_msgTypes[6]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -451,20 +431,17 @@ func (x *Payout) GetPaidAt() *timestamppb.Timestamp {
 }
 
 type Payouts struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Payouts       []*Payout              `protobuf:"bytes,1,rep,name=payouts,proto3" json:"payouts,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Payouts []*Payout `protobuf:"bytes,1,rep,name=payouts,proto3" json:"payouts,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *Payouts) Reset() {
 	*x = Payouts{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_payouts_proto_msgTypes[7]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_payouts_proto_msgTypes[7]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Payouts) String() string {
@@ -475,7 +452,7 @@ func (*Payouts) ProtoMessage() {}
 
 func (x *Payouts) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_payouts_proto_msgTypes[7]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -498,21 +475,18 @@ func (x *Payouts) GetPayouts() []*Payout {
 }
 
 type PayoutsList struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState         `protogen:"open.v1"`
+	Payouts       *Payouts                       `protobuf:"bytes,1,opt,name=payouts,proto3" json:"payouts,omitempty"`
+	Pagination    *pagination.PaginationResponse `protobuf:"bytes,2,opt,name=pagination,proto3" json:"pagination,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Payouts    *Payouts                       `protobuf:"bytes,1,opt,name=payouts,proto3" json:"payouts,omitempty"`
-	Pagination *pagination.PaginationResponse `protobuf:"bytes,2,opt,name=pagination,proto3" json:"pagination,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *PayoutsList) Reset() {
 	*x = PayoutsList{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_payouts_proto_msgTypes[8]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_payouts_proto_msgTypes[8]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *PayoutsList) String() string {
@@ -523,7 +497,7 @@ func (*PayoutsList) ProtoMessage() {}
 
 func (x *PayoutsList) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_payouts_proto_msgTypes[8]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -553,20 +527,17 @@ func (x *PayoutsList) GetPagination() *pagination.PaginationResponse {
 }
 
 type PayoutsMap struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Payouts       map[string]*Payouts    `protobuf:"bytes,1,rep,name=payouts,proto3" json:"payouts,omitempty" protobuf_key:"bytes,1,opt,name=key" protobuf_val:"bytes,2,opt,name=value"`
 	unknownFields protoimpl.UnknownFields
-
-	Payouts map[string]*Payouts `protobuf:"bytes,1,rep,name=payouts,proto3" json:"payouts,omitempty" protobuf_key:"bytes,1,opt,name=key,proto3" protobuf_val:"bytes,2,opt,name=value,proto3"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *PayoutsMap) Reset() {
 	*x = PayoutsMap{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_payouts_proto_msgTypes[9]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_payouts_proto_msgTypes[9]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *PayoutsMap) String() string {
@@ -577,7 +548,7 @@ func (*PayoutsMap) ProtoMessage() {}
 
 func (x *PayoutsMap) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_payouts_proto_msgTypes[9]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -600,25 +571,22 @@ func (x *PayoutsMap) GetPayouts() map[string]*Payouts {
 }
 
 type MinedBlocksSorts struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	Miner            *sorts.SortOrder `protobuf:"bytes,1,opt,name=miner,proto3,oneof" json:"miner,omitempty"`
-	MinerHashrate    *sorts.SortOrder `protobuf:"bytes,2,opt,name=miner_hashrate,json=minerHashrate,proto3,oneof" json:"miner_hashrate,omitempty"`
-	BlockHeight      *sorts.SortOrder `protobuf:"bytes,3,opt,name=block_height,json=blockHeight,proto3,oneof" json:"block_height,omitempty"`
-	BlockHash        *sorts.SortOrder `protobuf:"bytes,4,opt,name=block_hash,json=blockHash,proto3,oneof" json:"block_hash,omitempty"`
-	RoundMinersCount *sorts.SortOrder `protobuf:"bytes,5,opt,name=round_miners_count,json=roundMinersCount,proto3,oneof" json:"round_miners_count,omitempty"`
-	MinedAt          *sorts.SortOrder `protobuf:"bytes,6,opt,name=mined_at,json=minedAt,proto3,oneof" json:"mined_at,omitempty"`
+	state            protoimpl.MessageState `protogen:"open.v1"`
+	Miner            *sorts.SortOrder       `protobuf:"bytes,1,opt,name=miner,proto3,oneof" json:"miner,omitempty"`
+	MinerHashrate    *sorts.SortOrder       `protobuf:"bytes,2,opt,name=miner_hashrate,json=minerHashrate,proto3,oneof" json:"miner_hashrate,omitempty"`
+	BlockHeight      *sorts.SortOrder       `protobuf:"bytes,3,opt,name=block_height,json=blockHeight,proto3,oneof" json:"block_height,omitempty"`
+	BlockHash        *sorts.SortOrder       `protobuf:"bytes,4,opt,name=block_hash,json=blockHash,proto3,oneof" json:"block_hash,omitempty"`
+	RoundMinersCount *sorts.SortOrder       `protobuf:"bytes,5,opt,name=round_miners_count,json=roundMinersCount,proto3,oneof" json:"round_miners_count,omitempty"`
+	MinedAt          *sorts.SortOrder       `protobuf:"bytes,6,opt,name=mined_at,json=minedAt,proto3,oneof" json:"mined_at,omitempty"`
+	unknownFields    protoimpl.UnknownFields
+	sizeCache        protoimpl.SizeCache
 }
 
 func (x *MinedBlocksSorts) Reset() {
 	*x = MinedBlocksSorts{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_payouts_proto_msgTypes[10]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_payouts_proto_msgTypes[10]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *MinedBlocksSorts) String() string {
@@ -629,7 +597,7 @@ func (*MinedBlocksSorts) ProtoMessage() {}
 
 func (x *MinedBlocksSorts) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_payouts_proto_msgTypes[10]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -687,25 +655,22 @@ func (x *MinedBlocksSorts) GetMinedAt() *sorts.SortOrder {
 }
 
 type MinedBlocksFilters struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
+	state            protoimpl.MessageState       `protogen:"open.v1"`
 	Miner            *string                      `protobuf:"bytes,1,opt,name=miner,proto3,oneof" json:"miner,omitempty"`
 	MinerHashrate    *filters.BigIntRangeFilter   `protobuf:"bytes,2,opt,name=miner_hashrate,json=minerHashrate,proto3,oneof" json:"miner_hashrate,omitempty"`
 	BlockHeight      *filters.UInt64Filter        `protobuf:"bytes,3,opt,name=block_height,json=blockHeight,proto3,oneof" json:"block_height,omitempty"`
 	BlockHash        *string                      `protobuf:"bytes,4,opt,name=block_hash,json=blockHash,proto3,oneof" json:"block_hash,omitempty"`
 	RoundMinersCount *filters.UInt32Filter        `protobuf:"bytes,5,opt,name=round_miners_count,json=roundMinersCount,proto3,oneof" json:"round_miners_count,omitempty"`
 	MinedAt          *filters.DateTimeRangeFilter `protobuf:"bytes,6,opt,name=mined_at,json=minedAt,proto3,oneof" json:"mined_at,omitempty"`
+	unknownFields    protoimpl.UnknownFields
+	sizeCache        protoimpl.SizeCache
 }
 
 func (x *MinedBlocksFilters) Reset() {
 	*x = MinedBlocksFilters{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_payouts_proto_msgTypes[11]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_payouts_proto_msgTypes[11]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *MinedBlocksFilters) String() string {
@@ -716,7 +681,7 @@ func (*MinedBlocksFilters) ProtoMessage() {}
 
 func (x *MinedBlocksFilters) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_payouts_proto_msgTypes[11]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -774,22 +739,19 @@ func (x *MinedBlocksFilters) GetMinedAt() *filters.DateTimeRangeFilter {
 }
 
 type GetBlocksRequest struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState        `protogen:"open.v1"`
+	Sorts         *MinedBlocksSorts             `protobuf:"bytes,1,opt,name=sorts,proto3" json:"sorts,omitempty"`
+	Filters       *MinedBlocksFilters           `protobuf:"bytes,2,opt,name=filters,proto3" json:"filters,omitempty"`
+	Pagination    *pagination.PaginationRequest `protobuf:"bytes,3,opt,name=pagination,proto3" json:"pagination,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Sorts      *MinedBlocksSorts             `protobuf:"bytes,1,opt,name=sorts,proto3" json:"sorts,omitempty"`
-	Filters    *MinedBlocksFilters           `protobuf:"bytes,2,opt,name=filters,proto3" json:"filters,omitempty"`
-	Pagination *pagination.PaginationRequest `protobuf:"bytes,3,opt,name=pagination,proto3" json:"pagination,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *GetBlocksRequest) Reset() {
 	*x = GetBlocksRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_payouts_proto_msgTypes[12]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_payouts_proto_msgTypes[12]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *GetBlocksRequest) String() string {
@@ -800,7 +762,7 @@ func (*GetBlocksRequest) ProtoMessage() {}
 
 func (x *GetBlocksRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_payouts_proto_msgTypes[12]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -837,25 +799,22 @@ func (x *GetBlocksRequest) GetPagination() *pagination.PaginationRequest {
 }
 
 type MinedBlock struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
+	state            protoimpl.MessageState `protogen:"open.v1"`
 	Miner            string                 `protobuf:"bytes,1,opt,name=miner,proto3" json:"miner,omitempty"`
 	MinerHashrate    []byte                 `protobuf:"bytes,2,opt,name=miner_hashrate,json=minerHashrate,proto3" json:"miner_hashrate,omitempty"`
 	BlockHeight      uint64                 `protobuf:"varint,3,opt,name=block_height,json=blockHeight,proto3" json:"block_height,omitempty"`
 	BlockHash        string                 `protobuf:"bytes,4,opt,name=block_hash,json=blockHash,proto3" json:"block_hash,omitempty"`
 	RoundMinersCount uint32                 `protobuf:"varint,5,opt,name=round_miners_count,json=roundMinersCount,proto3" json:"round_miners_count,omitempty"`
 	MinedAt          *timestamppb.Timestamp `protobuf:"bytes,6,opt,name=mined_at,json=minedAt,proto3" json:"mined_at,omitempty"`
+	unknownFields    protoimpl.UnknownFields
+	sizeCache        protoimpl.SizeCache
 }
 
 func (x *MinedBlock) Reset() {
 	*x = MinedBlock{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_payouts_proto_msgTypes[13]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_payouts_proto_msgTypes[13]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *MinedBlock) String() string {
@@ -866,7 +825,7 @@ func (*MinedBlock) ProtoMessage() {}
 
 func (x *MinedBlock) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_payouts_proto_msgTypes[13]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -924,21 +883,18 @@ func (x *MinedBlock) GetMinedAt() *timestamppb.Timestamp {
 }
 
 type MinedBlocksList struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState         `protogen:"open.v1"`
+	Blocks        []*MinedBlock                  `protobuf:"bytes,1,rep,name=blocks,proto3" json:"blocks,omitempty"`
+	Pagination    *pagination.PaginationResponse `protobuf:"bytes,2,opt,name=pagination,proto3" json:"pagination,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Blocks     []*MinedBlock                  `protobuf:"bytes,1,rep,name=blocks,proto3" json:"blocks,omitempty"`
-	Pagination *pagination.PaginationResponse `protobuf:"bytes,2,opt,name=pagination,proto3" json:"pagination,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *MinedBlocksList) Reset() {
 	*x = MinedBlocksList{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_payouts_proto_msgTypes[14]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_payouts_proto_msgTypes[14]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *MinedBlocksList) String() string {
@@ -949,7 +905,7 @@ func (*MinedBlocksList) ProtoMessage() {}
 
 func (x *MinedBlocksList) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_payouts_proto_msgTypes[14]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -979,26 +935,23 @@ func (x *MinedBlocksList) GetPagination() *pagination.PaginationResponse {
 }
 
 type MinedSoloBlocksSorts struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Miner         *sorts.SortOrder       `protobuf:"bytes,1,opt,name=miner,proto3,oneof" json:"miner,omitempty"`
+	MinerHashrate *sorts.SortOrder       `protobuf:"bytes,2,opt,name=miner_hashrate,json=minerHashrate,proto3,oneof" json:"miner_hashrate,omitempty"`
+	BlockHeight   *sorts.SortOrder       `protobuf:"bytes,3,opt,name=block_height,json=blockHeight,proto3,oneof" json:"block_height,omitempty"`
+	BlockHash     *sorts.SortOrder       `protobuf:"bytes,4,opt,name=block_hash,json=blockHash,proto3,oneof" json:"block_hash,omitempty"`
+	Reward        *sorts.SortOrder       `protobuf:"bytes,5,opt,name=reward,proto3,oneof" json:"reward,omitempty"`
+	TxHash        *sorts.SortOrder       `protobuf:"bytes,6,opt,name=tx_hash,json=txHash,proto3,oneof" json:"tx_hash,omitempty"`
+	MinedAt       *sorts.SortOrder       `protobuf:"bytes,7,opt,name=mined_at,json=minedAt,proto3,oneof" json:"mined_at,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Miner         *sorts.SortOrder `protobuf:"bytes,1,opt,name=miner,proto3,oneof" json:"miner,omitempty"`
-	MinerHashrate *sorts.SortOrder `protobuf:"bytes,2,opt,name=miner_hashrate,json=minerHashrate,proto3,oneof" json:"miner_hashrate,omitempty"`
-	BlockHeight   *sorts.SortOrder `protobuf:"bytes,3,opt,name=block_height,json=blockHeight,proto3,oneof" json:"block_height,omitempty"`
-	BlockHash     *sorts.SortOrder `protobuf:"bytes,4,opt,name=block_hash,json=blockHash,proto3,oneof" json:"block_hash,omitempty"`
-	Reward        *sorts.SortOrder `protobuf:"bytes,5,opt,name=reward,proto3,oneof" json:"reward,omitempty"`
-	TxHash        *sorts.SortOrder `protobuf:"bytes,6,opt,name=tx_hash,json=txHash,proto3,oneof" json:"tx_hash,omitempty"`
-	MinedAt       *sorts.SortOrder `protobuf:"bytes,7,opt,name=mined_at,json=minedAt,proto3,oneof" json:"mined_at,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *MinedSoloBlocksSorts) Reset() {
 	*x = MinedSoloBlocksSorts{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_payouts_proto_msgTypes[15]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_payouts_proto_msgTypes[15]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *MinedSoloBlocksSorts) String() string {
@@ -1009,7 +962,7 @@ func (*MinedSoloBlocksSorts) ProtoMessage() {}
 
 func (x *MinedSoloBlocksSorts) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_payouts_proto_msgTypes[15]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1074,10 +1027,7 @@ func (x *MinedSoloBlocksSorts) GetMinedAt() *sorts.SortOrder {
 }
 
 type MinedSoloBlocksFilters struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
+	state         protoimpl.MessageState       `protogen:"open.v1"`
 	Miner         *string                      `protobuf:"bytes,1,opt,name=miner,proto3,oneof" json:"miner,omitempty"`
 	MinerHashrate *filters.BigIntRangeFilter   `protobuf:"bytes,2,opt,name=miner_hashrate,json=minerHashrate,proto3,oneof" json:"miner_hashrate,omitempty"`
 	BlockHeight   *filters.UInt64Filter        `protobuf:"bytes,3,opt,name=block_height,json=blockHeight,proto3,oneof" json:"block_height,omitempty"`
@@ -1085,15 +1035,15 @@ type MinedSoloBlocksFilters struct {
 	Reward        *filters.UInt64RangeFilter   `protobuf:"bytes,5,opt,name=reward,proto3,oneof" json:"reward,omitempty"`
 	TxHash        *string                      `protobuf:"bytes,6,opt,name=tx_hash,json=txHash,proto3,oneof" json:"tx_hash,omitempty"`
 	MinedAt       *filters.DateTimeRangeFilter `protobuf:"bytes,7,opt,name=mined_at,json=minedAt,proto3,oneof" json:"mined_at,omitempty"`
+	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *MinedSoloBlocksFilters) Reset() {
 	*x = MinedSoloBlocksFilters{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_payouts_proto_msgTypes[16]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_payouts_proto_msgTypes[16]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *MinedSoloBlocksFilters) String() string {
@@ -1104,7 +1054,7 @@ func (*MinedSoloBlocksFilters) ProtoMessage() {}
 
 func (x *MinedSoloBlocksFilters) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_payouts_proto_msgTypes[16]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1169,22 +1119,19 @@ func (x *MinedSoloBlocksFilters) GetMinedAt() *filters.DateTimeRangeFilter {
 }
 
 type GetSoloBlocksRequest struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState        `protogen:"open.v1"`
+	Sorts         *MinedSoloBlocksSorts         `protobuf:"bytes,1,opt,name=sorts,proto3" json:"sorts,omitempty"`
+	Filters       *MinedSoloBlocksFilters       `protobuf:"bytes,2,opt,name=filters,proto3" json:"filters,omitempty"`
+	Pagination    *pagination.PaginationRequest `protobuf:"bytes,3,opt,name=pagination,proto3" json:"pagination,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Sorts      *MinedSoloBlocksSorts         `protobuf:"bytes,1,opt,name=sorts,proto3" json:"sorts,omitempty"`
-	Filters    *MinedSoloBlocksFilters       `protobuf:"bytes,2,opt,name=filters,proto3" json:"filters,omitempty"`
-	Pagination *pagination.PaginationRequest `protobuf:"bytes,3,opt,name=pagination,proto3" json:"pagination,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *GetSoloBlocksRequest) Reset() {
 	*x = GetSoloBlocksRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_payouts_proto_msgTypes[17]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_payouts_proto_msgTypes[17]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *GetSoloBlocksRequest) String() string {
@@ -1195,7 +1142,7 @@ func (*GetSoloBlocksRequest) ProtoMessage() {}
 
 func (x *GetSoloBlocksRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_payouts_proto_msgTypes[17]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1232,21 +1179,18 @@ func (x *GetSoloBlocksRequest) GetPagination() *pagination.PaginationRequest {
 }
 
 type GetSoloBlocksFromListRequest struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState  `protogen:"open.v1"`
+	Miners        []string                `protobuf:"bytes,1,rep,name=miners,proto3" json:"miners,omitempty"`
+	Filters       *MinedSoloBlocksFilters `protobuf:"bytes,2,opt,name=filters,proto3" json:"filters,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Miners  []string                `protobuf:"bytes,1,rep,name=miners,proto3" json:"miners,omitempty"`
-	Filters *MinedSoloBlocksFilters `protobuf:"bytes,2,opt,name=filters,proto3" json:"filters,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *GetSoloBlocksFromListRequest) Reset() {
 	*x = GetSoloBlocksFromListRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_payouts_proto_msgTypes[18]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_payouts_proto_msgTypes[18]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *GetSoloBlocksFromListRequest) String() string {
@@ -1257,7 +1201,7 @@ func (*GetSoloBlocksFromListRequest) ProtoMessage() {}
 
 func (x *GetSoloBlocksFromListRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_payouts_proto_msgTypes[18]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1287,10 +1231,7 @@ func (x *GetSoloBlocksFromListRequest) GetFilters() *MinedSoloBlocksFilters {
 }
 
 type MinedSoloBlock struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
+	state         protoimpl.MessageState `protogen:"open.v1"`
 	Miner         string                 `protobuf:"bytes,1,opt,name=miner,proto3" json:"miner,omitempty"`
 	MinerHashrate []byte                 `protobuf:"bytes,2,opt,name=miner_hashrate,json=minerHashrate,proto3" json:"miner_hashrate,omitempty"`
 	BlockHeight   uint64                 `protobuf:"varint,3,opt,name=block_height,json=blockHeight,proto3" json:"block_height,omitempty"`
@@ -1298,15 +1239,15 @@ type MinedSoloBlock struct {
 	Reward        uint64                 `protobuf:"varint,5,opt,name=reward,proto3" json:"reward,omitempty"`
 	TxHash        string                 `protobuf:"bytes,6,opt,name=tx_hash,json=txHash,proto3" json:"tx_hash,omitempty"`
 	MinedAt       *timestamppb.Timestamp `protobuf:"bytes,7,opt,name=mined_at,json=minedAt,proto3" json:"mined_at,omitempty"`
+	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *MinedSoloBlock) Reset() {
 	*x = MinedSoloBlock{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_payouts_proto_msgTypes[19]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_payouts_proto_msgTypes[19]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *MinedSoloBlock) String() string {
@@ -1317,7 +1258,7 @@ func (*MinedSoloBlock) ProtoMessage() {}
 
 func (x *MinedSoloBlock) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_payouts_proto_msgTypes[19]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1382,20 +1323,17 @@ func (x *MinedSoloBlock) GetMinedAt() *timestamppb.Timestamp {
 }
 
 type MinedSoloBlocks struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Blocks        []*MinedSoloBlock      `protobuf:"bytes,1,rep,name=blocks,proto3" json:"blocks,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Blocks []*MinedSoloBlock `protobuf:"bytes,1,rep,name=blocks,proto3" json:"blocks,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *MinedSoloBlocks) Reset() {
 	*x = MinedSoloBlocks{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_payouts_proto_msgTypes[20]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_payouts_proto_msgTypes[20]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *MinedSoloBlocks) String() string {
@@ -1406,7 +1344,7 @@ func (*MinedSoloBlocks) ProtoMessage() {}
 
 func (x *MinedSoloBlocks) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_payouts_proto_msgTypes[20]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1429,21 +1367,18 @@ func (x *MinedSoloBlocks) GetBlocks() []*MinedSoloBlock {
 }
 
 type MinedSoloBlocksList struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState         `protogen:"open.v1"`
+	Blocks        *MinedSoloBlocks               `protobuf:"bytes,1,opt,name=blocks,proto3" json:"blocks,omitempty"`
+	Pagination    *pagination.PaginationResponse `protobuf:"bytes,2,opt,name=pagination,proto3" json:"pagination,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Blocks     *MinedSoloBlocks               `protobuf:"bytes,1,opt,name=blocks,proto3" json:"blocks,omitempty"`
-	Pagination *pagination.PaginationResponse `protobuf:"bytes,2,opt,name=pagination,proto3" json:"pagination,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *MinedSoloBlocksList) Reset() {
 	*x = MinedSoloBlocksList{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_payouts_proto_msgTypes[21]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_payouts_proto_msgTypes[21]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *MinedSoloBlocksList) String() string {
@@ -1454,7 +1389,7 @@ func (*MinedSoloBlocksList) ProtoMessage() {}
 
 func (x *MinedSoloBlocksList) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_payouts_proto_msgTypes[21]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1484,20 +1419,17 @@ func (x *MinedSoloBlocksList) GetPagination() *pagination.PaginationResponse {
 }
 
 type MinedSoloBlocksMap struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState      `protogen:"open.v1"`
+	Blocks        map[string]*MinedSoloBlocks `protobuf:"bytes,1,rep,name=blocks,proto3" json:"blocks,omitempty" protobuf_key:"bytes,1,opt,name=key" protobuf_val:"bytes,2,opt,name=value"`
 	unknownFields protoimpl.UnknownFields
-
-	Blocks map[string]*MinedSoloBlocks `protobuf:"bytes,1,rep,name=blocks,proto3" json:"blocks,omitempty" protobuf_key:"bytes,1,opt,name=key,proto3" protobuf_val:"bytes,2,opt,name=value,proto3"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *MinedSoloBlocksMap) Reset() {
 	*x = MinedSoloBlocksMap{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_pool_payouts_proto_msgTypes[22]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_pool_payouts_proto_msgTypes[22]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *MinedSoloBlocksMap) String() string {
@@ -1508,7 +1440,7 @@ func (*MinedSoloBlocksMap) ProtoMessage() {}
 
 func (x *MinedSoloBlocksMap) ProtoReflect() protoreflect.Message {
 	mi := &file_pool_payouts_proto_msgTypes[22]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1532,381 +1464,200 @@ func (x *MinedSoloBlocksMap) GetBlocks() map[string]*MinedSoloBlocks {
 
 var File_pool_payouts_proto protoreflect.FileDescriptor
 
-var file_pool_payouts_proto_rawDesc = []byte{
-	0x0a, 0x12, 0x70, 0x6f, 0x6f, 0x6c, 0x5f, 0x70, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73, 0x2e, 0x70,
-	0x72, 0x6f, 0x74, 0x6f, 0x12, 0x0c, 0x70, 0x6f, 0x6f, 0x6c, 0x5f, 0x70, 0x61, 0x79, 0x6f, 0x75,
-	0x74, 0x73, 0x1a, 0x1f, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2f, 0x70, 0x72, 0x6f, 0x74, 0x6f,
-	0x62, 0x75, 0x66, 0x2f, 0x74, 0x69, 0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x2e, 0x70, 0x72,
-	0x6f, 0x74, 0x6f, 0x1a, 0x11, 0x70, 0x6f, 0x6f, 0x6c, 0x5f, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x73,
-	0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x1a, 0x16, 0x75, 0x74, 0x69, 0x6c, 0x73, 0x2f, 0x70, 0x61,
-	0x67, 0x69, 0x6e, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x1a, 0x11,
-	0x75, 0x74, 0x69, 0x6c, 0x73, 0x2f, 0x73, 0x6f, 0x72, 0x74, 0x73, 0x2e, 0x70, 0x72, 0x6f, 0x74,
-	0x6f, 0x1a, 0x13, 0x75, 0x74, 0x69, 0x6c, 0x73, 0x2f, 0x66, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x73,
-	0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x22, 0x28, 0x0a, 0x0c, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x42,
-	0x61, 0x6c, 0x61, 0x6e, 0x63, 0x65, 0x12, 0x18, 0x0a, 0x07, 0x62, 0x61, 0x6c, 0x61, 0x6e, 0x63,
-	0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x04, 0x52, 0x07, 0x62, 0x61, 0x6c, 0x61, 0x6e, 0x63, 0x65,
-	0x22, 0xb7, 0x01, 0x0a, 0x11, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x73, 0x42, 0x61, 0x6c, 0x61, 0x6e,
-	0x63, 0x65, 0x73, 0x4d, 0x61, 0x70, 0x12, 0x49, 0x0a, 0x08, 0x62, 0x61, 0x6c, 0x61, 0x6e, 0x63,
-	0x65, 0x73, 0x18, 0x01, 0x20, 0x03, 0x28, 0x0b, 0x32, 0x2d, 0x2e, 0x70, 0x6f, 0x6f, 0x6c, 0x5f,
-	0x70, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73, 0x2e, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x73, 0x42, 0x61,
-	0x6c, 0x61, 0x6e, 0x63, 0x65, 0x73, 0x4d, 0x61, 0x70, 0x2e, 0x42, 0x61, 0x6c, 0x61, 0x6e, 0x63,
-	0x65, 0x73, 0x45, 0x6e, 0x74, 0x72, 0x79, 0x52, 0x08, 0x62, 0x61, 0x6c, 0x61, 0x6e, 0x63, 0x65,
-	0x73, 0x1a, 0x57, 0x0a, 0x0d, 0x42, 0x61, 0x6c, 0x61, 0x6e, 0x63, 0x65, 0x73, 0x45, 0x6e, 0x74,
-	0x72, 0x79, 0x12, 0x10, 0x0a, 0x03, 0x6b, 0x65, 0x79, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52,
-	0x03, 0x6b, 0x65, 0x79, 0x12, 0x30, 0x0a, 0x05, 0x76, 0x61, 0x6c, 0x75, 0x65, 0x18, 0x02, 0x20,
-	0x01, 0x28, 0x0b, 0x32, 0x1a, 0x2e, 0x70, 0x6f, 0x6f, 0x6c, 0x5f, 0x70, 0x61, 0x79, 0x6f, 0x75,
-	0x74, 0x73, 0x2e, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x42, 0x61, 0x6c, 0x61, 0x6e, 0x63, 0x65, 0x52,
-	0x05, 0x76, 0x61, 0x6c, 0x75, 0x65, 0x3a, 0x02, 0x38, 0x01, 0x22, 0xeb, 0x01, 0x0a, 0x0e, 0x50,
-	0x61, 0x79, 0x6f, 0x75, 0x74, 0x73, 0x46, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x73, 0x12, 0x19, 0x0a,
-	0x05, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x48, 0x00, 0x52, 0x05,
-	0x6d, 0x69, 0x6e, 0x65, 0x72, 0x88, 0x01, 0x01, 0x12, 0x1c, 0x0a, 0x07, 0x74, 0x78, 0x5f, 0x68,
-	0x61, 0x73, 0x68, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x48, 0x01, 0x52, 0x06, 0x74, 0x78, 0x48,
-	0x61, 0x73, 0x68, 0x88, 0x01, 0x01, 0x12, 0x37, 0x0a, 0x06, 0x61, 0x6d, 0x6f, 0x75, 0x6e, 0x74,
-	0x18, 0x03, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1a, 0x2e, 0x66, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x73,
-	0x2e, 0x55, 0x49, 0x6e, 0x74, 0x36, 0x34, 0x52, 0x61, 0x6e, 0x67, 0x65, 0x46, 0x69, 0x6c, 0x74,
-	0x65, 0x72, 0x48, 0x02, 0x52, 0x06, 0x61, 0x6d, 0x6f, 0x75, 0x6e, 0x74, 0x88, 0x01, 0x01, 0x12,
-	0x3a, 0x0a, 0x07, 0x70, 0x61, 0x69, 0x64, 0x5f, 0x61, 0x74, 0x18, 0x04, 0x20, 0x01, 0x28, 0x0b,
-	0x32, 0x1c, 0x2e, 0x66, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x73, 0x2e, 0x44, 0x61, 0x74, 0x65, 0x54,
-	0x69, 0x6d, 0x65, 0x52, 0x61, 0x6e, 0x67, 0x65, 0x46, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x48, 0x03,
-	0x52, 0x06, 0x70, 0x61, 0x69, 0x64, 0x41, 0x74, 0x88, 0x01, 0x01, 0x42, 0x08, 0x0a, 0x06, 0x5f,
-	0x6d, 0x69, 0x6e, 0x65, 0x72, 0x42, 0x0a, 0x0a, 0x08, 0x5f, 0x74, 0x78, 0x5f, 0x68, 0x61, 0x73,
-	0x68, 0x42, 0x09, 0x0a, 0x07, 0x5f, 0x61, 0x6d, 0x6f, 0x75, 0x6e, 0x74, 0x42, 0x0a, 0x0a, 0x08,
-	0x5f, 0x70, 0x61, 0x69, 0x64, 0x5f, 0x61, 0x74, 0x22, 0xf7, 0x01, 0x0a, 0x0c, 0x50, 0x61, 0x79,
-	0x6f, 0x75, 0x74, 0x73, 0x53, 0x6f, 0x72, 0x74, 0x73, 0x12, 0x2b, 0x0a, 0x05, 0x6d, 0x69, 0x6e,
-	0x65, 0x72, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x10, 0x2e, 0x73, 0x6f, 0x72, 0x74, 0x73,
-	0x2e, 0x53, 0x6f, 0x72, 0x74, 0x4f, 0x72, 0x64, 0x65, 0x72, 0x48, 0x00, 0x52, 0x05, 0x6d, 0x69,
-	0x6e, 0x65, 0x72, 0x88, 0x01, 0x01, 0x12, 0x2e, 0x0a, 0x07, 0x74, 0x78, 0x5f, 0x68, 0x61, 0x73,
-	0x68, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x10, 0x2e, 0x73, 0x6f, 0x72, 0x74, 0x73, 0x2e,
-	0x53, 0x6f, 0x72, 0x74, 0x4f, 0x72, 0x64, 0x65, 0x72, 0x48, 0x01, 0x52, 0x06, 0x74, 0x78, 0x48,
-	0x61, 0x73, 0x68, 0x88, 0x01, 0x01, 0x12, 0x2d, 0x0a, 0x06, 0x61, 0x6d, 0x6f, 0x75, 0x6e, 0x74,
-	0x18, 0x03, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x10, 0x2e, 0x73, 0x6f, 0x72, 0x74, 0x73, 0x2e, 0x53,
-	0x6f, 0x72, 0x74, 0x4f, 0x72, 0x64, 0x65, 0x72, 0x48, 0x02, 0x52, 0x06, 0x61, 0x6d, 0x6f, 0x75,
-	0x6e, 0x74, 0x88, 0x01, 0x01, 0x12, 0x2e, 0x0a, 0x07, 0x70, 0x61, 0x69, 0x64, 0x5f, 0x61, 0x74,
-	0x18, 0x04, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x10, 0x2e, 0x73, 0x6f, 0x72, 0x74, 0x73, 0x2e, 0x53,
-	0x6f, 0x72, 0x74, 0x4f, 0x72, 0x64, 0x65, 0x72, 0x48, 0x03, 0x52, 0x06, 0x70, 0x61, 0x69, 0x64,
-	0x41, 0x74, 0x88, 0x01, 0x01, 0x42, 0x08, 0x0a, 0x06, 0x5f, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x42,
-	0x0a, 0x0a, 0x08, 0x5f, 0x74, 0x78, 0x5f, 0x68, 0x61, 0x73, 0x68, 0x42, 0x09, 0x0a, 0x07, 0x5f,
-	0x61, 0x6d, 0x6f, 0x75, 0x6e, 0x74, 0x42, 0x0a, 0x0a, 0x08, 0x5f, 0x70, 0x61, 0x69, 0x64, 0x5f,
-	0x61, 0x74, 0x22, 0xbc, 0x01, 0x0a, 0x11, 0x47, 0x65, 0x74, 0x50, 0x61, 0x79, 0x6f, 0x75, 0x74,
-	0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x30, 0x0a, 0x05, 0x73, 0x6f, 0x72, 0x74,
-	0x73, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1a, 0x2e, 0x70, 0x6f, 0x6f, 0x6c, 0x5f, 0x70,
-	0x61, 0x79, 0x6f, 0x75, 0x74, 0x73, 0x2e, 0x50, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73, 0x53, 0x6f,
-	0x72, 0x74, 0x73, 0x52, 0x05, 0x73, 0x6f, 0x72, 0x74, 0x73, 0x12, 0x36, 0x0a, 0x07, 0x66, 0x69,
-	0x6c, 0x74, 0x65, 0x72, 0x73, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x70, 0x6f,
-	0x6f, 0x6c, 0x5f, 0x70, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73, 0x2e, 0x50, 0x61, 0x79, 0x6f, 0x75,
-	0x74, 0x73, 0x46, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x73, 0x52, 0x07, 0x66, 0x69, 0x6c, 0x74, 0x65,
-	0x72, 0x73, 0x12, 0x3d, 0x0a, 0x0a, 0x70, 0x61, 0x67, 0x69, 0x6e, 0x61, 0x74, 0x69, 0x6f, 0x6e,
-	0x18, 0x03, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1d, 0x2e, 0x70, 0x61, 0x67, 0x69, 0x6e, 0x61, 0x74,
-	0x69, 0x6f, 0x6e, 0x2e, 0x50, 0x61, 0x67, 0x69, 0x6e, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x52, 0x65,
-	0x71, 0x75, 0x65, 0x73, 0x74, 0x52, 0x0a, 0x70, 0x61, 0x67, 0x69, 0x6e, 0x61, 0x74, 0x69, 0x6f,
-	0x6e, 0x22, 0x6b, 0x0a, 0x19, 0x47, 0x65, 0x74, 0x50, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73, 0x46,
-	0x72, 0x6f, 0x6d, 0x4c, 0x69, 0x73, 0x74, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x16,
-	0x0a, 0x06, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x73, 0x18, 0x01, 0x20, 0x03, 0x28, 0x09, 0x52, 0x06,
-	0x6d, 0x69, 0x6e, 0x65, 0x72, 0x73, 0x12, 0x36, 0x0a, 0x07, 0x66, 0x69, 0x6c, 0x74, 0x65, 0x72,
-	0x73, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x70, 0x6f, 0x6f, 0x6c, 0x5f, 0x70,
-	0x61, 0x79, 0x6f, 0x75, 0x74, 0x73, 0x2e, 0x50, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73, 0x46, 0x69,
-	0x6c, 0x74, 0x65, 0x72, 0x73, 0x52, 0x07, 0x66, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x73, 0x22, 0x84,
-	0x01, 0x0a, 0x06, 0x50, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x12, 0x14, 0x0a, 0x05, 0x6d, 0x69, 0x6e,
-	0x65, 0x72, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x05, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x12,
-	0x17, 0x0a, 0x07, 0x74, 0x78, 0x5f, 0x68, 0x61, 0x73, 0x68, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09,
-	0x52, 0x06, 0x74, 0x78, 0x48, 0x61, 0x73, 0x68, 0x12, 0x16, 0x0a, 0x06, 0x61, 0x6d, 0x6f, 0x75,
-	0x6e, 0x74, 0x18, 0x03, 0x20, 0x01, 0x28, 0x04, 0x52, 0x06, 0x61, 0x6d, 0x6f, 0x75, 0x6e, 0x74,
-	0x12, 0x33, 0x0a, 0x07, 0x70, 0x61, 0x69, 0x64, 0x5f, 0x61, 0x74, 0x18, 0x04, 0x20, 0x01, 0x28,
-	0x0b, 0x32, 0x1a, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f,
-	0x62, 0x75, 0x66, 0x2e, 0x54, 0x69, 0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x52, 0x06, 0x70,
-	0x61, 0x69, 0x64, 0x41, 0x74, 0x22, 0x39, 0x0a, 0x07, 0x50, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73,
-	0x12, 0x2e, 0x0a, 0x07, 0x70, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73, 0x18, 0x01, 0x20, 0x03, 0x28,
-	0x0b, 0x32, 0x14, 0x2e, 0x70, 0x6f, 0x6f, 0x6c, 0x5f, 0x70, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73,
-	0x2e, 0x50, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x52, 0x07, 0x70, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73,
-	0x22, 0x7e, 0x0a, 0x0b, 0x50, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73, 0x4c, 0x69, 0x73, 0x74, 0x12,
-	0x2f, 0x0a, 0x07, 0x70, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0b,
-	0x32, 0x15, 0x2e, 0x70, 0x6f, 0x6f, 0x6c, 0x5f, 0x70, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73, 0x2e,
-	0x50, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73, 0x52, 0x07, 0x70, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73,
-	0x12, 0x3e, 0x0a, 0x0a, 0x70, 0x61, 0x67, 0x69, 0x6e, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x18, 0x02,
-	0x20, 0x01, 0x28, 0x0b, 0x32, 0x1e, 0x2e, 0x70, 0x61, 0x67, 0x69, 0x6e, 0x61, 0x74, 0x69, 0x6f,
-	0x6e, 0x2e, 0x50, 0x61, 0x67, 0x69, 0x6e, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x52, 0x65, 0x73, 0x70,
-	0x6f, 0x6e, 0x73, 0x65, 0x52, 0x0a, 0x70, 0x61, 0x67, 0x69, 0x6e, 0x61, 0x74, 0x69, 0x6f, 0x6e,
-	0x22, 0xa0, 0x01, 0x0a, 0x0a, 0x50, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73, 0x4d, 0x61, 0x70, 0x12,
-	0x3f, 0x0a, 0x07, 0x70, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73, 0x18, 0x01, 0x20, 0x03, 0x28, 0x0b,
-	0x32, 0x25, 0x2e, 0x70, 0x6f, 0x6f, 0x6c, 0x5f, 0x70, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73, 0x2e,
-	0x50, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73, 0x4d, 0x61, 0x70, 0x2e, 0x50, 0x61, 0x79, 0x6f, 0x75,
-	0x74, 0x73, 0x45, 0x6e, 0x74, 0x72, 0x79, 0x52, 0x07, 0x70, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73,
-	0x1a, 0x51, 0x0a, 0x0c, 0x50, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73, 0x45, 0x6e, 0x74, 0x72, 0x79,
-	0x12, 0x10, 0x0a, 0x03, 0x6b, 0x65, 0x79, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x03, 0x6b,
-	0x65, 0x79, 0x12, 0x2b, 0x0a, 0x05, 0x76, 0x61, 0x6c, 0x75, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28,
-	0x0b, 0x32, 0x15, 0x2e, 0x70, 0x6f, 0x6f, 0x6c, 0x5f, 0x70, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73,
-	0x2e, 0x50, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73, 0x52, 0x05, 0x76, 0x61, 0x6c, 0x75, 0x65, 0x3a,
-	0x02, 0x38, 0x01, 0x22, 0xc5, 0x03, 0x0a, 0x10, 0x4d, 0x69, 0x6e, 0x65, 0x64, 0x42, 0x6c, 0x6f,
-	0x63, 0x6b, 0x73, 0x53, 0x6f, 0x72, 0x74, 0x73, 0x12, 0x2b, 0x0a, 0x05, 0x6d, 0x69, 0x6e, 0x65,
-	0x72, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x10, 0x2e, 0x73, 0x6f, 0x72, 0x74, 0x73, 0x2e,
-	0x53, 0x6f, 0x72, 0x74, 0x4f, 0x72, 0x64, 0x65, 0x72, 0x48, 0x00, 0x52, 0x05, 0x6d, 0x69, 0x6e,
-	0x65, 0x72, 0x88, 0x01, 0x01, 0x12, 0x3c, 0x0a, 0x0e, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x5f, 0x68,
-	0x61, 0x73, 0x68, 0x72, 0x61, 0x74, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x10, 0x2e,
-	0x73, 0x6f, 0x72, 0x74, 0x73, 0x2e, 0x53, 0x6f, 0x72, 0x74, 0x4f, 0x72, 0x64, 0x65, 0x72, 0x48,
-	0x01, 0x52, 0x0d, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x48, 0x61, 0x73, 0x68, 0x72, 0x61, 0x74, 0x65,
-	0x88, 0x01, 0x01, 0x12, 0x38, 0x0a, 0x0c, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x5f, 0x68, 0x65, 0x69,
-	0x67, 0x68, 0x74, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x10, 0x2e, 0x73, 0x6f, 0x72, 0x74,
-	0x73, 0x2e, 0x53, 0x6f, 0x72, 0x74, 0x4f, 0x72, 0x64, 0x65, 0x72, 0x48, 0x02, 0x52, 0x0b, 0x62,
-	0x6c, 0x6f, 0x63, 0x6b, 0x48, 0x65, 0x69, 0x67, 0x68, 0x74, 0x88, 0x01, 0x01, 0x12, 0x34, 0x0a,
-	0x0a, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x5f, 0x68, 0x61, 0x73, 0x68, 0x18, 0x04, 0x20, 0x01, 0x28,
-	0x0b, 0x32, 0x10, 0x2e, 0x73, 0x6f, 0x72, 0x74, 0x73, 0x2e, 0x53, 0x6f, 0x72, 0x74, 0x4f, 0x72,
-	0x64, 0x65, 0x72, 0x48, 0x03, 0x52, 0x09, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x48, 0x61, 0x73, 0x68,
-	0x88, 0x01, 0x01, 0x12, 0x43, 0x0a, 0x12, 0x72, 0x6f, 0x75, 0x6e, 0x64, 0x5f, 0x6d, 0x69, 0x6e,
-	0x65, 0x72, 0x73, 0x5f, 0x63, 0x6f, 0x75, 0x6e, 0x74, 0x18, 0x05, 0x20, 0x01, 0x28, 0x0b, 0x32,
-	0x10, 0x2e, 0x73, 0x6f, 0x72, 0x74, 0x73, 0x2e, 0x53, 0x6f, 0x72, 0x74, 0x4f, 0x72, 0x64, 0x65,
-	0x72, 0x48, 0x04, 0x52, 0x10, 0x72, 0x6f, 0x75, 0x6e, 0x64, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x73,
-	0x43, 0x6f, 0x75, 0x6e, 0x74, 0x88, 0x01, 0x01, 0x12, 0x30, 0x0a, 0x08, 0x6d, 0x69, 0x6e, 0x65,
-	0x64, 0x5f, 0x61, 0x74, 0x18, 0x06, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x10, 0x2e, 0x73, 0x6f, 0x72,
-	0x74, 0x73, 0x2e, 0x53, 0x6f, 0x72, 0x74, 0x4f, 0x72, 0x64, 0x65, 0x72, 0x48, 0x05, 0x52, 0x07,
-	0x6d, 0x69, 0x6e, 0x65, 0x64, 0x41, 0x74, 0x88, 0x01, 0x01, 0x42, 0x08, 0x0a, 0x06, 0x5f, 0x6d,
-	0x69, 0x6e, 0x65, 0x72, 0x42, 0x11, 0x0a, 0x0f, 0x5f, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x5f, 0x68,
-	0x61, 0x73, 0x68, 0x72, 0x61, 0x74, 0x65, 0x42, 0x0f, 0x0a, 0x0d, 0x5f, 0x62, 0x6c, 0x6f, 0x63,
-	0x6b, 0x5f, 0x68, 0x65, 0x69, 0x67, 0x68, 0x74, 0x42, 0x0d, 0x0a, 0x0b, 0x5f, 0x62, 0x6c, 0x6f,
-	0x63, 0x6b, 0x5f, 0x68, 0x61, 0x73, 0x68, 0x42, 0x15, 0x0a, 0x13, 0x5f, 0x72, 0x6f, 0x75, 0x6e,
-	0x64, 0x5f, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x73, 0x5f, 0x63, 0x6f, 0x75, 0x6e, 0x74, 0x42, 0x0b,
-	0x0a, 0x09, 0x5f, 0x6d, 0x69, 0x6e, 0x65, 0x64, 0x5f, 0x61, 0x74, 0x22, 0xc3, 0x03, 0x0a, 0x12,
-	0x4d, 0x69, 0x6e, 0x65, 0x64, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x46, 0x69, 0x6c, 0x74, 0x65,
-	0x72, 0x73, 0x12, 0x19, 0x0a, 0x05, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x18, 0x01, 0x20, 0x01, 0x28,
-	0x09, 0x48, 0x00, 0x52, 0x05, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x88, 0x01, 0x01, 0x12, 0x46, 0x0a,
-	0x0e, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x5f, 0x68, 0x61, 0x73, 0x68, 0x72, 0x61, 0x74, 0x65, 0x18,
-	0x02, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1a, 0x2e, 0x66, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x73, 0x2e,
-	0x42, 0x69, 0x67, 0x49, 0x6e, 0x74, 0x52, 0x61, 0x6e, 0x67, 0x65, 0x46, 0x69, 0x6c, 0x74, 0x65,
-	0x72, 0x48, 0x01, 0x52, 0x0d, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x48, 0x61, 0x73, 0x68, 0x72, 0x61,
-	0x74, 0x65, 0x88, 0x01, 0x01, 0x12, 0x3d, 0x0a, 0x0c, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x5f, 0x68,
-	0x65, 0x69, 0x67, 0x68, 0x74, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x15, 0x2e, 0x66, 0x69,
-	0x6c, 0x74, 0x65, 0x72, 0x73, 0x2e, 0x55, 0x49, 0x6e, 0x74, 0x36, 0x34, 0x46, 0x69, 0x6c, 0x74,
-	0x65, 0x72, 0x48, 0x02, 0x52, 0x0b, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x48, 0x65, 0x69, 0x67, 0x68,
-	0x74, 0x88, 0x01, 0x01, 0x12, 0x22, 0x0a, 0x0a, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x5f, 0x68, 0x61,
-	0x73, 0x68, 0x18, 0x04, 0x20, 0x01, 0x28, 0x09, 0x48, 0x03, 0x52, 0x09, 0x62, 0x6c, 0x6f, 0x63,
-	0x6b, 0x48, 0x61, 0x73, 0x68, 0x88, 0x01, 0x01, 0x12, 0x48, 0x0a, 0x12, 0x72, 0x6f, 0x75, 0x6e,
-	0x64, 0x5f, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x73, 0x5f, 0x63, 0x6f, 0x75, 0x6e, 0x74, 0x18, 0x05,
-	0x20, 0x01, 0x28, 0x0b, 0x32, 0x15, 0x2e, 0x66, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x73, 0x2e, 0x55,
-	0x49, 0x6e, 0x74, 0x33, 0x32, 0x46, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x48, 0x04, 0x52, 0x10, 0x72,
-	0x6f, 0x75, 0x6e, 0x64, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x73, 0x43, 0x6f, 0x75, 0x6e, 0x74, 0x88,
-	0x01, 0x01, 0x12, 0x3c, 0x0a, 0x08, 0x6d, 0x69, 0x6e, 0x65, 0x64, 0x5f, 0x61, 0x74, 0x18, 0x06,
-	0x20, 0x01, 0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x66, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x73, 0x2e, 0x44,
-	0x61, 0x74, 0x65, 0x54, 0x69, 0x6d, 0x65, 0x52, 0x61, 0x6e, 0x67, 0x65, 0x46, 0x69, 0x6c, 0x74,
-	0x65, 0x72, 0x48, 0x05, 0x52, 0x07, 0x6d, 0x69, 0x6e, 0x65, 0x64, 0x41, 0x74, 0x88, 0x01, 0x01,
-	0x42, 0x08, 0x0a, 0x06, 0x5f, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x42, 0x11, 0x0a, 0x0f, 0x5f, 0x6d,
-	0x69, 0x6e, 0x65, 0x72, 0x5f, 0x68, 0x61, 0x73, 0x68, 0x72, 0x61, 0x74, 0x65, 0x42, 0x0f, 0x0a,
-	0x0d, 0x5f, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x5f, 0x68, 0x65, 0x69, 0x67, 0x68, 0x74, 0x42, 0x0d,
-	0x0a, 0x0b, 0x5f, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x5f, 0x68, 0x61, 0x73, 0x68, 0x42, 0x15, 0x0a,
-	0x13, 0x5f, 0x72, 0x6f, 0x75, 0x6e, 0x64, 0x5f, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x73, 0x5f, 0x63,
-	0x6f, 0x75, 0x6e, 0x74, 0x42, 0x0b, 0x0a, 0x09, 0x5f, 0x6d, 0x69, 0x6e, 0x65, 0x64, 0x5f, 0x61,
-	0x74, 0x22, 0xc3, 0x01, 0x0a, 0x10, 0x47, 0x65, 0x74, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x52,
-	0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x34, 0x0a, 0x05, 0x73, 0x6f, 0x72, 0x74, 0x73, 0x18,
-	0x01, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1e, 0x2e, 0x70, 0x6f, 0x6f, 0x6c, 0x5f, 0x70, 0x61, 0x79,
-	0x6f, 0x75, 0x74, 0x73, 0x2e, 0x4d, 0x69, 0x6e, 0x65, 0x64, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x73,
-	0x53, 0x6f, 0x72, 0x74, 0x73, 0x52, 0x05, 0x73, 0x6f, 0x72, 0x74, 0x73, 0x12, 0x3a, 0x0a, 0x07,
-	0x66, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x73, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x20, 0x2e,
-	0x70, 0x6f, 0x6f, 0x6c, 0x5f, 0x70, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73, 0x2e, 0x4d, 0x69, 0x6e,
-	0x65, 0x64, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x46, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x73, 0x52,
-	0x07, 0x66, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x73, 0x12, 0x3d, 0x0a, 0x0a, 0x70, 0x61, 0x67, 0x69,
-	0x6e, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1d, 0x2e, 0x70,
-	0x61, 0x67, 0x69, 0x6e, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x2e, 0x50, 0x61, 0x67, 0x69, 0x6e, 0x61,
-	0x74, 0x69, 0x6f, 0x6e, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x52, 0x0a, 0x70, 0x61, 0x67,
-	0x69, 0x6e, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x22, 0xf0, 0x01, 0x0a, 0x0a, 0x4d, 0x69, 0x6e, 0x65,
-	0x64, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x12, 0x14, 0x0a, 0x05, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x18,
-	0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x05, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x12, 0x25, 0x0a, 0x0e,
-	0x6d, 0x69, 0x6e, 0x65, 0x72, 0x5f, 0x68, 0x61, 0x73, 0x68, 0x72, 0x61, 0x74, 0x65, 0x18, 0x02,
-	0x20, 0x01, 0x28, 0x0c, 0x52, 0x0d, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x48, 0x61, 0x73, 0x68, 0x72,
-	0x61, 0x74, 0x65, 0x12, 0x21, 0x0a, 0x0c, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x5f, 0x68, 0x65, 0x69,
-	0x67, 0x68, 0x74, 0x18, 0x03, 0x20, 0x01, 0x28, 0x04, 0x52, 0x0b, 0x62, 0x6c, 0x6f, 0x63, 0x6b,
-	0x48, 0x65, 0x69, 0x67, 0x68, 0x74, 0x12, 0x1d, 0x0a, 0x0a, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x5f,
-	0x68, 0x61, 0x73, 0x68, 0x18, 0x04, 0x20, 0x01, 0x28, 0x09, 0x52, 0x09, 0x62, 0x6c, 0x6f, 0x63,
-	0x6b, 0x48, 0x61, 0x73, 0x68, 0x12, 0x2c, 0x0a, 0x12, 0x72, 0x6f, 0x75, 0x6e, 0x64, 0x5f, 0x6d,
-	0x69, 0x6e, 0x65, 0x72, 0x73, 0x5f, 0x63, 0x6f, 0x75, 0x6e, 0x74, 0x18, 0x05, 0x20, 0x01, 0x28,
-	0x0d, 0x52, 0x10, 0x72, 0x6f, 0x75, 0x6e, 0x64, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x73, 0x43, 0x6f,
-	0x75, 0x6e, 0x74, 0x12, 0x35, 0x0a, 0x08, 0x6d, 0x69, 0x6e, 0x65, 0x64, 0x5f, 0x61, 0x74, 0x18,
-	0x06, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1a, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70,
-	0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x54, 0x69, 0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d,
-	0x70, 0x52, 0x07, 0x6d, 0x69, 0x6e, 0x65, 0x64, 0x41, 0x74, 0x22, 0x83, 0x01, 0x0a, 0x0f, 0x4d,
-	0x69, 0x6e, 0x65, 0x64, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x4c, 0x69, 0x73, 0x74, 0x12, 0x30,
-	0x0a, 0x06, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x18, 0x01, 0x20, 0x03, 0x28, 0x0b, 0x32, 0x18,
-	0x2e, 0x70, 0x6f, 0x6f, 0x6c, 0x5f, 0x70, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73, 0x2e, 0x4d, 0x69,
-	0x6e, 0x65, 0x64, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x52, 0x06, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x73,
-	0x12, 0x3e, 0x0a, 0x0a, 0x70, 0x61, 0x67, 0x69, 0x6e, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x18, 0x02,
-	0x20, 0x01, 0x28, 0x0b, 0x32, 0x1e, 0x2e, 0x70, 0x61, 0x67, 0x69, 0x6e, 0x61, 0x74, 0x69, 0x6f,
-	0x6e, 0x2e, 0x50, 0x61, 0x67, 0x69, 0x6e, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x52, 0x65, 0x73, 0x70,
-	0x6f, 0x6e, 0x73, 0x65, 0x52, 0x0a, 0x70, 0x61, 0x67, 0x69, 0x6e, 0x61, 0x74, 0x69, 0x6f, 0x6e,
-	0x22, 0xe3, 0x03, 0x0a, 0x14, 0x4d, 0x69, 0x6e, 0x65, 0x64, 0x53, 0x6f, 0x6c, 0x6f, 0x42, 0x6c,
-	0x6f, 0x63, 0x6b, 0x73, 0x53, 0x6f, 0x72, 0x74, 0x73, 0x12, 0x2b, 0x0a, 0x05, 0x6d, 0x69, 0x6e,
-	0x65, 0x72, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x10, 0x2e, 0x73, 0x6f, 0x72, 0x74, 0x73,
-	0x2e, 0x53, 0x6f, 0x72, 0x74, 0x4f, 0x72, 0x64, 0x65, 0x72, 0x48, 0x00, 0x52, 0x05, 0x6d, 0x69,
-	0x6e, 0x65, 0x72, 0x88, 0x01, 0x01, 0x12, 0x3c, 0x0a, 0x0e, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x5f,
-	0x68, 0x61, 0x73, 0x68, 0x72, 0x61, 0x74, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x10,
-	0x2e, 0x73, 0x6f, 0x72, 0x74, 0x73, 0x2e, 0x53, 0x6f, 0x72, 0x74, 0x4f, 0x72, 0x64, 0x65, 0x72,
-	0x48, 0x01, 0x52, 0x0d, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x48, 0x61, 0x73, 0x68, 0x72, 0x61, 0x74,
-	0x65, 0x88, 0x01, 0x01, 0x12, 0x38, 0x0a, 0x0c, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x5f, 0x68, 0x65,
-	0x69, 0x67, 0x68, 0x74, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x10, 0x2e, 0x73, 0x6f, 0x72,
-	0x74, 0x73, 0x2e, 0x53, 0x6f, 0x72, 0x74, 0x4f, 0x72, 0x64, 0x65, 0x72, 0x48, 0x02, 0x52, 0x0b,
-	0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x48, 0x65, 0x69, 0x67, 0x68, 0x74, 0x88, 0x01, 0x01, 0x12, 0x34,
-	0x0a, 0x0a, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x5f, 0x68, 0x61, 0x73, 0x68, 0x18, 0x04, 0x20, 0x01,
-	0x28, 0x0b, 0x32, 0x10, 0x2e, 0x73, 0x6f, 0x72, 0x74, 0x73, 0x2e, 0x53, 0x6f, 0x72, 0x74, 0x4f,
-	0x72, 0x64, 0x65, 0x72, 0x48, 0x03, 0x52, 0x09, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x48, 0x61, 0x73,
-	0x68, 0x88, 0x01, 0x01, 0x12, 0x2d, 0x0a, 0x06, 0x72, 0x65, 0x77, 0x61, 0x72, 0x64, 0x18, 0x05,
-	0x20, 0x01, 0x28, 0x0b, 0x32, 0x10, 0x2e, 0x73, 0x6f, 0x72, 0x74, 0x73, 0x2e, 0x53, 0x6f, 0x72,
-	0x74, 0x4f, 0x72, 0x64, 0x65, 0x72, 0x48, 0x04, 0x52, 0x06, 0x72, 0x65, 0x77, 0x61, 0x72, 0x64,
-	0x88, 0x01, 0x01, 0x12, 0x2e, 0x0a, 0x07, 0x74, 0x78, 0x5f, 0x68, 0x61, 0x73, 0x68, 0x18, 0x06,
-	0x20, 0x01, 0x28, 0x0b, 0x32, 0x10, 0x2e, 0x73, 0x6f, 0x72, 0x74, 0x73, 0x2e, 0x53, 0x6f, 0x72,
-	0x74, 0x4f, 0x72, 0x64, 0x65, 0x72, 0x48, 0x05, 0x52, 0x06, 0x74, 0x78, 0x48, 0x61, 0x73, 0x68,
-	0x88, 0x01, 0x01, 0x12, 0x30, 0x0a, 0x08, 0x6d, 0x69, 0x6e, 0x65, 0x64, 0x5f, 0x61, 0x74, 0x18,
-	0x07, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x10, 0x2e, 0x73, 0x6f, 0x72, 0x74, 0x73, 0x2e, 0x53, 0x6f,
-	0x72, 0x74, 0x4f, 0x72, 0x64, 0x65, 0x72, 0x48, 0x06, 0x52, 0x07, 0x6d, 0x69, 0x6e, 0x65, 0x64,
-	0x41, 0x74, 0x88, 0x01, 0x01, 0x42, 0x08, 0x0a, 0x06, 0x5f, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x42,
-	0x11, 0x0a, 0x0f, 0x5f, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x5f, 0x68, 0x61, 0x73, 0x68, 0x72, 0x61,
-	0x74, 0x65, 0x42, 0x0f, 0x0a, 0x0d, 0x5f, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x5f, 0x68, 0x65, 0x69,
-	0x67, 0x68, 0x74, 0x42, 0x0d, 0x0a, 0x0b, 0x5f, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x5f, 0x68, 0x61,
-	0x73, 0x68, 0x42, 0x09, 0x0a, 0x07, 0x5f, 0x72, 0x65, 0x77, 0x61, 0x72, 0x64, 0x42, 0x0a, 0x0a,
-	0x08, 0x5f, 0x74, 0x78, 0x5f, 0x68, 0x61, 0x73, 0x68, 0x42, 0x0b, 0x0a, 0x09, 0x5f, 0x6d, 0x69,
-	0x6e, 0x65, 0x64, 0x5f, 0x61, 0x74, 0x22, 0xd4, 0x03, 0x0a, 0x16, 0x4d, 0x69, 0x6e, 0x65, 0x64,
-	0x53, 0x6f, 0x6c, 0x6f, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x46, 0x69, 0x6c, 0x74, 0x65, 0x72,
-	0x73, 0x12, 0x19, 0x0a, 0x05, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09,
-	0x48, 0x00, 0x52, 0x05, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x88, 0x01, 0x01, 0x12, 0x46, 0x0a, 0x0e,
-	0x6d, 0x69, 0x6e, 0x65, 0x72, 0x5f, 0x68, 0x61, 0x73, 0x68, 0x72, 0x61, 0x74, 0x65, 0x18, 0x02,
-	0x20, 0x01, 0x28, 0x0b, 0x32, 0x1a, 0x2e, 0x66, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x73, 0x2e, 0x42,
-	0x69, 0x67, 0x49, 0x6e, 0x74, 0x52, 0x61, 0x6e, 0x67, 0x65, 0x46, 0x69, 0x6c, 0x74, 0x65, 0x72,
-	0x48, 0x01, 0x52, 0x0d, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x48, 0x61, 0x73, 0x68, 0x72, 0x61, 0x74,
-	0x65, 0x88, 0x01, 0x01, 0x12, 0x3d, 0x0a, 0x0c, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x5f, 0x68, 0x65,
-	0x69, 0x67, 0x68, 0x74, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x15, 0x2e, 0x66, 0x69, 0x6c,
-	0x74, 0x65, 0x72, 0x73, 0x2e, 0x55, 0x49, 0x6e, 0x74, 0x36, 0x34, 0x46, 0x69, 0x6c, 0x74, 0x65,
-	0x72, 0x48, 0x02, 0x52, 0x0b, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x48, 0x65, 0x69, 0x67, 0x68, 0x74,
-	0x88, 0x01, 0x01, 0x12, 0x22, 0x0a, 0x0a, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x5f, 0x68, 0x61, 0x73,
-	0x68, 0x18, 0x04, 0x20, 0x01, 0x28, 0x09, 0x48, 0x03, 0x52, 0x09, 0x62, 0x6c, 0x6f, 0x63, 0x6b,
-	0x48, 0x61, 0x73, 0x68, 0x88, 0x01, 0x01, 0x12, 0x37, 0x0a, 0x06, 0x72, 0x65, 0x77, 0x61, 0x72,
-	0x64, 0x18, 0x05, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1a, 0x2e, 0x66, 0x69, 0x6c, 0x74, 0x65, 0x72,
-	0x73, 0x2e, 0x55, 0x49, 0x6e, 0x74, 0x36, 0x34, 0x52, 0x61, 0x6e, 0x67, 0x65, 0x46, 0x69, 0x6c,
-	0x74, 0x65, 0x72, 0x48, 0x04, 0x52, 0x06, 0x72, 0x65, 0x77, 0x61, 0x72, 0x64, 0x88, 0x01, 0x01,
-	0x12, 0x1c, 0x0a, 0x07, 0x74, 0x78, 0x5f, 0x68, 0x61, 0x73, 0x68, 0x18, 0x06, 0x20, 0x01, 0x28,
-	0x09, 0x48, 0x05, 0x52, 0x06, 0x74, 0x78, 0x48, 0x61, 0x73, 0x68, 0x88, 0x01, 0x01, 0x12, 0x3c,
-	0x0a, 0x08, 0x6d, 0x69, 0x6e, 0x65, 0x64, 0x5f, 0x61, 0x74, 0x18, 0x07, 0x20, 0x01, 0x28, 0x0b,
-	0x32, 0x1c, 0x2e, 0x66, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x73, 0x2e, 0x44, 0x61, 0x74, 0x65, 0x54,
-	0x69, 0x6d, 0x65, 0x52, 0x61, 0x6e, 0x67, 0x65, 0x46, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x48, 0x06,
-	0x52, 0x07, 0x6d, 0x69, 0x6e, 0x65, 0x64, 0x41, 0x74, 0x88, 0x01, 0x01, 0x42, 0x08, 0x0a, 0x06,
-	0x5f, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x42, 0x11, 0x0a, 0x0f, 0x5f, 0x6d, 0x69, 0x6e, 0x65, 0x72,
-	0x5f, 0x68, 0x61, 0x73, 0x68, 0x72, 0x61, 0x74, 0x65, 0x42, 0x0f, 0x0a, 0x0d, 0x5f, 0x62, 0x6c,
-	0x6f, 0x63, 0x6b, 0x5f, 0x68, 0x65, 0x69, 0x67, 0x68, 0x74, 0x42, 0x0d, 0x0a, 0x0b, 0x5f, 0x62,
-	0x6c, 0x6f, 0x63, 0x6b, 0x5f, 0x68, 0x61, 0x73, 0x68, 0x42, 0x09, 0x0a, 0x07, 0x5f, 0x72, 0x65,
-	0x77, 0x61, 0x72, 0x64, 0x42, 0x0a, 0x0a, 0x08, 0x5f, 0x74, 0x78, 0x5f, 0x68, 0x61, 0x73, 0x68,
-	0x42, 0x0b, 0x0a, 0x09, 0x5f, 0x6d, 0x69, 0x6e, 0x65, 0x64, 0x5f, 0x61, 0x74, 0x22, 0xcf, 0x01,
-	0x0a, 0x14, 0x47, 0x65, 0x74, 0x53, 0x6f, 0x6c, 0x6f, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x52,
-	0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x38, 0x0a, 0x05, 0x73, 0x6f, 0x72, 0x74, 0x73, 0x18,
-	0x01, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x22, 0x2e, 0x70, 0x6f, 0x6f, 0x6c, 0x5f, 0x70, 0x61, 0x79,
-	0x6f, 0x75, 0x74, 0x73, 0x2e, 0x4d, 0x69, 0x6e, 0x65, 0x64, 0x53, 0x6f, 0x6c, 0x6f, 0x42, 0x6c,
-	0x6f, 0x63, 0x6b, 0x73, 0x53, 0x6f, 0x72, 0x74, 0x73, 0x52, 0x05, 0x73, 0x6f, 0x72, 0x74, 0x73,
-	0x12, 0x3e, 0x0a, 0x07, 0x66, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x73, 0x18, 0x02, 0x20, 0x01, 0x28,
-	0x0b, 0x32, 0x24, 0x2e, 0x70, 0x6f, 0x6f, 0x6c, 0x5f, 0x70, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73,
-	0x2e, 0x4d, 0x69, 0x6e, 0x65, 0x64, 0x53, 0x6f, 0x6c, 0x6f, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x73,
-	0x46, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x73, 0x52, 0x07, 0x66, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x73,
-	0x12, 0x3d, 0x0a, 0x0a, 0x70, 0x61, 0x67, 0x69, 0x6e, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x18, 0x03,
-	0x20, 0x01, 0x28, 0x0b, 0x32, 0x1d, 0x2e, 0x70, 0x61, 0x67, 0x69, 0x6e, 0x61, 0x74, 0x69, 0x6f,
-	0x6e, 0x2e, 0x50, 0x61, 0x67, 0x69, 0x6e, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x52, 0x65, 0x71, 0x75,
-	0x65, 0x73, 0x74, 0x52, 0x0a, 0x70, 0x61, 0x67, 0x69, 0x6e, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x22,
-	0x76, 0x0a, 0x1c, 0x47, 0x65, 0x74, 0x53, 0x6f, 0x6c, 0x6f, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x73,
-	0x46, 0x72, 0x6f, 0x6d, 0x4c, 0x69, 0x73, 0x74, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12,
-	0x16, 0x0a, 0x06, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x73, 0x18, 0x01, 0x20, 0x03, 0x28, 0x09, 0x52,
-	0x06, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x73, 0x12, 0x3e, 0x0a, 0x07, 0x66, 0x69, 0x6c, 0x74, 0x65,
-	0x72, 0x73, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x24, 0x2e, 0x70, 0x6f, 0x6f, 0x6c, 0x5f,
-	0x70, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73, 0x2e, 0x4d, 0x69, 0x6e, 0x65, 0x64, 0x53, 0x6f, 0x6c,
-	0x6f, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x46, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x73, 0x52, 0x07,
-	0x66, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x73, 0x22, 0xf7, 0x01, 0x0a, 0x0e, 0x4d, 0x69, 0x6e, 0x65,
-	0x64, 0x53, 0x6f, 0x6c, 0x6f, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x12, 0x14, 0x0a, 0x05, 0x6d, 0x69,
-	0x6e, 0x65, 0x72, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x05, 0x6d, 0x69, 0x6e, 0x65, 0x72,
-	0x12, 0x25, 0x0a, 0x0e, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x5f, 0x68, 0x61, 0x73, 0x68, 0x72, 0x61,
-	0x74, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x0d, 0x6d, 0x69, 0x6e, 0x65, 0x72, 0x48,
-	0x61, 0x73, 0x68, 0x72, 0x61, 0x74, 0x65, 0x12, 0x21, 0x0a, 0x0c, 0x62, 0x6c, 0x6f, 0x63, 0x6b,
-	0x5f, 0x68, 0x65, 0x69, 0x67, 0x68, 0x74, 0x18, 0x03, 0x20, 0x01, 0x28, 0x04, 0x52, 0x0b, 0x62,
-	0x6c, 0x6f, 0x63, 0x6b, 0x48, 0x65, 0x69, 0x67, 0x68, 0x74, 0x12, 0x1d, 0x0a, 0x0a, 0x62, 0x6c,
-	0x6f, 0x63, 0x6b, 0x5f, 0x68, 0x61, 0x73, 0x68, 0x18, 0x04, 0x20, 0x01, 0x28, 0x09, 0x52, 0x09,
-	0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x48, 0x61, 0x73, 0x68, 0x12, 0x16, 0x0a, 0x06, 0x72, 0x65, 0x77,
-	0x61, 0x72, 0x64, 0x18, 0x05, 0x20, 0x01, 0x28, 0x04, 0x52, 0x06, 0x72, 0x65, 0x77, 0x61, 0x72,
-	0x64, 0x12, 0x17, 0x0a, 0x07, 0x74, 0x78, 0x5f, 0x68, 0x61, 0x73, 0x68, 0x18, 0x06, 0x20, 0x01,
-	0x28, 0x09, 0x52, 0x06, 0x74, 0x78, 0x48, 0x61, 0x73, 0x68, 0x12, 0x35, 0x0a, 0x08, 0x6d, 0x69,
-	0x6e, 0x65, 0x64, 0x5f, 0x61, 0x74, 0x18, 0x07, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1a, 0x2e, 0x67,
-	0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x54,
-	0x69, 0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x52, 0x07, 0x6d, 0x69, 0x6e, 0x65, 0x64, 0x41,
-	0x74, 0x22, 0x47, 0x0a, 0x0f, 0x4d, 0x69, 0x6e, 0x65, 0x64, 0x53, 0x6f, 0x6c, 0x6f, 0x42, 0x6c,
-	0x6f, 0x63, 0x6b, 0x73, 0x12, 0x34, 0x0a, 0x06, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x18, 0x01,
-	0x20, 0x03, 0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x70, 0x6f, 0x6f, 0x6c, 0x5f, 0x70, 0x61, 0x79, 0x6f,
-	0x75, 0x74, 0x73, 0x2e, 0x4d, 0x69, 0x6e, 0x65, 0x64, 0x53, 0x6f, 0x6c, 0x6f, 0x42, 0x6c, 0x6f,
-	0x63, 0x6b, 0x52, 0x06, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x22, 0x8c, 0x01, 0x0a, 0x13, 0x4d,
-	0x69, 0x6e, 0x65, 0x64, 0x53, 0x6f, 0x6c, 0x6f, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x4c, 0x69,
-	0x73, 0x74, 0x12, 0x35, 0x0a, 0x06, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x18, 0x01, 0x20, 0x01,
-	0x28, 0x0b, 0x32, 0x1d, 0x2e, 0x70, 0x6f, 0x6f, 0x6c, 0x5f, 0x70, 0x61, 0x79, 0x6f, 0x75, 0x74,
-	0x73, 0x2e, 0x4d, 0x69, 0x6e, 0x65, 0x64, 0x53, 0x6f, 0x6c, 0x6f, 0x42, 0x6c, 0x6f, 0x63, 0x6b,
-	0x73, 0x52, 0x06, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x12, 0x3e, 0x0a, 0x0a, 0x70, 0x61, 0x67,
-	0x69, 0x6e, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1e, 0x2e,
-	0x70, 0x61, 0x67, 0x69, 0x6e, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x2e, 0x50, 0x61, 0x67, 0x69, 0x6e,
-	0x61, 0x74, 0x69, 0x6f, 0x6e, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x52, 0x0a, 0x70,
-	0x61, 0x67, 0x69, 0x6e, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x22, 0xb4, 0x01, 0x0a, 0x12, 0x4d, 0x69,
-	0x6e, 0x65, 0x64, 0x53, 0x6f, 0x6c, 0x6f, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x4d, 0x61, 0x70,
-	0x12, 0x44, 0x0a, 0x06, 0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x18, 0x01, 0x20, 0x03, 0x28, 0x0b,
-	0x32, 0x2c, 0x2e, 0x70, 0x6f, 0x6f, 0x6c, 0x5f, 0x70, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73, 0x2e,
-	0x4d, 0x69, 0x6e, 0x65, 0x64, 0x53, 0x6f, 0x6c, 0x6f, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x4d,
-	0x61, 0x70, 0x2e, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x45, 0x6e, 0x74, 0x72, 0x79, 0x52, 0x06,
-	0x62, 0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x1a, 0x58, 0x0a, 0x0b, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x73,
-	0x45, 0x6e, 0x74, 0x72, 0x79, 0x12, 0x10, 0x0a, 0x03, 0x6b, 0x65, 0x79, 0x18, 0x01, 0x20, 0x01,
-	0x28, 0x09, 0x52, 0x03, 0x6b, 0x65, 0x79, 0x12, 0x33, 0x0a, 0x05, 0x76, 0x61, 0x6c, 0x75, 0x65,
-	0x18, 0x02, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1d, 0x2e, 0x70, 0x6f, 0x6f, 0x6c, 0x5f, 0x70, 0x61,
-	0x79, 0x6f, 0x75, 0x74, 0x73, 0x2e, 0x4d, 0x69, 0x6e, 0x65, 0x64, 0x53, 0x6f, 0x6c, 0x6f, 0x42,
-	0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x52, 0x05, 0x76, 0x61, 0x6c, 0x75, 0x65, 0x3a, 0x02, 0x38, 0x01,
-	0x32, 0xa4, 0x04, 0x0a, 0x12, 0x50, 0x6f, 0x6f, 0x6c, 0x50, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73,
-	0x53, 0x65, 0x72, 0x76, 0x69, 0x63, 0x65, 0x12, 0x60, 0x0a, 0x19, 0x47, 0x65, 0x74, 0x4d, 0x69,
-	0x6e, 0x65, 0x72, 0x73, 0x42, 0x61, 0x6c, 0x61, 0x6e, 0x63, 0x65, 0x73, 0x46, 0x72, 0x6f, 0x6d,
-	0x4c, 0x69, 0x73, 0x74, 0x12, 0x22, 0x2e, 0x70, 0x6f, 0x6f, 0x6c, 0x5f, 0x6d, 0x69, 0x6e, 0x65,
-	0x72, 0x73, 0x2e, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x41, 0x64, 0x64, 0x72, 0x65, 0x73, 0x73, 0x65,
-	0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x1f, 0x2e, 0x70, 0x6f, 0x6f, 0x6c, 0x5f,
-	0x70, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73, 0x2e, 0x4d, 0x69, 0x6e, 0x65, 0x72, 0x73, 0x42, 0x61,
-	0x6c, 0x61, 0x6e, 0x63, 0x65, 0x73, 0x4d, 0x61, 0x70, 0x12, 0x48, 0x0a, 0x0a, 0x47, 0x65, 0x74,
-	0x50, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73, 0x12, 0x1f, 0x2e, 0x70, 0x6f, 0x6f, 0x6c, 0x5f, 0x70,
-	0x61, 0x79, 0x6f, 0x75, 0x74, 0x73, 0x2e, 0x47, 0x65, 0x74, 0x50, 0x61, 0x79, 0x6f, 0x75, 0x74,
-	0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x19, 0x2e, 0x70, 0x6f, 0x6f, 0x6c, 0x5f,
-	0x70, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73, 0x2e, 0x50, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73, 0x4c,
-	0x69, 0x73, 0x74, 0x12, 0x57, 0x0a, 0x12, 0x47, 0x65, 0x74, 0x50, 0x61, 0x79, 0x6f, 0x75, 0x74,
-	0x73, 0x46, 0x72, 0x6f, 0x6d, 0x4c, 0x69, 0x73, 0x74, 0x12, 0x27, 0x2e, 0x70, 0x6f, 0x6f, 0x6c,
-	0x5f, 0x70, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73, 0x2e, 0x47, 0x65, 0x74, 0x50, 0x61, 0x79, 0x6f,
-	0x75, 0x74, 0x73, 0x46, 0x72, 0x6f, 0x6d, 0x4c, 0x69, 0x73, 0x74, 0x52, 0x65, 0x71, 0x75, 0x65,
-	0x73, 0x74, 0x1a, 0x18, 0x2e, 0x70, 0x6f, 0x6f, 0x6c, 0x5f, 0x70, 0x61, 0x79, 0x6f, 0x75, 0x74,
-	0x73, 0x2e, 0x50, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73, 0x4d, 0x61, 0x70, 0x12, 0x4a, 0x0a, 0x09,
-	0x47, 0x65, 0x74, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x12, 0x1e, 0x2e, 0x70, 0x6f, 0x6f, 0x6c,
-	0x5f, 0x70, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73, 0x2e, 0x47, 0x65, 0x74, 0x42, 0x6c, 0x6f, 0x63,
-	0x6b, 0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x1d, 0x2e, 0x70, 0x6f, 0x6f, 0x6c,
-	0x5f, 0x70, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73, 0x2e, 0x4d, 0x69, 0x6e, 0x65, 0x64, 0x42, 0x6c,
-	0x6f, 0x63, 0x6b, 0x73, 0x4c, 0x69, 0x73, 0x74, 0x12, 0x56, 0x0a, 0x0d, 0x47, 0x65, 0x74, 0x53,
-	0x6f, 0x6c, 0x6f, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x12, 0x22, 0x2e, 0x70, 0x6f, 0x6f, 0x6c,
-	0x5f, 0x70, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73, 0x2e, 0x47, 0x65, 0x74, 0x53, 0x6f, 0x6c, 0x6f,
-	0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x21, 0x2e,
-	0x70, 0x6f, 0x6f, 0x6c, 0x5f, 0x70, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73, 0x2e, 0x4d, 0x69, 0x6e,
-	0x65, 0x64, 0x53, 0x6f, 0x6c, 0x6f, 0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x4c, 0x69, 0x73, 0x74,
-	0x12, 0x65, 0x0a, 0x15, 0x47, 0x65, 0x74, 0x53, 0x6f, 0x6c, 0x6f, 0x42, 0x6c, 0x6f, 0x63, 0x6b,
-	0x73, 0x46, 0x72, 0x6f, 0x6d, 0x4c, 0x69, 0x73, 0x74, 0x12, 0x2a, 0x2e, 0x70, 0x6f, 0x6f, 0x6c,
-	0x5f, 0x70, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73, 0x2e, 0x47, 0x65, 0x74, 0x53, 0x6f, 0x6c, 0x6f,
-	0x42, 0x6c, 0x6f, 0x63, 0x6b, 0x73, 0x46, 0x72, 0x6f, 0x6d, 0x4c, 0x69, 0x73, 0x74, 0x52, 0x65,
-	0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x20, 0x2e, 0x70, 0x6f, 0x6f, 0x6c, 0x5f, 0x70, 0x61, 0x79,
-	0x6f, 0x75, 0x74, 0x73, 0x2e, 0x4d, 0x69, 0x6e, 0x65, 0x64, 0x53, 0x6f, 0x6c, 0x6f, 0x42, 0x6c,
-	0x6f, 0x63, 0x6b, 0x73, 0x4d, 0x61, 0x70, 0x42, 0x53, 0x5a, 0x51, 0x67, 0x69, 0x74, 0x68, 0x75,
-	0x62, 0x2e, 0x63, 0x6f, 0x6d, 0x2f, 0x67, 0x72, 0x61, 0x6e, 0x64, 0x6d, 0x69, 0x6e, 0x69, 0x6e,
-	0x67, 0x70, 0x6f, 0x6f, 0x6c, 0x2f, 0x70, 0x6f, 0x6f, 0x6c, 0x2d, 0x61, 0x70, 0x69, 0x2d, 0x70,
-	0x72, 0x6f, 0x74, 0x6f, 0x2f, 0x67, 0x65, 0x6e, 0x65, 0x72, 0x61, 0x74, 0x65, 0x64, 0x2f, 0x70,
-	0x6f, 0x6f, 0x6c, 0x5f, 0x70, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73, 0x3b, 0x70, 0x6f, 0x6f, 0x6c,
-	0x50, 0x61, 0x79, 0x6f, 0x75, 0x74, 0x73, 0x50, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x06, 0x70, 0x72,
-	0x6f, 0x74, 0x6f, 0x33,
-}
+const file_pool_payouts_proto_rawDesc = "" +
+	"\n" +
+	"\x12pool_payouts.proto\x12\fpool_payouts\x1a\x1fgoogle/protobuf/timestamp.proto\x1a\x11pool_miners.proto\x1a\x16utils/pagination.proto\x1a\x11utils/sorts.proto\x1a\x13utils/filters.proto\"(\n" +
+	"\fMinerBalance\x12\x18\n" +
+	"\abalance\x18\x01 \x01(\x04R\abalance\"\xb7\x01\n" +
+	"\x11MinersBalancesMap\x12I\n" +
+	"\bbalances\x18\x01 \x03(\v2-.pool_payouts.MinersBalancesMap.BalancesEntryR\bbalances\x1aW\n" +
+	"\rBalancesEntry\x12\x10\n" +
+	"\x03key\x18\x01 \x01(\tR\x03key\x120\n" +
+	"\x05value\x18\x02 \x01(\v2\x1a.pool_payouts.MinerBalanceR\x05value:\x028\x01\"\xeb\x01\n" +
+	"\x0ePayoutsFilters\x12\x19\n" +
+	"\x05miner\x18\x01 \x01(\tH\x00R\x05miner\x88\x01\x01\x12\x1c\n" +
+	"\atx_hash\x18\x02 \x01(\tH\x01R\x06txHash\x88\x01\x01\x127\n" +
+	"\x06amount\x18\x03 \x01(\v2\x1a.filters.UInt64RangeFilterH\x02R\x06amount\x88\x01\x01\x12:\n" +
+	"\apaid_at\x18\x04 \x01(\v2\x1c.filters.DateTimeRangeFilterH\x03R\x06paidAt\x88\x01\x01B\b\n" +
+	"\x06_minerB\n" +
+	"\n" +
+	"\b_tx_hashB\t\n" +
+	"\a_amountB\n" +
+	"\n" +
+	"\b_paid_at\"\xf7\x01\n" +
+	"\fPayoutsSorts\x12+\n" +
+	"\x05miner\x18\x01 \x01(\v2\x10.sorts.SortOrderH\x00R\x05miner\x88\x01\x01\x12.\n" +
+	"\atx_hash\x18\x02 \x01(\v2\x10.sorts.SortOrderH\x01R\x06txHash\x88\x01\x01\x12-\n" +
+	"\x06amount\x18\x03 \x01(\v2\x10.sorts.SortOrderH\x02R\x06amount\x88\x01\x01\x12.\n" +
+	"\apaid_at\x18\x04 \x01(\v2\x10.sorts.SortOrderH\x03R\x06paidAt\x88\x01\x01B\b\n" +
+	"\x06_minerB\n" +
+	"\n" +
+	"\b_tx_hashB\t\n" +
+	"\a_amountB\n" +
+	"\n" +
+	"\b_paid_at\"\xbc\x01\n" +
+	"\x11GetPayoutsRequest\x120\n" +
+	"\x05sorts\x18\x01 \x01(\v2\x1a.pool_payouts.PayoutsSortsR\x05sorts\x126\n" +
+	"\afilters\x18\x02 \x01(\v2\x1c.pool_payouts.PayoutsFiltersR\afilters\x12=\n" +
+	"\n" +
+	"pagination\x18\x03 \x01(\v2\x1d.pagination.PaginationRequestR\n" +
+	"pagination\"k\n" +
+	"\x19GetPayoutsFromListRequest\x12\x16\n" +
+	"\x06miners\x18\x01 \x03(\tR\x06miners\x126\n" +
+	"\afilters\x18\x02 \x01(\v2\x1c.pool_payouts.PayoutsFiltersR\afilters\"\x84\x01\n" +
+	"\x06Payout\x12\x14\n" +
+	"\x05miner\x18\x01 \x01(\tR\x05miner\x12\x17\n" +
+	"\atx_hash\x18\x02 \x01(\tR\x06txHash\x12\x16\n" +
+	"\x06amount\x18\x03 \x01(\x04R\x06amount\x123\n" +
+	"\apaid_at\x18\x04 \x01(\v2\x1a.google.protobuf.TimestampR\x06paidAt\"9\n" +
+	"\aPayouts\x12.\n" +
+	"\apayouts\x18\x01 \x03(\v2\x14.pool_payouts.PayoutR\apayouts\"~\n" +
+	"\vPayoutsList\x12/\n" +
+	"\apayouts\x18\x01 \x01(\v2\x15.pool_payouts.PayoutsR\apayouts\x12>\n" +
+	"\n" +
+	"pagination\x18\x02 \x01(\v2\x1e.pagination.PaginationResponseR\n" +
+	"pagination\"\xa0\x01\n" +
+	"\n" +
+	"PayoutsMap\x12?\n" +
+	"\apayouts\x18\x01 \x03(\v2%.pool_payouts.PayoutsMap.PayoutsEntryR\apayouts\x1aQ\n" +
+	"\fPayoutsEntry\x12\x10\n" +
+	"\x03key\x18\x01 \x01(\tR\x03key\x12+\n" +
+	"\x05value\x18\x02 \x01(\v2\x15.pool_payouts.PayoutsR\x05value:\x028\x01\"\xc5\x03\n" +
+	"\x10MinedBlocksSorts\x12+\n" +
+	"\x05miner\x18\x01 \x01(\v2\x10.sorts.SortOrderH\x00R\x05miner\x88\x01\x01\x12<\n" +
+	"\x0eminer_hashrate\x18\x02 \x01(\v2\x10.sorts.SortOrderH\x01R\rminerHashrate\x88\x01\x01\x128\n" +
+	"\fblock_height\x18\x03 \x01(\v2\x10.sorts.SortOrderH\x02R\vblockHeight\x88\x01\x01\x124\n" +
+	"\n" +
+	"block_hash\x18\x04 \x01(\v2\x10.sorts.SortOrderH\x03R\tblockHash\x88\x01\x01\x12C\n" +
+	"\x12round_miners_count\x18\x05 \x01(\v2\x10.sorts.SortOrderH\x04R\x10roundMinersCount\x88\x01\x01\x120\n" +
+	"\bmined_at\x18\x06 \x01(\v2\x10.sorts.SortOrderH\x05R\aminedAt\x88\x01\x01B\b\n" +
+	"\x06_minerB\x11\n" +
+	"\x0f_miner_hashrateB\x0f\n" +
+	"\r_block_heightB\r\n" +
+	"\v_block_hashB\x15\n" +
+	"\x13_round_miners_countB\v\n" +
+	"\t_mined_at\"\xc3\x03\n" +
+	"\x12MinedBlocksFilters\x12\x19\n" +
+	"\x05miner\x18\x01 \x01(\tH\x00R\x05miner\x88\x01\x01\x12F\n" +
+	"\x0eminer_hashrate\x18\x02 \x01(\v2\x1a.filters.BigIntRangeFilterH\x01R\rminerHashrate\x88\x01\x01\x12=\n" +
+	"\fblock_height\x18\x03 \x01(\v2\x15.filters.UInt64FilterH\x02R\vblockHeight\x88\x01\x01\x12\"\n" +
+	"\n" +
+	"block_hash\x18\x04 \x01(\tH\x03R\tblockHash\x88\x01\x01\x12H\n" +
+	"\x12round_miners_count\x18\x05 \x01(\v2\x15.filters.UInt32FilterH\x04R\x10roundMinersCount\x88\x01\x01\x12<\n" +
+	"\bmined_at\x18\x06 \x01(\v2\x1c.filters.DateTimeRangeFilterH\x05R\aminedAt\x88\x01\x01B\b\n" +
+	"\x06_minerB\x11\n" +
+	"\x0f_miner_hashrateB\x0f\n" +
+	"\r_block_heightB\r\n" +
+	"\v_block_hashB\x15\n" +
+	"\x13_round_miners_countB\v\n" +
+	"\t_mined_at\"\xc3\x01\n" +
+	"\x10GetBlocksRequest\x124\n" +
+	"\x05sorts\x18\x01 \x01(\v2\x1e.pool_payouts.MinedBlocksSortsR\x05sorts\x12:\n" +
+	"\afilters\x18\x02 \x01(\v2 .pool_payouts.MinedBlocksFiltersR\afilters\x12=\n" +
+	"\n" +
+	"pagination\x18\x03 \x01(\v2\x1d.pagination.PaginationRequestR\n" +
+	"pagination\"\xf0\x01\n" +
+	"\n" +
+	"MinedBlock\x12\x14\n" +
+	"\x05miner\x18\x01 \x01(\tR\x05miner\x12%\n" +
+	"\x0eminer_hashrate\x18\x02 \x01(\fR\rminerHashrate\x12!\n" +
+	"\fblock_height\x18\x03 \x01(\x04R\vblockHeight\x12\x1d\n" +
+	"\n" +
+	"block_hash\x18\x04 \x01(\tR\tblockHash\x12,\n" +
+	"\x12round_miners_count\x18\x05 \x01(\rR\x10roundMinersCount\x125\n" +
+	"\bmined_at\x18\x06 \x01(\v2\x1a.google.protobuf.TimestampR\aminedAt\"\x83\x01\n" +
+	"\x0fMinedBlocksList\x120\n" +
+	"\x06blocks\x18\x01 \x03(\v2\x18.pool_payouts.MinedBlockR\x06blocks\x12>\n" +
+	"\n" +
+	"pagination\x18\x02 \x01(\v2\x1e.pagination.PaginationResponseR\n" +
+	"pagination\"\xe3\x03\n" +
+	"\x14MinedSoloBlocksSorts\x12+\n" +
+	"\x05miner\x18\x01 \x01(\v2\x10.sorts.SortOrderH\x00R\x05miner\x88\x01\x01\x12<\n" +
+	"\x0eminer_hashrate\x18\x02 \x01(\v2\x10.sorts.SortOrderH\x01R\rminerHashrate\x88\x01\x01\x128\n" +
+	"\fblock_height\x18\x03 \x01(\v2\x10.sorts.SortOrderH\x02R\vblockHeight\x88\x01\x01\x124\n" +
+	"\n" +
+	"block_hash\x18\x04 \x01(\v2\x10.sorts.SortOrderH\x03R\tblockHash\x88\x01\x01\x12-\n" +
+	"\x06reward\x18\x05 \x01(\v2\x10.sorts.SortOrderH\x04R\x06reward\x88\x01\x01\x12.\n" +
+	"\atx_hash\x18\x06 \x01(\v2\x10.sorts.SortOrderH\x05R\x06txHash\x88\x01\x01\x120\n" +
+	"\bmined_at\x18\a \x01(\v2\x10.sorts.SortOrderH\x06R\aminedAt\x88\x01\x01B\b\n" +
+	"\x06_minerB\x11\n" +
+	"\x0f_miner_hashrateB\x0f\n" +
+	"\r_block_heightB\r\n" +
+	"\v_block_hashB\t\n" +
+	"\a_rewardB\n" +
+	"\n" +
+	"\b_tx_hashB\v\n" +
+	"\t_mined_at\"\xd4\x03\n" +
+	"\x16MinedSoloBlocksFilters\x12\x19\n" +
+	"\x05miner\x18\x01 \x01(\tH\x00R\x05miner\x88\x01\x01\x12F\n" +
+	"\x0eminer_hashrate\x18\x02 \x01(\v2\x1a.filters.BigIntRangeFilterH\x01R\rminerHashrate\x88\x01\x01\x12=\n" +
+	"\fblock_height\x18\x03 \x01(\v2\x15.filters.UInt64FilterH\x02R\vblockHeight\x88\x01\x01\x12\"\n" +
+	"\n" +
+	"block_hash\x18\x04 \x01(\tH\x03R\tblockHash\x88\x01\x01\x127\n" +
+	"\x06reward\x18\x05 \x01(\v2\x1a.filters.UInt64RangeFilterH\x04R\x06reward\x88\x01\x01\x12\x1c\n" +
+	"\atx_hash\x18\x06 \x01(\tH\x05R\x06txHash\x88\x01\x01\x12<\n" +
+	"\bmined_at\x18\a \x01(\v2\x1c.filters.DateTimeRangeFilterH\x06R\aminedAt\x88\x01\x01B\b\n" +
+	"\x06_minerB\x11\n" +
+	"\x0f_miner_hashrateB\x0f\n" +
+	"\r_block_heightB\r\n" +
+	"\v_block_hashB\t\n" +
+	"\a_rewardB\n" +
+	"\n" +
+	"\b_tx_hashB\v\n" +
+	"\t_mined_at\"\xcf\x01\n" +
+	"\x14GetSoloBlocksRequest\x128\n" +
+	"\x05sorts\x18\x01 \x01(\v2\".pool_payouts.MinedSoloBlocksSortsR\x05sorts\x12>\n" +
+	"\afilters\x18\x02 \x01(\v2$.pool_payouts.MinedSoloBlocksFiltersR\afilters\x12=\n" +
+	"\n" +
+	"pagination\x18\x03 \x01(\v2\x1d.pagination.PaginationRequestR\n" +
+	"pagination\"v\n" +
+	"\x1cGetSoloBlocksFromListRequest\x12\x16\n" +
+	"\x06miners\x18\x01 \x03(\tR\x06miners\x12>\n" +
+	"\afilters\x18\x02 \x01(\v2$.pool_payouts.MinedSoloBlocksFiltersR\afilters\"\xf7\x01\n" +
+	"\x0eMinedSoloBlock\x12\x14\n" +
+	"\x05miner\x18\x01 \x01(\tR\x05miner\x12%\n" +
+	"\x0eminer_hashrate\x18\x02 \x01(\fR\rminerHashrate\x12!\n" +
+	"\fblock_height\x18\x03 \x01(\x04R\vblockHeight\x12\x1d\n" +
+	"\n" +
+	"block_hash\x18\x04 \x01(\tR\tblockHash\x12\x16\n" +
+	"\x06reward\x18\x05 \x01(\x04R\x06reward\x12\x17\n" +
+	"\atx_hash\x18\x06 \x01(\tR\x06txHash\x125\n" +
+	"\bmined_at\x18\a \x01(\v2\x1a.google.protobuf.TimestampR\aminedAt\"G\n" +
+	"\x0fMinedSoloBlocks\x124\n" +
+	"\x06blocks\x18\x01 \x03(\v2\x1c.pool_payouts.MinedSoloBlockR\x06blocks\"\x8c\x01\n" +
+	"\x13MinedSoloBlocksList\x125\n" +
+	"\x06blocks\x18\x01 \x01(\v2\x1d.pool_payouts.MinedSoloBlocksR\x06blocks\x12>\n" +
+	"\n" +
+	"pagination\x18\x02 \x01(\v2\x1e.pagination.PaginationResponseR\n" +
+	"pagination\"\xb4\x01\n" +
+	"\x12MinedSoloBlocksMap\x12D\n" +
+	"\x06blocks\x18\x01 \x03(\v2,.pool_payouts.MinedSoloBlocksMap.BlocksEntryR\x06blocks\x1aX\n" +
+	"\vBlocksEntry\x12\x10\n" +
+	"\x03key\x18\x01 \x01(\tR\x03key\x123\n" +
+	"\x05value\x18\x02 \x01(\v2\x1d.pool_payouts.MinedSoloBlocksR\x05value:\x028\x012\xa4\x04\n" +
+	"\x12PoolPayoutsService\x12`\n" +
+	"\x19GetMinersBalancesFromList\x12\".pool_miners.MinerAddressesRequest\x1a\x1f.pool_payouts.MinersBalancesMap\x12H\n" +
+	"\n" +
+	"GetPayouts\x12\x1f.pool_payouts.GetPayoutsRequest\x1a\x19.pool_payouts.PayoutsList\x12W\n" +
+	"\x12GetPayoutsFromList\x12'.pool_payouts.GetPayoutsFromListRequest\x1a\x18.pool_payouts.PayoutsMap\x12J\n" +
+	"\tGetBlocks\x12\x1e.pool_payouts.GetBlocksRequest\x1a\x1d.pool_payouts.MinedBlocksList\x12V\n" +
+	"\rGetSoloBlocks\x12\".pool_payouts.GetSoloBlocksRequest\x1a!.pool_payouts.MinedSoloBlocksList\x12e\n" +
+	"\x15GetSoloBlocksFromList\x12*.pool_payouts.GetSoloBlocksFromListRequest\x1a .pool_payouts.MinedSoloBlocksMapBOZMgithub.com/grandminingpool/pool-api-proto/generated/pool_payouts;pool_payoutsb\x06proto3"
 
 var (
 	file_pool_payouts_proto_rawDescOnce sync.Once
-	file_pool_payouts_proto_rawDescData = file_pool_payouts_proto_rawDesc
+	file_pool_payouts_proto_rawDescData []byte
 )
 
 func file_pool_payouts_proto_rawDescGZIP() []byte {
 	file_pool_payouts_proto_rawDescOnce.Do(func() {
-		file_pool_payouts_proto_rawDescData = protoimpl.X.CompressGZIP(file_pool_payouts_proto_rawDescData)
+		file_pool_payouts_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_pool_payouts_proto_rawDesc), len(file_pool_payouts_proto_rawDesc)))
 	})
 	return file_pool_payouts_proto_rawDescData
 }
 
 var file_pool_payouts_proto_msgTypes = make([]protoimpl.MessageInfo, 26)
-var file_pool_payouts_proto_goTypes = []interface{}{
+var file_pool_payouts_proto_goTypes = []any{
 	(*MinerBalance)(nil),                      // 0: pool_payouts.MinerBalance
 	(*MinersBalancesMap)(nil),                 // 1: pool_payouts.MinersBalancesMap
 	(*PayoutsFilters)(nil),                    // 2: pool_payouts.PayoutsFilters
@@ -2024,295 +1775,17 @@ func file_pool_payouts_proto_init() {
 	if File_pool_payouts_proto != nil {
 		return
 	}
-	if !protoimpl.UnsafeEnabled {
-		file_pool_payouts_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*MinerBalance); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_pool_payouts_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*MinersBalancesMap); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_pool_payouts_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*PayoutsFilters); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_pool_payouts_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*PayoutsSorts); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_pool_payouts_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*GetPayoutsRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_pool_payouts_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*GetPayoutsFromListRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_pool_payouts_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Payout); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_pool_payouts_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Payouts); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_pool_payouts_proto_msgTypes[8].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*PayoutsList); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_pool_payouts_proto_msgTypes[9].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*PayoutsMap); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_pool_payouts_proto_msgTypes[10].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*MinedBlocksSorts); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_pool_payouts_proto_msgTypes[11].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*MinedBlocksFilters); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_pool_payouts_proto_msgTypes[12].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*GetBlocksRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_pool_payouts_proto_msgTypes[13].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*MinedBlock); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_pool_payouts_proto_msgTypes[14].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*MinedBlocksList); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_pool_payouts_proto_msgTypes[15].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*MinedSoloBlocksSorts); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_pool_payouts_proto_msgTypes[16].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*MinedSoloBlocksFilters); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_pool_payouts_proto_msgTypes[17].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*GetSoloBlocksRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_pool_payouts_proto_msgTypes[18].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*GetSoloBlocksFromListRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_pool_payouts_proto_msgTypes[19].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*MinedSoloBlock); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_pool_payouts_proto_msgTypes[20].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*MinedSoloBlocks); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_pool_payouts_proto_msgTypes[21].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*MinedSoloBlocksList); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_pool_payouts_proto_msgTypes[22].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*MinedSoloBlocksMap); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-	}
-	file_pool_payouts_proto_msgTypes[2].OneofWrappers = []interface{}{}
-	file_pool_payouts_proto_msgTypes[3].OneofWrappers = []interface{}{}
-	file_pool_payouts_proto_msgTypes[10].OneofWrappers = []interface{}{}
-	file_pool_payouts_proto_msgTypes[11].OneofWrappers = []interface{}{}
-	file_pool_payouts_proto_msgTypes[15].OneofWrappers = []interface{}{}
-	file_pool_payouts_proto_msgTypes[16].OneofWrappers = []interface{}{}
+	file_pool_payouts_proto_msgTypes[2].OneofWrappers = []any{}
+	file_pool_payouts_proto_msgTypes[3].OneofWrappers = []any{}
+	file_pool_payouts_proto_msgTypes[10].OneofWrappers = []any{}
+	file_pool_payouts_proto_msgTypes[11].OneofWrappers = []any{}
+	file_pool_payouts_proto_msgTypes[15].OneofWrappers = []any{}
+	file_pool_payouts_proto_msgTypes[16].OneofWrappers = []any{}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
-			RawDescriptor: file_pool_payouts_proto_rawDesc,
+			RawDescriptor: unsafe.Slice(unsafe.StringData(file_pool_payouts_proto_rawDesc), len(file_pool_payouts_proto_rawDesc)),
 			NumEnums:      0,
 			NumMessages:   26,
 			NumExtensions: 0,
@@ -2323,7 +1796,6 @@ func file_pool_payouts_proto_init() {
 		MessageInfos:      file_pool_payouts_proto_msgTypes,
 	}.Build()
 	File_pool_payouts_proto = out.File
-	file_pool_payouts_proto_rawDesc = nil
 	file_pool_payouts_proto_goTypes = nil
 	file_pool_payouts_proto_depIdxs = nil
 }
