@@ -1,10 +1,10 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.28.1
-// 	protoc        v5.26.1
+// 	protoc-gen-go v1.36.6
+// 	protoc        v5.29.3
 // source: filters.proto
 
-package filtersProto
+package filters
 
 import (
 	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
@@ -12,6 +12,7 @@ import (
 	timestamppb "google.golang.org/protobuf/types/known/timestamppb"
 	reflect "reflect"
 	sync "sync"
+	unsafe "unsafe"
 )
 
 const (
@@ -22,21 +23,18 @@ const (
 )
 
 type UInt32Range struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Min           *uint32                `protobuf:"varint,2,opt,name=min,proto3,oneof" json:"min,omitempty"`
+	Max           *uint32                `protobuf:"varint,3,opt,name=max,proto3,oneof" json:"max,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Min *uint32 `protobuf:"varint,2,opt,name=min,proto3,oneof" json:"min,omitempty"`
-	Max *uint32 `protobuf:"varint,3,opt,name=max,proto3,oneof" json:"max,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *UInt32Range) Reset() {
 	*x = UInt32Range{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_filters_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_filters_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *UInt32Range) String() string {
@@ -47,7 +45,7 @@ func (*UInt32Range) ProtoMessage() {}
 
 func (x *UInt32Range) ProtoReflect() protoreflect.Message {
 	mi := &file_filters_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -77,24 +75,21 @@ func (x *UInt32Range) GetMax() uint32 {
 }
 
 type UInt32Filter struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	// Types that are assignable to Operators:
+	state protoimpl.MessageState `protogen:"open.v1"`
+	// Types that are valid to be assigned to Operators:
 	//
 	//	*UInt32Filter_Equals
 	//	*UInt32Filter_Range
-	Operators isUInt32Filter_Operators `protobuf_oneof:"operators"`
+	Operators     isUInt32Filter_Operators `protobuf_oneof:"operators"`
+	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *UInt32Filter) Reset() {
 	*x = UInt32Filter{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_filters_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_filters_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *UInt32Filter) String() string {
@@ -105,7 +100,7 @@ func (*UInt32Filter) ProtoMessage() {}
 
 func (x *UInt32Filter) ProtoReflect() protoreflect.Message {
 	mi := &file_filters_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -120,23 +115,27 @@ func (*UInt32Filter) Descriptor() ([]byte, []int) {
 	return file_filters_proto_rawDescGZIP(), []int{1}
 }
 
-func (m *UInt32Filter) GetOperators() isUInt32Filter_Operators {
-	if m != nil {
-		return m.Operators
+func (x *UInt32Filter) GetOperators() isUInt32Filter_Operators {
+	if x != nil {
+		return x.Operators
 	}
 	return nil
 }
 
 func (x *UInt32Filter) GetEquals() uint32 {
-	if x, ok := x.GetOperators().(*UInt32Filter_Equals); ok {
-		return x.Equals
+	if x != nil {
+		if x, ok := x.Operators.(*UInt32Filter_Equals); ok {
+			return x.Equals
+		}
 	}
 	return 0
 }
 
 func (x *UInt32Filter) GetRange() *UInt32Range {
-	if x, ok := x.GetOperators().(*UInt32Filter_Range); ok {
-		return x.Range
+	if x != nil {
+		if x, ok := x.Operators.(*UInt32Filter_Range); ok {
+			return x.Range
+		}
 	}
 	return nil
 }
@@ -158,21 +157,18 @@ func (*UInt32Filter_Equals) isUInt32Filter_Operators() {}
 func (*UInt32Filter_Range) isUInt32Filter_Operators() {}
 
 type UInt64RangeFilter struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Min           *uint64                `protobuf:"varint,1,opt,name=min,proto3,oneof" json:"min,omitempty"`
+	Max           *uint64                `protobuf:"varint,2,opt,name=max,proto3,oneof" json:"max,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Min *uint64 `protobuf:"varint,1,opt,name=min,proto3,oneof" json:"min,omitempty"`
-	Max *uint64 `protobuf:"varint,2,opt,name=max,proto3,oneof" json:"max,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *UInt64RangeFilter) Reset() {
 	*x = UInt64RangeFilter{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_filters_proto_msgTypes[2]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_filters_proto_msgTypes[2]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *UInt64RangeFilter) String() string {
@@ -183,7 +179,7 @@ func (*UInt64RangeFilter) ProtoMessage() {}
 
 func (x *UInt64RangeFilter) ProtoReflect() protoreflect.Message {
 	mi := &file_filters_proto_msgTypes[2]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -213,24 +209,21 @@ func (x *UInt64RangeFilter) GetMax() uint64 {
 }
 
 type UInt64Filter struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	// Types that are assignable to Operators:
+	state protoimpl.MessageState `protogen:"open.v1"`
+	// Types that are valid to be assigned to Operators:
 	//
 	//	*UInt64Filter_Equals
 	//	*UInt64Filter_Range
-	Operators isUInt64Filter_Operators `protobuf_oneof:"operators"`
+	Operators     isUInt64Filter_Operators `protobuf_oneof:"operators"`
+	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *UInt64Filter) Reset() {
 	*x = UInt64Filter{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_filters_proto_msgTypes[3]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_filters_proto_msgTypes[3]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *UInt64Filter) String() string {
@@ -241,7 +234,7 @@ func (*UInt64Filter) ProtoMessage() {}
 
 func (x *UInt64Filter) ProtoReflect() protoreflect.Message {
 	mi := &file_filters_proto_msgTypes[3]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -256,23 +249,27 @@ func (*UInt64Filter) Descriptor() ([]byte, []int) {
 	return file_filters_proto_rawDescGZIP(), []int{3}
 }
 
-func (m *UInt64Filter) GetOperators() isUInt64Filter_Operators {
-	if m != nil {
-		return m.Operators
+func (x *UInt64Filter) GetOperators() isUInt64Filter_Operators {
+	if x != nil {
+		return x.Operators
 	}
 	return nil
 }
 
 func (x *UInt64Filter) GetEquals() uint64 {
-	if x, ok := x.GetOperators().(*UInt64Filter_Equals); ok {
-		return x.Equals
+	if x != nil {
+		if x, ok := x.Operators.(*UInt64Filter_Equals); ok {
+			return x.Equals
+		}
 	}
 	return 0
 }
 
 func (x *UInt64Filter) GetRange() *UInt64RangeFilter {
-	if x, ok := x.GetOperators().(*UInt64Filter_Range); ok {
-		return x.Range
+	if x != nil {
+		if x, ok := x.Operators.(*UInt64Filter_Range); ok {
+			return x.Range
+		}
 	}
 	return nil
 }
@@ -294,21 +291,18 @@ func (*UInt64Filter_Equals) isUInt64Filter_Operators() {}
 func (*UInt64Filter_Range) isUInt64Filter_Operators() {}
 
 type BigIntRangeFilter struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Min           []byte                 `protobuf:"bytes,1,opt,name=min,proto3,oneof" json:"min,omitempty"`
+	Max           []byte                 `protobuf:"bytes,2,opt,name=max,proto3,oneof" json:"max,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Min []byte `protobuf:"bytes,1,opt,name=min,proto3,oneof" json:"min,omitempty"`
-	Max []byte `protobuf:"bytes,2,opt,name=max,proto3,oneof" json:"max,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *BigIntRangeFilter) Reset() {
 	*x = BigIntRangeFilter{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_filters_proto_msgTypes[4]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_filters_proto_msgTypes[4]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *BigIntRangeFilter) String() string {
@@ -319,7 +313,7 @@ func (*BigIntRangeFilter) ProtoMessage() {}
 
 func (x *BigIntRangeFilter) ProtoReflect() protoreflect.Message {
 	mi := &file_filters_proto_msgTypes[4]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -349,21 +343,18 @@ func (x *BigIntRangeFilter) GetMax() []byte {
 }
 
 type DateTimeRangeFilter struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Start         *timestamppb.Timestamp `protobuf:"bytes,2,opt,name=start,proto3,oneof" json:"start,omitempty"`
+	End           *timestamppb.Timestamp `protobuf:"bytes,3,opt,name=end,proto3,oneof" json:"end,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Start *timestamppb.Timestamp `protobuf:"bytes,2,opt,name=start,proto3,oneof" json:"start,omitempty"`
-	End   *timestamppb.Timestamp `protobuf:"bytes,3,opt,name=end,proto3,oneof" json:"end,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *DateTimeRangeFilter) Reset() {
 	*x = DateTimeRangeFilter{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_filters_proto_msgTypes[5]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_filters_proto_msgTypes[5]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *DateTimeRangeFilter) String() string {
@@ -374,7 +365,7 @@ func (*DateTimeRangeFilter) ProtoMessage() {}
 
 func (x *DateTimeRangeFilter) ProtoReflect() protoreflect.Message {
 	mi := &file_filters_proto_msgTypes[5]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -405,70 +396,52 @@ func (x *DateTimeRangeFilter) GetEnd() *timestamppb.Timestamp {
 
 var File_filters_proto protoreflect.FileDescriptor
 
-var file_filters_proto_rawDesc = []byte{
-	0x0a, 0x0d, 0x66, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x73, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12,
-	0x07, 0x66, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x73, 0x1a, 0x1f, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65,
-	0x2f, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2f, 0x74, 0x69, 0x6d, 0x65, 0x73, 0x74,
-	0x61, 0x6d, 0x70, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x22, 0x4b, 0x0a, 0x0b, 0x55, 0x49, 0x6e,
-	0x74, 0x33, 0x32, 0x52, 0x61, 0x6e, 0x67, 0x65, 0x12, 0x15, 0x0a, 0x03, 0x6d, 0x69, 0x6e, 0x18,
-	0x02, 0x20, 0x01, 0x28, 0x0d, 0x48, 0x00, 0x52, 0x03, 0x6d, 0x69, 0x6e, 0x88, 0x01, 0x01, 0x12,
-	0x15, 0x0a, 0x03, 0x6d, 0x61, 0x78, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0d, 0x48, 0x01, 0x52, 0x03,
-	0x6d, 0x61, 0x78, 0x88, 0x01, 0x01, 0x42, 0x06, 0x0a, 0x04, 0x5f, 0x6d, 0x69, 0x6e, 0x42, 0x06,
-	0x0a, 0x04, 0x5f, 0x6d, 0x61, 0x78, 0x22, 0x63, 0x0a, 0x0c, 0x55, 0x49, 0x6e, 0x74, 0x33, 0x32,
-	0x46, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x12, 0x18, 0x0a, 0x06, 0x65, 0x71, 0x75, 0x61, 0x6c, 0x73,
-	0x18, 0x01, 0x20, 0x01, 0x28, 0x0d, 0x48, 0x00, 0x52, 0x06, 0x65, 0x71, 0x75, 0x61, 0x6c, 0x73,
-	0x12, 0x2c, 0x0a, 0x05, 0x72, 0x61, 0x6e, 0x67, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0b, 0x32,
-	0x14, 0x2e, 0x66, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x73, 0x2e, 0x55, 0x49, 0x6e, 0x74, 0x33, 0x32,
-	0x52, 0x61, 0x6e, 0x67, 0x65, 0x48, 0x00, 0x52, 0x05, 0x72, 0x61, 0x6e, 0x67, 0x65, 0x42, 0x0b,
-	0x0a, 0x09, 0x6f, 0x70, 0x65, 0x72, 0x61, 0x74, 0x6f, 0x72, 0x73, 0x22, 0x51, 0x0a, 0x11, 0x55,
-	0x49, 0x6e, 0x74, 0x36, 0x34, 0x52, 0x61, 0x6e, 0x67, 0x65, 0x46, 0x69, 0x6c, 0x74, 0x65, 0x72,
-	0x12, 0x15, 0x0a, 0x03, 0x6d, 0x69, 0x6e, 0x18, 0x01, 0x20, 0x01, 0x28, 0x04, 0x48, 0x00, 0x52,
-	0x03, 0x6d, 0x69, 0x6e, 0x88, 0x01, 0x01, 0x12, 0x15, 0x0a, 0x03, 0x6d, 0x61, 0x78, 0x18, 0x02,
-	0x20, 0x01, 0x28, 0x04, 0x48, 0x01, 0x52, 0x03, 0x6d, 0x61, 0x78, 0x88, 0x01, 0x01, 0x42, 0x06,
-	0x0a, 0x04, 0x5f, 0x6d, 0x69, 0x6e, 0x42, 0x06, 0x0a, 0x04, 0x5f, 0x6d, 0x61, 0x78, 0x22, 0x69,
-	0x0a, 0x0c, 0x55, 0x49, 0x6e, 0x74, 0x36, 0x34, 0x46, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x12, 0x18,
-	0x0a, 0x06, 0x65, 0x71, 0x75, 0x61, 0x6c, 0x73, 0x18, 0x01, 0x20, 0x01, 0x28, 0x04, 0x48, 0x00,
-	0x52, 0x06, 0x65, 0x71, 0x75, 0x61, 0x6c, 0x73, 0x12, 0x32, 0x0a, 0x05, 0x72, 0x61, 0x6e, 0x67,
-	0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1a, 0x2e, 0x66, 0x69, 0x6c, 0x74, 0x65, 0x72,
-	0x73, 0x2e, 0x55, 0x49, 0x6e, 0x74, 0x36, 0x34, 0x52, 0x61, 0x6e, 0x67, 0x65, 0x46, 0x69, 0x6c,
-	0x74, 0x65, 0x72, 0x48, 0x00, 0x52, 0x05, 0x72, 0x61, 0x6e, 0x67, 0x65, 0x42, 0x0b, 0x0a, 0x09,
-	0x6f, 0x70, 0x65, 0x72, 0x61, 0x74, 0x6f, 0x72, 0x73, 0x22, 0x51, 0x0a, 0x11, 0x42, 0x69, 0x67,
-	0x49, 0x6e, 0x74, 0x52, 0x61, 0x6e, 0x67, 0x65, 0x46, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x12, 0x15,
-	0x0a, 0x03, 0x6d, 0x69, 0x6e, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0c, 0x48, 0x00, 0x52, 0x03, 0x6d,
-	0x69, 0x6e, 0x88, 0x01, 0x01, 0x12, 0x15, 0x0a, 0x03, 0x6d, 0x61, 0x78, 0x18, 0x02, 0x20, 0x01,
-	0x28, 0x0c, 0x48, 0x01, 0x52, 0x03, 0x6d, 0x61, 0x78, 0x88, 0x01, 0x01, 0x42, 0x06, 0x0a, 0x04,
-	0x5f, 0x6d, 0x69, 0x6e, 0x42, 0x06, 0x0a, 0x04, 0x5f, 0x6d, 0x61, 0x78, 0x22, 0x91, 0x01, 0x0a,
-	0x13, 0x44, 0x61, 0x74, 0x65, 0x54, 0x69, 0x6d, 0x65, 0x52, 0x61, 0x6e, 0x67, 0x65, 0x46, 0x69,
-	0x6c, 0x74, 0x65, 0x72, 0x12, 0x35, 0x0a, 0x05, 0x73, 0x74, 0x61, 0x72, 0x74, 0x18, 0x02, 0x20,
-	0x01, 0x28, 0x0b, 0x32, 0x1a, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f,
-	0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x54, 0x69, 0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x48,
-	0x00, 0x52, 0x05, 0x73, 0x74, 0x61, 0x72, 0x74, 0x88, 0x01, 0x01, 0x12, 0x31, 0x0a, 0x03, 0x65,
-	0x6e, 0x64, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1a, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c,
-	0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x54, 0x69, 0x6d, 0x65, 0x73,
-	0x74, 0x61, 0x6d, 0x70, 0x48, 0x01, 0x52, 0x03, 0x65, 0x6e, 0x64, 0x88, 0x01, 0x01, 0x42, 0x08,
-	0x0a, 0x06, 0x5f, 0x73, 0x74, 0x61, 0x72, 0x74, 0x42, 0x06, 0x0a, 0x04, 0x5f, 0x65, 0x6e, 0x64,
-	0x42, 0x50, 0x5a, 0x4e, 0x67, 0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d, 0x2f, 0x67,
-	0x72, 0x61, 0x6e, 0x64, 0x6d, 0x69, 0x6e, 0x69, 0x6e, 0x67, 0x70, 0x6f, 0x6f, 0x6c, 0x2f, 0x70,
-	0x6f, 0x6f, 0x6c, 0x2d, 0x61, 0x70, 0x69, 0x2d, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2f, 0x67, 0x65,
-	0x6e, 0x65, 0x72, 0x61, 0x74, 0x65, 0x64, 0x2f, 0x75, 0x74, 0x69, 0x6c, 0x73, 0x2f, 0x66, 0x69,
-	0x6c, 0x74, 0x65, 0x72, 0x73, 0x3b, 0x66, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x73, 0x50, 0x72, 0x6f,
-	0x74, 0x6f, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
-}
+const file_filters_proto_rawDesc = "" +
+	"\n" +
+	"\rfilters.proto\x12\afilters\x1a\x1fgoogle/protobuf/timestamp.proto\"K\n" +
+	"\vUInt32Range\x12\x15\n" +
+	"\x03min\x18\x02 \x01(\rH\x00R\x03min\x88\x01\x01\x12\x15\n" +
+	"\x03max\x18\x03 \x01(\rH\x01R\x03max\x88\x01\x01B\x06\n" +
+	"\x04_minB\x06\n" +
+	"\x04_max\"c\n" +
+	"\fUInt32Filter\x12\x18\n" +
+	"\x06equals\x18\x01 \x01(\rH\x00R\x06equals\x12,\n" +
+	"\x05range\x18\x02 \x01(\v2\x14.filters.UInt32RangeH\x00R\x05rangeB\v\n" +
+	"\toperators\"Q\n" +
+	"\x11UInt64RangeFilter\x12\x15\n" +
+	"\x03min\x18\x01 \x01(\x04H\x00R\x03min\x88\x01\x01\x12\x15\n" +
+	"\x03max\x18\x02 \x01(\x04H\x01R\x03max\x88\x01\x01B\x06\n" +
+	"\x04_minB\x06\n" +
+	"\x04_max\"i\n" +
+	"\fUInt64Filter\x12\x18\n" +
+	"\x06equals\x18\x01 \x01(\x04H\x00R\x06equals\x122\n" +
+	"\x05range\x18\x02 \x01(\v2\x1a.filters.UInt64RangeFilterH\x00R\x05rangeB\v\n" +
+	"\toperators\"Q\n" +
+	"\x11BigIntRangeFilter\x12\x15\n" +
+	"\x03min\x18\x01 \x01(\fH\x00R\x03min\x88\x01\x01\x12\x15\n" +
+	"\x03max\x18\x02 \x01(\fH\x01R\x03max\x88\x01\x01B\x06\n" +
+	"\x04_minB\x06\n" +
+	"\x04_max\"\x91\x01\n" +
+	"\x13DateTimeRangeFilter\x125\n" +
+	"\x05start\x18\x02 \x01(\v2\x1a.google.protobuf.TimestampH\x00R\x05start\x88\x01\x01\x121\n" +
+	"\x03end\x18\x03 \x01(\v2\x1a.google.protobuf.TimestampH\x01R\x03end\x88\x01\x01B\b\n" +
+	"\x06_startB\x06\n" +
+	"\x04_endBKZIgithub.com/grandminingpool/pool-api-proto/generated/utils/filters;filtersb\x06proto3"
 
 var (
 	file_filters_proto_rawDescOnce sync.Once
-	file_filters_proto_rawDescData = file_filters_proto_rawDesc
+	file_filters_proto_rawDescData []byte
 )
 
 func file_filters_proto_rawDescGZIP() []byte {
 	file_filters_proto_rawDescOnce.Do(func() {
-		file_filters_proto_rawDescData = protoimpl.X.CompressGZIP(file_filters_proto_rawDescData)
+		file_filters_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_filters_proto_rawDesc), len(file_filters_proto_rawDesc)))
 	})
 	return file_filters_proto_rawDescData
 }
 
 var file_filters_proto_msgTypes = make([]protoimpl.MessageInfo, 6)
-var file_filters_proto_goTypes = []interface{}{
+var file_filters_proto_goTypes = []any{
 	(*UInt32Range)(nil),           // 0: filters.UInt32Range
 	(*UInt32Filter)(nil),          // 1: filters.UInt32Filter
 	(*UInt64RangeFilter)(nil),     // 2: filters.UInt64RangeFilter
@@ -494,97 +467,23 @@ func file_filters_proto_init() {
 	if File_filters_proto != nil {
 		return
 	}
-	if !protoimpl.UnsafeEnabled {
-		file_filters_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*UInt32Range); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_filters_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*UInt32Filter); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_filters_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*UInt64RangeFilter); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_filters_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*UInt64Filter); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_filters_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*BigIntRangeFilter); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_filters_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*DateTimeRangeFilter); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-	}
-	file_filters_proto_msgTypes[0].OneofWrappers = []interface{}{}
-	file_filters_proto_msgTypes[1].OneofWrappers = []interface{}{
+	file_filters_proto_msgTypes[0].OneofWrappers = []any{}
+	file_filters_proto_msgTypes[1].OneofWrappers = []any{
 		(*UInt32Filter_Equals)(nil),
 		(*UInt32Filter_Range)(nil),
 	}
-	file_filters_proto_msgTypes[2].OneofWrappers = []interface{}{}
-	file_filters_proto_msgTypes[3].OneofWrappers = []interface{}{
+	file_filters_proto_msgTypes[2].OneofWrappers = []any{}
+	file_filters_proto_msgTypes[3].OneofWrappers = []any{
 		(*UInt64Filter_Equals)(nil),
 		(*UInt64Filter_Range)(nil),
 	}
-	file_filters_proto_msgTypes[4].OneofWrappers = []interface{}{}
-	file_filters_proto_msgTypes[5].OneofWrappers = []interface{}{}
+	file_filters_proto_msgTypes[4].OneofWrappers = []any{}
+	file_filters_proto_msgTypes[5].OneofWrappers = []any{}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
-			RawDescriptor: file_filters_proto_rawDesc,
+			RawDescriptor: unsafe.Slice(unsafe.StringData(file_filters_proto_rawDesc), len(file_filters_proto_rawDesc)),
 			NumEnums:      0,
 			NumMessages:   6,
 			NumExtensions: 0,
@@ -595,7 +494,6 @@ func file_filters_proto_init() {
 		MessageInfos:      file_filters_proto_msgTypes,
 	}.Build()
 	File_filters_proto = out.File
-	file_filters_proto_rawDesc = nil
 	file_filters_proto_goTypes = nil
 	file_filters_proto_depIdxs = nil
 }
