@@ -1,16 +1,17 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.28.1
-// 	protoc        v5.26.1
+// 	protoc-gen-go v1.36.6
+// 	protoc        v5.29.3
 // source: sorts.proto
 
-package sortsProto
+package sorts
 
 import (
 	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	reflect "reflect"
 	sync "sync"
+	unsafe "unsafe"
 )
 
 const (
@@ -67,20 +68,17 @@ func (SortDirection) EnumDescriptor() ([]byte, []int) {
 }
 
 type SortOrder struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Direction     SortDirection          `protobuf:"varint,1,opt,name=direction,proto3,enum=sorts.SortDirection" json:"direction,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Direction SortDirection `protobuf:"varint,1,opt,name=direction,proto3,enum=sorts.SortDirection" json:"direction,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *SortOrder) Reset() {
 	*x = SortOrder{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_sorts_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_sorts_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *SortOrder) String() string {
@@ -91,7 +89,7 @@ func (*SortOrder) ProtoMessage() {}
 
 func (x *SortOrder) ProtoReflect() protoreflect.Message {
 	mi := &file_sorts_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -115,37 +113,30 @@ func (x *SortOrder) GetDirection() SortDirection {
 
 var File_sorts_proto protoreflect.FileDescriptor
 
-var file_sorts_proto_rawDesc = []byte{
-	0x0a, 0x0b, 0x73, 0x6f, 0x72, 0x74, 0x73, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12, 0x05, 0x73,
-	0x6f, 0x72, 0x74, 0x73, 0x22, 0x3f, 0x0a, 0x09, 0x53, 0x6f, 0x72, 0x74, 0x4f, 0x72, 0x64, 0x65,
-	0x72, 0x12, 0x32, 0x0a, 0x09, 0x64, 0x69, 0x72, 0x65, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x18, 0x01,
-	0x20, 0x01, 0x28, 0x0e, 0x32, 0x14, 0x2e, 0x73, 0x6f, 0x72, 0x74, 0x73, 0x2e, 0x53, 0x6f, 0x72,
-	0x74, 0x44, 0x69, 0x72, 0x65, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x52, 0x09, 0x64, 0x69, 0x72, 0x65,
-	0x63, 0x74, 0x69, 0x6f, 0x6e, 0x2a, 0x22, 0x0a, 0x0d, 0x53, 0x6f, 0x72, 0x74, 0x44, 0x69, 0x72,
-	0x65, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x12, 0x07, 0x0a, 0x03, 0x41, 0x53, 0x43, 0x10, 0x00, 0x12,
-	0x08, 0x0a, 0x04, 0x44, 0x45, 0x53, 0x43, 0x10, 0x01, 0x42, 0x4c, 0x5a, 0x4a, 0x67, 0x69, 0x74,
-	0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d, 0x2f, 0x67, 0x72, 0x61, 0x6e, 0x64, 0x6d, 0x69, 0x6e,
-	0x69, 0x6e, 0x67, 0x70, 0x6f, 0x6f, 0x6c, 0x2f, 0x70, 0x6f, 0x6f, 0x6c, 0x2d, 0x61, 0x70, 0x69,
-	0x2d, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2f, 0x67, 0x65, 0x6e, 0x65, 0x72, 0x61, 0x74, 0x65, 0x64,
-	0x2f, 0x75, 0x74, 0x69, 0x6c, 0x73, 0x2f, 0x73, 0x6f, 0x72, 0x74, 0x73, 0x3b, 0x73, 0x6f, 0x72,
-	0x74, 0x73, 0x50, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
-}
+const file_sorts_proto_rawDesc = "" +
+	"\n" +
+	"\vsorts.proto\x12\x05sorts\"?\n" +
+	"\tSortOrder\x122\n" +
+	"\tdirection\x18\x01 \x01(\x0e2\x14.sorts.SortDirectionR\tdirection*\"\n" +
+	"\rSortDirection\x12\a\n" +
+	"\x03ASC\x10\x00\x12\b\n" +
+	"\x04DESC\x10\x01BGZEgithub.com/grandminingpool/pool-api-proto/generated/utils/sorts;sortsb\x06proto3"
 
 var (
 	file_sorts_proto_rawDescOnce sync.Once
-	file_sorts_proto_rawDescData = file_sorts_proto_rawDesc
+	file_sorts_proto_rawDescData []byte
 )
 
 func file_sorts_proto_rawDescGZIP() []byte {
 	file_sorts_proto_rawDescOnce.Do(func() {
-		file_sorts_proto_rawDescData = protoimpl.X.CompressGZIP(file_sorts_proto_rawDescData)
+		file_sorts_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_sorts_proto_rawDesc), len(file_sorts_proto_rawDesc)))
 	})
 	return file_sorts_proto_rawDescData
 }
 
 var file_sorts_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
 var file_sorts_proto_msgTypes = make([]protoimpl.MessageInfo, 1)
-var file_sorts_proto_goTypes = []interface{}{
+var file_sorts_proto_goTypes = []any{
 	(SortDirection)(0), // 0: sorts.SortDirection
 	(*SortOrder)(nil),  // 1: sorts.SortOrder
 }
@@ -163,25 +154,11 @@ func file_sorts_proto_init() {
 	if File_sorts_proto != nil {
 		return
 	}
-	if !protoimpl.UnsafeEnabled {
-		file_sorts_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*SortOrder); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
-			RawDescriptor: file_sorts_proto_rawDesc,
+			RawDescriptor: unsafe.Slice(unsafe.StringData(file_sorts_proto_rawDesc), len(file_sorts_proto_rawDesc)),
 			NumEnums:      1,
 			NumMessages:   1,
 			NumExtensions: 0,
@@ -193,7 +170,6 @@ func file_sorts_proto_init() {
 		MessageInfos:      file_sorts_proto_msgTypes,
 	}.Build()
 	File_sorts_proto = out.File
-	file_sorts_proto_rawDesc = nil
 	file_sorts_proto_goTypes = nil
 	file_sorts_proto_depIdxs = nil
 }
