@@ -1,10 +1,10 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.2.0
-// - protoc             v5.26.1
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.29.3
 // source: charts.proto
 
-package chartsProto
+package charts
 
 import (
 	context "context"
@@ -15,8 +15,19 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
+
+const (
+	ChartsService_GetPoolStats_FullMethodName            = "/charts.ChartsService/GetPoolStats"
+	ChartsService_GetNetworkDifficulties_FullMethodName  = "/charts.ChartsService/GetNetworkDifficulties"
+	ChartsService_GetRounds_FullMethodName               = "/charts.ChartsService/GetRounds"
+	ChartsService_GetMinerProfitabilities_FullMethodName = "/charts.ChartsService/GetMinerProfitabilities"
+	ChartsService_GetMinerHashrates_FullMethodName       = "/charts.ChartsService/GetMinerHashrates"
+	ChartsService_GetMinerWorkerHashrates_FullMethodName = "/charts.ChartsService/GetMinerWorkerHashrates"
+	ChartsService_GetMinerShares_FullMethodName          = "/charts.ChartsService/GetMinerShares"
+	ChartsService_GetMinerWorkerShares_FullMethodName    = "/charts.ChartsService/GetMinerWorkerShares"
+)
 
 // ChartsServiceClient is the client API for ChartsService service.
 //
@@ -41,8 +52,9 @@ func NewChartsServiceClient(cc grpc.ClientConnInterface) ChartsServiceClient {
 }
 
 func (c *chartsServiceClient) GetPoolStats(ctx context.Context, in *GetPoolStatsRequest, opts ...grpc.CallOption) (*PoolStatsPoints, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PoolStatsPoints)
-	err := c.cc.Invoke(ctx, "/charts.ChartsService/GetPoolStats", in, out, opts...)
+	err := c.cc.Invoke(ctx, ChartsService_GetPoolStats_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -50,8 +62,9 @@ func (c *chartsServiceClient) GetPoolStats(ctx context.Context, in *GetPoolStats
 }
 
 func (c *chartsServiceClient) GetNetworkDifficulties(ctx context.Context, in *GetNetworkDifficultiesRequest, opts ...grpc.CallOption) (*NetworkDifficultiesPoints, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(NetworkDifficultiesPoints)
-	err := c.cc.Invoke(ctx, "/charts.ChartsService/GetNetworkDifficulties", in, out, opts...)
+	err := c.cc.Invoke(ctx, ChartsService_GetNetworkDifficulties_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -59,8 +72,9 @@ func (c *chartsServiceClient) GetNetworkDifficulties(ctx context.Context, in *Ge
 }
 
 func (c *chartsServiceClient) GetRounds(ctx context.Context, in *GetRoundsRequest, opts ...grpc.CallOption) (*RoundsPoints, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(RoundsPoints)
-	err := c.cc.Invoke(ctx, "/charts.ChartsService/GetRounds", in, out, opts...)
+	err := c.cc.Invoke(ctx, ChartsService_GetRounds_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -68,8 +82,9 @@ func (c *chartsServiceClient) GetRounds(ctx context.Context, in *GetRoundsReques
 }
 
 func (c *chartsServiceClient) GetMinerProfitabilities(ctx context.Context, in *GetMinerChartRequest, opts ...grpc.CallOption) (*MinerProfitabilitiesPoints, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(MinerProfitabilitiesPoints)
-	err := c.cc.Invoke(ctx, "/charts.ChartsService/GetMinerProfitabilities", in, out, opts...)
+	err := c.cc.Invoke(ctx, ChartsService_GetMinerProfitabilities_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -77,8 +92,9 @@ func (c *chartsServiceClient) GetMinerProfitabilities(ctx context.Context, in *G
 }
 
 func (c *chartsServiceClient) GetMinerHashrates(ctx context.Context, in *GetMinerChartRequest, opts ...grpc.CallOption) (*MinerHashratesPoints, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(MinerHashratesPoints)
-	err := c.cc.Invoke(ctx, "/charts.ChartsService/GetMinerHashrates", in, out, opts...)
+	err := c.cc.Invoke(ctx, ChartsService_GetMinerHashrates_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -86,8 +102,9 @@ func (c *chartsServiceClient) GetMinerHashrates(ctx context.Context, in *GetMine
 }
 
 func (c *chartsServiceClient) GetMinerWorkerHashrates(ctx context.Context, in *GetMinerWorkerChartRequest, opts ...grpc.CallOption) (*MinerHashratesPoints, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(MinerHashratesPoints)
-	err := c.cc.Invoke(ctx, "/charts.ChartsService/GetMinerWorkerHashrates", in, out, opts...)
+	err := c.cc.Invoke(ctx, ChartsService_GetMinerWorkerHashrates_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -95,8 +112,9 @@ func (c *chartsServiceClient) GetMinerWorkerHashrates(ctx context.Context, in *G
 }
 
 func (c *chartsServiceClient) GetMinerShares(ctx context.Context, in *GetMinerChartRequest, opts ...grpc.CallOption) (*MinerSharesPoints, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(MinerSharesPoints)
-	err := c.cc.Invoke(ctx, "/charts.ChartsService/GetMinerShares", in, out, opts...)
+	err := c.cc.Invoke(ctx, ChartsService_GetMinerShares_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -104,8 +122,9 @@ func (c *chartsServiceClient) GetMinerShares(ctx context.Context, in *GetMinerCh
 }
 
 func (c *chartsServiceClient) GetMinerWorkerShares(ctx context.Context, in *GetMinerWorkerChartRequest, opts ...grpc.CallOption) (*MinerSharesPoints, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(MinerSharesPoints)
-	err := c.cc.Invoke(ctx, "/charts.ChartsService/GetMinerWorkerShares", in, out, opts...)
+	err := c.cc.Invoke(ctx, ChartsService_GetMinerWorkerShares_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -114,7 +133,7 @@ func (c *chartsServiceClient) GetMinerWorkerShares(ctx context.Context, in *GetM
 
 // ChartsServiceServer is the server API for ChartsService service.
 // All implementations must embed UnimplementedChartsServiceServer
-// for forward compatibility
+// for forward compatibility.
 type ChartsServiceServer interface {
 	GetPoolStats(context.Context, *GetPoolStatsRequest) (*PoolStatsPoints, error)
 	GetNetworkDifficulties(context.Context, *GetNetworkDifficultiesRequest) (*NetworkDifficultiesPoints, error)
@@ -127,9 +146,12 @@ type ChartsServiceServer interface {
 	mustEmbedUnimplementedChartsServiceServer()
 }
 
-// UnimplementedChartsServiceServer must be embedded to have forward compatible implementations.
-type UnimplementedChartsServiceServer struct {
-}
+// UnimplementedChartsServiceServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedChartsServiceServer struct{}
 
 func (UnimplementedChartsServiceServer) GetPoolStats(context.Context, *GetPoolStatsRequest) (*PoolStatsPoints, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetPoolStats not implemented")
@@ -156,6 +178,7 @@ func (UnimplementedChartsServiceServer) GetMinerWorkerShares(context.Context, *G
 	return nil, status.Errorf(codes.Unimplemented, "method GetMinerWorkerShares not implemented")
 }
 func (UnimplementedChartsServiceServer) mustEmbedUnimplementedChartsServiceServer() {}
+func (UnimplementedChartsServiceServer) testEmbeddedByValue()                       {}
 
 // UnsafeChartsServiceServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to ChartsServiceServer will
@@ -165,6 +188,13 @@ type UnsafeChartsServiceServer interface {
 }
 
 func RegisterChartsServiceServer(s grpc.ServiceRegistrar, srv ChartsServiceServer) {
+	// If the following call pancis, it indicates UnimplementedChartsServiceServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&ChartsService_ServiceDesc, srv)
 }
 
@@ -178,7 +208,7 @@ func _ChartsService_GetPoolStats_Handler(srv interface{}, ctx context.Context, d
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/charts.ChartsService/GetPoolStats",
+		FullMethod: ChartsService_GetPoolStats_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(ChartsServiceServer).GetPoolStats(ctx, req.(*GetPoolStatsRequest))
@@ -196,7 +226,7 @@ func _ChartsService_GetNetworkDifficulties_Handler(srv interface{}, ctx context.
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/charts.ChartsService/GetNetworkDifficulties",
+		FullMethod: ChartsService_GetNetworkDifficulties_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(ChartsServiceServer).GetNetworkDifficulties(ctx, req.(*GetNetworkDifficultiesRequest))
@@ -214,7 +244,7 @@ func _ChartsService_GetRounds_Handler(srv interface{}, ctx context.Context, dec 
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/charts.ChartsService/GetRounds",
+		FullMethod: ChartsService_GetRounds_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(ChartsServiceServer).GetRounds(ctx, req.(*GetRoundsRequest))
@@ -232,7 +262,7 @@ func _ChartsService_GetMinerProfitabilities_Handler(srv interface{}, ctx context
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/charts.ChartsService/GetMinerProfitabilities",
+		FullMethod: ChartsService_GetMinerProfitabilities_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(ChartsServiceServer).GetMinerProfitabilities(ctx, req.(*GetMinerChartRequest))
@@ -250,7 +280,7 @@ func _ChartsService_GetMinerHashrates_Handler(srv interface{}, ctx context.Conte
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/charts.ChartsService/GetMinerHashrates",
+		FullMethod: ChartsService_GetMinerHashrates_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(ChartsServiceServer).GetMinerHashrates(ctx, req.(*GetMinerChartRequest))
@@ -268,7 +298,7 @@ func _ChartsService_GetMinerWorkerHashrates_Handler(srv interface{}, ctx context
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/charts.ChartsService/GetMinerWorkerHashrates",
+		FullMethod: ChartsService_GetMinerWorkerHashrates_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(ChartsServiceServer).GetMinerWorkerHashrates(ctx, req.(*GetMinerWorkerChartRequest))
@@ -286,7 +316,7 @@ func _ChartsService_GetMinerShares_Handler(srv interface{}, ctx context.Context,
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/charts.ChartsService/GetMinerShares",
+		FullMethod: ChartsService_GetMinerShares_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(ChartsServiceServer).GetMinerShares(ctx, req.(*GetMinerChartRequest))
@@ -304,7 +334,7 @@ func _ChartsService_GetMinerWorkerShares_Handler(srv interface{}, ctx context.Co
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/charts.ChartsService/GetMinerWorkerShares",
+		FullMethod: ChartsService_GetMinerWorkerShares_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(ChartsServiceServer).GetMinerWorkerShares(ctx, req.(*GetMinerWorkerChartRequest))
