@@ -1,10 +1,10 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.2.0
-// - protoc             v5.26.1
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.29.3
 // source: pool.proto
 
-package poolProto
+package pool
 
 import (
 	context "context"
@@ -16,8 +16,15 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
+
+const (
+	PoolService_GetPoolInfo_FullMethodName    = "/pool.PoolService/GetPoolInfo"
+	PoolService_GetNetworkInfo_FullMethodName = "/pool.PoolService/GetNetworkInfo"
+	PoolService_GetPoolSlaves_FullMethodName  = "/pool.PoolService/GetPoolSlaves"
+	PoolService_GetPoolStats_FullMethodName   = "/pool.PoolService/GetPoolStats"
+)
 
 // PoolServiceClient is the client API for PoolService service.
 //
@@ -38,8 +45,9 @@ func NewPoolServiceClient(cc grpc.ClientConnInterface) PoolServiceClient {
 }
 
 func (c *poolServiceClient) GetPoolInfo(ctx context.Context, in *emptypb.Empty, opts ...grpc.CallOption) (*PoolInfo, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PoolInfo)
-	err := c.cc.Invoke(ctx, "/pool.PoolService/GetPoolInfo", in, out, opts...)
+	err := c.cc.Invoke(ctx, PoolService_GetPoolInfo_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -47,8 +55,9 @@ func (c *poolServiceClient) GetPoolInfo(ctx context.Context, in *emptypb.Empty, 
 }
 
 func (c *poolServiceClient) GetNetworkInfo(ctx context.Context, in *emptypb.Empty, opts ...grpc.CallOption) (*NetworkInfo, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(NetworkInfo)
-	err := c.cc.Invoke(ctx, "/pool.PoolService/GetNetworkInfo", in, out, opts...)
+	err := c.cc.Invoke(ctx, PoolService_GetNetworkInfo_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -56,8 +65,9 @@ func (c *poolServiceClient) GetNetworkInfo(ctx context.Context, in *emptypb.Empt
 }
 
 func (c *poolServiceClient) GetPoolSlaves(ctx context.Context, in *GetPoolAssetRequest, opts ...grpc.CallOption) (*PoolSlaves, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PoolSlaves)
-	err := c.cc.Invoke(ctx, "/pool.PoolService/GetPoolSlaves", in, out, opts...)
+	err := c.cc.Invoke(ctx, PoolService_GetPoolSlaves_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -65,8 +75,9 @@ func (c *poolServiceClient) GetPoolSlaves(ctx context.Context, in *GetPoolAssetR
 }
 
 func (c *poolServiceClient) GetPoolStats(ctx context.Context, in *GetPoolAssetRequest, opts ...grpc.CallOption) (*PoolStats, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PoolStats)
-	err := c.cc.Invoke(ctx, "/pool.PoolService/GetPoolStats", in, out, opts...)
+	err := c.cc.Invoke(ctx, PoolService_GetPoolStats_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -75,7 +86,7 @@ func (c *poolServiceClient) GetPoolStats(ctx context.Context, in *GetPoolAssetRe
 
 // PoolServiceServer is the server API for PoolService service.
 // All implementations must embed UnimplementedPoolServiceServer
-// for forward compatibility
+// for forward compatibility.
 type PoolServiceServer interface {
 	GetPoolInfo(context.Context, *emptypb.Empty) (*PoolInfo, error)
 	GetNetworkInfo(context.Context, *emptypb.Empty) (*NetworkInfo, error)
@@ -84,9 +95,12 @@ type PoolServiceServer interface {
 	mustEmbedUnimplementedPoolServiceServer()
 }
 
-// UnimplementedPoolServiceServer must be embedded to have forward compatible implementations.
-type UnimplementedPoolServiceServer struct {
-}
+// UnimplementedPoolServiceServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedPoolServiceServer struct{}
 
 func (UnimplementedPoolServiceServer) GetPoolInfo(context.Context, *emptypb.Empty) (*PoolInfo, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetPoolInfo not implemented")
@@ -101,6 +115,7 @@ func (UnimplementedPoolServiceServer) GetPoolStats(context.Context, *GetPoolAsse
 	return nil, status.Errorf(codes.Unimplemented, "method GetPoolStats not implemented")
 }
 func (UnimplementedPoolServiceServer) mustEmbedUnimplementedPoolServiceServer() {}
+func (UnimplementedPoolServiceServer) testEmbeddedByValue()                     {}
 
 // UnsafePoolServiceServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to PoolServiceServer will
@@ -110,6 +125,13 @@ type UnsafePoolServiceServer interface {
 }
 
 func RegisterPoolServiceServer(s grpc.ServiceRegistrar, srv PoolServiceServer) {
+	// If the following call pancis, it indicates UnimplementedPoolServiceServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&PoolService_ServiceDesc, srv)
 }
 
@@ -123,7 +145,7 @@ func _PoolService_GetPoolInfo_Handler(srv interface{}, ctx context.Context, dec 
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/pool.PoolService/GetPoolInfo",
+		FullMethod: PoolService_GetPoolInfo_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(PoolServiceServer).GetPoolInfo(ctx, req.(*emptypb.Empty))
@@ -141,7 +163,7 @@ func _PoolService_GetNetworkInfo_Handler(srv interface{}, ctx context.Context, d
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/pool.PoolService/GetNetworkInfo",
+		FullMethod: PoolService_GetNetworkInfo_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(PoolServiceServer).GetNetworkInfo(ctx, req.(*emptypb.Empty))
@@ -159,7 +181,7 @@ func _PoolService_GetPoolSlaves_Handler(srv interface{}, ctx context.Context, de
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/pool.PoolService/GetPoolSlaves",
+		FullMethod: PoolService_GetPoolSlaves_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(PoolServiceServer).GetPoolSlaves(ctx, req.(*GetPoolAssetRequest))
@@ -177,7 +199,7 @@ func _PoolService_GetPoolStats_Handler(srv interface{}, ctx context.Context, dec
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/pool.PoolService/GetPoolStats",
+		FullMethod: PoolService_GetPoolStats_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(PoolServiceServer).GetPoolStats(ctx, req.(*GetPoolAssetRequest))
