@@ -1,10 +1,10 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.2.0
-// - protoc             v5.26.1
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.29.3
 // source: pool_miners.proto
 
-package poolMinersProto
+package pool_miners
 
 import (
 	context "context"
@@ -15,8 +15,16 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
+
+const (
+	PoolMinersService_ValidateAddress_FullMethodName          = "/pool_miners.PoolMinersService/ValidateAddress"
+	PoolMinersService_GetMiners_FullMethodName                = "/pool_miners.PoolMinersService/GetMiners"
+	PoolMinersService_GetMiner_FullMethodName                 = "/pool_miners.PoolMinersService/GetMiner"
+	PoolMinersService_GetMinerWorkers_FullMethodName          = "/pool_miners.PoolMinersService/GetMinerWorkers"
+	PoolMinersService_GetMinersWorkersFromList_FullMethodName = "/pool_miners.PoolMinersService/GetMinersWorkersFromList"
+)
 
 // PoolMinersServiceClient is the client API for PoolMinersService service.
 //
@@ -38,8 +46,9 @@ func NewPoolMinersServiceClient(cc grpc.ClientConnInterface) PoolMinersServiceCl
 }
 
 func (c *poolMinersServiceClient) ValidateAddress(ctx context.Context, in *MinerAddressRequest, opts ...grpc.CallOption) (*ValidateAddressResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(ValidateAddressResponse)
-	err := c.cc.Invoke(ctx, "/pool_miners.PoolMinersService/ValidateAddress", in, out, opts...)
+	err := c.cc.Invoke(ctx, PoolMinersService_ValidateAddress_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -47,8 +56,9 @@ func (c *poolMinersServiceClient) ValidateAddress(ctx context.Context, in *Miner
 }
 
 func (c *poolMinersServiceClient) GetMiners(ctx context.Context, in *GetMinersRequest, opts ...grpc.CallOption) (*MinersList, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(MinersList)
-	err := c.cc.Invoke(ctx, "/pool_miners.PoolMinersService/GetMiners", in, out, opts...)
+	err := c.cc.Invoke(ctx, PoolMinersService_GetMiners_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -56,8 +66,9 @@ func (c *poolMinersServiceClient) GetMiners(ctx context.Context, in *GetMinersRe
 }
 
 func (c *poolMinersServiceClient) GetMiner(ctx context.Context, in *MinerAddressRequest, opts ...grpc.CallOption) (*Miner, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Miner)
-	err := c.cc.Invoke(ctx, "/pool_miners.PoolMinersService/GetMiner", in, out, opts...)
+	err := c.cc.Invoke(ctx, PoolMinersService_GetMiner_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -65,8 +76,9 @@ func (c *poolMinersServiceClient) GetMiner(ctx context.Context, in *MinerAddress
 }
 
 func (c *poolMinersServiceClient) GetMinerWorkers(ctx context.Context, in *GetMinerWorkersRequest, opts ...grpc.CallOption) (*MinerWorkersList, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(MinerWorkersList)
-	err := c.cc.Invoke(ctx, "/pool_miners.PoolMinersService/GetMinerWorkers", in, out, opts...)
+	err := c.cc.Invoke(ctx, PoolMinersService_GetMinerWorkers_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -74,8 +86,9 @@ func (c *poolMinersServiceClient) GetMinerWorkers(ctx context.Context, in *GetMi
 }
 
 func (c *poolMinersServiceClient) GetMinersWorkersFromList(ctx context.Context, in *MinerAddressesRequest, opts ...grpc.CallOption) (*MinersWorkersMap, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(MinersWorkersMap)
-	err := c.cc.Invoke(ctx, "/pool_miners.PoolMinersService/GetMinersWorkersFromList", in, out, opts...)
+	err := c.cc.Invoke(ctx, PoolMinersService_GetMinersWorkersFromList_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -84,7 +97,7 @@ func (c *poolMinersServiceClient) GetMinersWorkersFromList(ctx context.Context, 
 
 // PoolMinersServiceServer is the server API for PoolMinersService service.
 // All implementations must embed UnimplementedPoolMinersServiceServer
-// for forward compatibility
+// for forward compatibility.
 type PoolMinersServiceServer interface {
 	ValidateAddress(context.Context, *MinerAddressRequest) (*ValidateAddressResponse, error)
 	GetMiners(context.Context, *GetMinersRequest) (*MinersList, error)
@@ -94,9 +107,12 @@ type PoolMinersServiceServer interface {
 	mustEmbedUnimplementedPoolMinersServiceServer()
 }
 
-// UnimplementedPoolMinersServiceServer must be embedded to have forward compatible implementations.
-type UnimplementedPoolMinersServiceServer struct {
-}
+// UnimplementedPoolMinersServiceServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedPoolMinersServiceServer struct{}
 
 func (UnimplementedPoolMinersServiceServer) ValidateAddress(context.Context, *MinerAddressRequest) (*ValidateAddressResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method ValidateAddress not implemented")
@@ -114,6 +130,7 @@ func (UnimplementedPoolMinersServiceServer) GetMinersWorkersFromList(context.Con
 	return nil, status.Errorf(codes.Unimplemented, "method GetMinersWorkersFromList not implemented")
 }
 func (UnimplementedPoolMinersServiceServer) mustEmbedUnimplementedPoolMinersServiceServer() {}
+func (UnimplementedPoolMinersServiceServer) testEmbeddedByValue()                           {}
 
 // UnsafePoolMinersServiceServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to PoolMinersServiceServer will
@@ -123,6 +140,13 @@ type UnsafePoolMinersServiceServer interface {
 }
 
 func RegisterPoolMinersServiceServer(s grpc.ServiceRegistrar, srv PoolMinersServiceServer) {
+	// If the following call pancis, it indicates UnimplementedPoolMinersServiceServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&PoolMinersService_ServiceDesc, srv)
 }
 
@@ -136,7 +160,7 @@ func _PoolMinersService_ValidateAddress_Handler(srv interface{}, ctx context.Con
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/pool_miners.PoolMinersService/ValidateAddress",
+		FullMethod: PoolMinersService_ValidateAddress_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(PoolMinersServiceServer).ValidateAddress(ctx, req.(*MinerAddressRequest))
@@ -154,7 +178,7 @@ func _PoolMinersService_GetMiners_Handler(srv interface{}, ctx context.Context, 
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/pool_miners.PoolMinersService/GetMiners",
+		FullMethod: PoolMinersService_GetMiners_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(PoolMinersServiceServer).GetMiners(ctx, req.(*GetMinersRequest))
@@ -172,7 +196,7 @@ func _PoolMinersService_GetMiner_Handler(srv interface{}, ctx context.Context, d
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/pool_miners.PoolMinersService/GetMiner",
+		FullMethod: PoolMinersService_GetMiner_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(PoolMinersServiceServer).GetMiner(ctx, req.(*MinerAddressRequest))
@@ -190,7 +214,7 @@ func _PoolMinersService_GetMinerWorkers_Handler(srv interface{}, ctx context.Con
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/pool_miners.PoolMinersService/GetMinerWorkers",
+		FullMethod: PoolMinersService_GetMinerWorkers_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(PoolMinersServiceServer).GetMinerWorkers(ctx, req.(*GetMinerWorkersRequest))
@@ -208,7 +232,7 @@ func _PoolMinersService_GetMinersWorkersFromList_Handler(srv interface{}, ctx co
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/pool_miners.PoolMinersService/GetMinersWorkersFromList",
+		FullMethod: PoolMinersService_GetMinersWorkersFromList_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(PoolMinersServiceServer).GetMinersWorkersFromList(ctx, req.(*MinerAddressesRequest))
