@@ -1,10 +1,10 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.2.0
-// - protoc             v5.26.1
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.29.3
 // source: pool_payouts.proto
 
-package poolPayoutsProto
+package pool_payouts
 
 import (
 	context "context"
@@ -16,8 +16,17 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
+
+const (
+	PoolPayoutsService_GetMinersBalancesFromList_FullMethodName = "/pool_payouts.PoolPayoutsService/GetMinersBalancesFromList"
+	PoolPayoutsService_GetPayouts_FullMethodName                = "/pool_payouts.PoolPayoutsService/GetPayouts"
+	PoolPayoutsService_GetPayoutsFromList_FullMethodName        = "/pool_payouts.PoolPayoutsService/GetPayoutsFromList"
+	PoolPayoutsService_GetBlocks_FullMethodName                 = "/pool_payouts.PoolPayoutsService/GetBlocks"
+	PoolPayoutsService_GetSoloBlocks_FullMethodName             = "/pool_payouts.PoolPayoutsService/GetSoloBlocks"
+	PoolPayoutsService_GetSoloBlocksFromList_FullMethodName     = "/pool_payouts.PoolPayoutsService/GetSoloBlocksFromList"
+)
 
 // PoolPayoutsServiceClient is the client API for PoolPayoutsService service.
 //
@@ -40,8 +49,9 @@ func NewPoolPayoutsServiceClient(cc grpc.ClientConnInterface) PoolPayoutsService
 }
 
 func (c *poolPayoutsServiceClient) GetMinersBalancesFromList(ctx context.Context, in *pool_miners.MinerAddressesRequest, opts ...grpc.CallOption) (*MinersBalancesMap, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(MinersBalancesMap)
-	err := c.cc.Invoke(ctx, "/pool_payouts.PoolPayoutsService/GetMinersBalancesFromList", in, out, opts...)
+	err := c.cc.Invoke(ctx, PoolPayoutsService_GetMinersBalancesFromList_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -49,8 +59,9 @@ func (c *poolPayoutsServiceClient) GetMinersBalancesFromList(ctx context.Context
 }
 
 func (c *poolPayoutsServiceClient) GetPayouts(ctx context.Context, in *GetPayoutsRequest, opts ...grpc.CallOption) (*PayoutsList, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PayoutsList)
-	err := c.cc.Invoke(ctx, "/pool_payouts.PoolPayoutsService/GetPayouts", in, out, opts...)
+	err := c.cc.Invoke(ctx, PoolPayoutsService_GetPayouts_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -58,8 +69,9 @@ func (c *poolPayoutsServiceClient) GetPayouts(ctx context.Context, in *GetPayout
 }
 
 func (c *poolPayoutsServiceClient) GetPayoutsFromList(ctx context.Context, in *GetPayoutsFromListRequest, opts ...grpc.CallOption) (*PayoutsMap, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PayoutsMap)
-	err := c.cc.Invoke(ctx, "/pool_payouts.PoolPayoutsService/GetPayoutsFromList", in, out, opts...)
+	err := c.cc.Invoke(ctx, PoolPayoutsService_GetPayoutsFromList_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -67,8 +79,9 @@ func (c *poolPayoutsServiceClient) GetPayoutsFromList(ctx context.Context, in *G
 }
 
 func (c *poolPayoutsServiceClient) GetBlocks(ctx context.Context, in *GetBlocksRequest, opts ...grpc.CallOption) (*MinedBlocksList, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(MinedBlocksList)
-	err := c.cc.Invoke(ctx, "/pool_payouts.PoolPayoutsService/GetBlocks", in, out, opts...)
+	err := c.cc.Invoke(ctx, PoolPayoutsService_GetBlocks_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -76,8 +89,9 @@ func (c *poolPayoutsServiceClient) GetBlocks(ctx context.Context, in *GetBlocksR
 }
 
 func (c *poolPayoutsServiceClient) GetSoloBlocks(ctx context.Context, in *GetSoloBlocksRequest, opts ...grpc.CallOption) (*MinedSoloBlocksList, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(MinedSoloBlocksList)
-	err := c.cc.Invoke(ctx, "/pool_payouts.PoolPayoutsService/GetSoloBlocks", in, out, opts...)
+	err := c.cc.Invoke(ctx, PoolPayoutsService_GetSoloBlocks_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -85,8 +99,9 @@ func (c *poolPayoutsServiceClient) GetSoloBlocks(ctx context.Context, in *GetSol
 }
 
 func (c *poolPayoutsServiceClient) GetSoloBlocksFromList(ctx context.Context, in *GetSoloBlocksFromListRequest, opts ...grpc.CallOption) (*MinedSoloBlocksMap, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(MinedSoloBlocksMap)
-	err := c.cc.Invoke(ctx, "/pool_payouts.PoolPayoutsService/GetSoloBlocksFromList", in, out, opts...)
+	err := c.cc.Invoke(ctx, PoolPayoutsService_GetSoloBlocksFromList_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -95,7 +110,7 @@ func (c *poolPayoutsServiceClient) GetSoloBlocksFromList(ctx context.Context, in
 
 // PoolPayoutsServiceServer is the server API for PoolPayoutsService service.
 // All implementations must embed UnimplementedPoolPayoutsServiceServer
-// for forward compatibility
+// for forward compatibility.
 type PoolPayoutsServiceServer interface {
 	GetMinersBalancesFromList(context.Context, *pool_miners.MinerAddressesRequest) (*MinersBalancesMap, error)
 	GetPayouts(context.Context, *GetPayoutsRequest) (*PayoutsList, error)
@@ -106,9 +121,12 @@ type PoolPayoutsServiceServer interface {
 	mustEmbedUnimplementedPoolPayoutsServiceServer()
 }
 
-// UnimplementedPoolPayoutsServiceServer must be embedded to have forward compatible implementations.
-type UnimplementedPoolPayoutsServiceServer struct {
-}
+// UnimplementedPoolPayoutsServiceServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedPoolPayoutsServiceServer struct{}
 
 func (UnimplementedPoolPayoutsServiceServer) GetMinersBalancesFromList(context.Context, *pool_miners.MinerAddressesRequest) (*MinersBalancesMap, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetMinersBalancesFromList not implemented")
@@ -129,6 +147,7 @@ func (UnimplementedPoolPayoutsServiceServer) GetSoloBlocksFromList(context.Conte
 	return nil, status.Errorf(codes.Unimplemented, "method GetSoloBlocksFromList not implemented")
 }
 func (UnimplementedPoolPayoutsServiceServer) mustEmbedUnimplementedPoolPayoutsServiceServer() {}
+func (UnimplementedPoolPayoutsServiceServer) testEmbeddedByValue()                            {}
 
 // UnsafePoolPayoutsServiceServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to PoolPayoutsServiceServer will
@@ -138,6 +157,13 @@ type UnsafePoolPayoutsServiceServer interface {
 }
 
 func RegisterPoolPayoutsServiceServer(s grpc.ServiceRegistrar, srv PoolPayoutsServiceServer) {
+	// If the following call pancis, it indicates UnimplementedPoolPayoutsServiceServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&PoolPayoutsService_ServiceDesc, srv)
 }
 
@@ -151,7 +177,7 @@ func _PoolPayoutsService_GetMinersBalancesFromList_Handler(srv interface{}, ctx 
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/pool_payouts.PoolPayoutsService/GetMinersBalancesFromList",
+		FullMethod: PoolPayoutsService_GetMinersBalancesFromList_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(PoolPayoutsServiceServer).GetMinersBalancesFromList(ctx, req.(*pool_miners.MinerAddressesRequest))
@@ -169,7 +195,7 @@ func _PoolPayoutsService_GetPayouts_Handler(srv interface{}, ctx context.Context
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/pool_payouts.PoolPayoutsService/GetPayouts",
+		FullMethod: PoolPayoutsService_GetPayouts_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(PoolPayoutsServiceServer).GetPayouts(ctx, req.(*GetPayoutsRequest))
@@ -187,7 +213,7 @@ func _PoolPayoutsService_GetPayoutsFromList_Handler(srv interface{}, ctx context
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/pool_payouts.PoolPayoutsService/GetPayoutsFromList",
+		FullMethod: PoolPayoutsService_GetPayoutsFromList_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(PoolPayoutsServiceServer).GetPayoutsFromList(ctx, req.(*GetPayoutsFromListRequest))
@@ -205,7 +231,7 @@ func _PoolPayoutsService_GetBlocks_Handler(srv interface{}, ctx context.Context,
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/pool_payouts.PoolPayoutsService/GetBlocks",
+		FullMethod: PoolPayoutsService_GetBlocks_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(PoolPayoutsServiceServer).GetBlocks(ctx, req.(*GetBlocksRequest))
@@ -223,7 +249,7 @@ func _PoolPayoutsService_GetSoloBlocks_Handler(srv interface{}, ctx context.Cont
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/pool_payouts.PoolPayoutsService/GetSoloBlocks",
+		FullMethod: PoolPayoutsService_GetSoloBlocks_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(PoolPayoutsServiceServer).GetSoloBlocks(ctx, req.(*GetSoloBlocksRequest))
@@ -241,7 +267,7 @@ func _PoolPayoutsService_GetSoloBlocksFromList_Handler(srv interface{}, ctx cont
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/pool_payouts.PoolPayoutsService/GetSoloBlocksFromList",
+		FullMethod: PoolPayoutsService_GetSoloBlocksFromList_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(PoolPayoutsServiceServer).GetSoloBlocksFromList(ctx, req.(*GetSoloBlocksFromListRequest))
